Differential Causal Effects - main function

Main function to compute differential causal effects and the pathway enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = "glm.dce.fit"),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  verbose = FALSE
)

Arguments

graph
valid object defining a directed acyclic graph

df_expr_wt
data frame with wild type expression values

df_expr_mt
data from with mutation type expression values

solver
character with name of solver function

solver_args
additional arguments for the solver function

adjustment_type
character string for the method to define the adjustment set Z for the regression

effect_type
method of computing causal effects

p_method
character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test
either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size
either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding
indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

verbose
logical for verbose output

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,X.mt)

#> Warning: Cycle(s) detected in network

#> Error in glm.dce.fit(x = structure(c(1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 778, 1100, 920, 974, 829, 1094, 942, 1129, 906, 1128, 848, 1043, 943, 864, 859, 991, 1495, 859, 758, 987, 1019, 1061, 953, 829, 944, 1078, 975, 879, 1084, 962, 1152, 1152, 863, 1037, 1091, 1054, 955, 1034, 1029, 1043, 1362, 1042, 965, 947, 901, 1125, 998, 939, 908, 981, 978, 1068, 1049, 1102, 913, 956, 717, 1021, 1073, 1033, 1007, 1245, 1042, 969, 1031, 908, 925, 1119, 1056, 984, 987, 902, 934, 980, 809, 1007, 817, 956, 953, 953, 1145, 1166, 771, 1098, 900, 922, 819, 922, 985, 884, 984, 907, 887, 989, 1052, 952, 1026, 1052, 966, 1222, 954, 795, 987, 1034, 886, 876, 1116, 925, 1106, 1043, 898, 877, 1073, 1008, 1098, 1023, 976, 877, 1010, 914, 874, 1266, 1094, 972, 1089, 1038, 1123, 1014, 1083, 1124, 1004, 1053, 1119, 1206, 816, 849, 1201, 1011, 890, 1132, 947, 949, 887, 1015, 929, 1094, 826, 1167, 741, 1133, 1053, 972, 1073, 1067, 843, 946, 804, 1091, 1056, 1007, 973, 1117, 1146, 892, 1031, 1097, 993, 1118, 1081, 1133, 850, 924, 1077, 994, 1049, 995, 1114, 955, 791, 932, 1073, 1077, 1067, 774, 992, 953, 1029, 944, 1032, 845, 1019, 907, 1055, 896, 1020, 1186, 1058, 1183, 918, 940, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 954, 795, 987, 1034, 886, 876, 1116, 925, 1106, 1043, 898, 877, 1073, 1008, 1098, 1023, 976, 877, 1010, 914, 874, 1266, 1094, 972, 1089, 1038, 1123, 1014, 1083, 1124, 1004, 1053, 1119, 1206, 816, 849, 1201, 1011, 890, 1132, 947, 949, 887, 1015, 929, 1094, 826, 1167, 741, 1133, 1053, 972, 1073, 1067, 843, 946, 804, 1091, 1056, 1007, 973, 1117, 1146, 892, 1031, 1097, 993, 1118, 1081, 1133, 850, 924, 1077, 994, 1049, 995, 1114, 955, 791, 932, 1073, 1077, 1067, 774, 992, 953, 1029, 944, 1032, 845, 1019, 907, 1055, 896, 1020, 1186, 1058, 1183, 918, 940), .Dim = c(200L, 4L), .Dimnames = list(c("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50", "51", "52", "53", "54", "55", "56", "57", "58", "59", "60", "61", "62", "63", "64", "65", "66", "67", "68", "69", "70", "71", "72", "73", "74", "75", "76", "77", "78", "79", "80", "81", "82", "83", "84", "85", "86", "87", "88", "89", "90", "91", "92", "93", "94", "95", "96", "97", "98", "99", "100", "101", "102", "103", "104", "105", "106", "107", "108", "109", "110", "111", "112", "113", "114", "115", "116", "117", "118", "119", "120", "121", "122", "123", "124", "125", "126", "127", "128", "129", "130", "131", "132", "133", "134", "135", "136", "137", "138", "139", "140", "141", "142", "143", "144", "145", "146", "147", "148", "149", "150", "151", "152", "153", "154", "155", "156", "157", "158", "159", "160", "161", "162", "163", "164", "165", "166", "167", "168", "169", "170", "171", "172", "173", "174", "175", "176", "177", "178", "179", "180", "181", "182", "183", "184", "185", "186", "187", "188", "189", "190", "191", "192", "193", "194", "195", "196", "197", "198", "199", "200"), c("(Intercept)", "N", "X", "N:X")), assign = 0:3), y = c(`1` = 425, `2` = 238, `3` = 326, `4` = 336, `5` = 355, `6` = 223, `7` = 320, `8` = 218, `9` = 333, `10` = 226, `11` = 387, `12` = 262, `13` = 330, `14` = 348, `15` = 401, `16` = 330, `17` = 2, `18` = 372, `19` = 408, `20` = 276, `21` = 279, `22` = 276, `23` = 328, `24` = 354, `25` = 286, `26` = 241, `27` = 280, `28` = 360, `29` = 228, `30` = 261, `31` = 191, `32` = 209, `33` = 387, `34` = 269, `35` = 227, `36` = 311, `37` = 289, `38` = 252, `39` = 269, `40` = 243, `41` = 82, `42` = 255, `43` = 321, `44` = 314, `45` = 331, `46` = 239, `47` = 296, `48` = 303, `49` = 377, `50` = 300, `51` = 279, `52` = 239, `53` = 251, `54` = 235, `55` = 319, `56` = 346, `57` = 441, `58` = 291, `59` = 242, `60` = 283, `61` = 296, `62` = 152, `63` = 268, `64` = 323, `65` = 275, `66` = 371, `67` = 345, `68` = 202, `69` = 228, `70` = 331, `71` = 332, `72` = 350, `73` = 330, `74` = 283, `75` = 399, `76` = 291, `77` = 374, `78` = 331, `79` = 326, `80` = 306, `81` = 201, `82` = 196, `83` = 424, `84` = 244, `85` = 361, `86` = 334, `87` = 403, `88` = 317, `89` = 290, `90` = 372, `91` = 305, `92` = 325, `93` = 347, `94` = 301, `95` = 258, `96` = 317, `97` = 274, `98` = 277, `99` = 295, `100` = 153, `101` = 158, `102` = 222, `103` = 171, `104` = 133, `105` = 202, `106` = 263, `107` = 106, `108` = 171, `109` = 111, `110` = 126, `111` = 229, `112` = 229, `113` = 114, `114` = 179, `115` = 90, `116` = 119, `117` = 171, `118` = 246, `119` = 145, `120` = 203, `121` = 205, `122` = 0, `123` = 87, `124` = 169, `125` = 93, `126` = 143, `127` = 91, `128` = 150, `129` = 102, `130` = 66, `131` = 158, `132` = 141, `133` = 74, `134` = 43, `135` = 267, `136` = 240, `137` = 45, `138` = 140, `139` = 223, `140` = 81, `141` = 168, `142` = 178, `143` = 205, `144` = 142, `145` = 199, `146` = 99, `147` = 256, `148` = 48, `149` = 304, `150` = 79, `151` = 118, `152` = 153, `153` = 99, `154` = 113, `155` = 269, `156` = 195, `157` = 282, `158` = 107, `159` = 142, `160` = 134, `161` = 166, `162` = 85, `163` = 65, `164` = 213, `165` = 147, `166` = 87, `167` = 168, `168` = 82, `169` = 103, `170` = 73, `171` = 240, `172` = 203, `173` = 108, `174` = 164, `175` = 130, `176` = 161, `177` = 86, `178` = 178, `179` = 275, `180` = 189, `181` = 109, `182` = 122, `183` = 102, `184` = 264, `185` = 139, `186` = 175, `187` = 131, `188` = 171, `189` = 144, `190` = 282, `191` = 146, `192` = 251, `193` = 128, `194` = 226, `195` = 136, `196` = 55, `197` = 105, `198` = 46, `199` = 222, `200` = 195), weights = NULL, start = NULL,     etastart = NULL, mustart = NULL, offset = NULL, family = structure(list(        family = "gaussian", link = "identity", linkfun = function (mu)         mu, linkinv = function (eta)         eta, variance = function (mu)         rep.int(1, length(mu)), dev.resids = function (y, mu,             wt)         wt * ((y - mu)^2), aic = function (y, n, mu, wt, dev)         {            nobs <- length(y)            nobs * (log(dev/nobs * 2 * pi) + 1) + 2 - sum(log(wt))        }, mu.eta = function (eta)         rep.int(1, length(eta)), initialize = expression({            n <- rep.int(1, nobs)            if (is.null(etastart) && is.null(start) && is.null(mustart) &&                 ((family$link == "inverse" && any(y == 0)) ||                   (family$link == "log" && any(y <= 0))))                 stop("cannot find valid starting values: please specify some")            mustart <- y        }), validmu = function (mu)         TRUE, valideta = function (eta)         TRUE), class = "family"), control = list(), intercept = TRUE,     singular.ok = TRUE): could not find function "glm.dce.fit"