Differential Causal Effects - main function

Main function to compute differential causal effects and the pathway enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "lm",
  solver_args = list(),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for igraph
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "lm",
  solver_args = list(),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for graphNEL
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "lm",
  solver_args = list(),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for matrix
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "lm",
  solver_args = list(),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph

valid object defining a directed acyclic graph

df_expr_wt

data frame with wild type expression values

df_expr_mt

data from with mutation type expression values

solver

character with name of solver function

solver_args

additional arguments for the solver function. please adress this argument, if you use your own solver function. the default argument works with glm functions in the packages MASS, stats and glm2

adjustment_type

character string for the method to define the adjustment set Z for the regression

effect_type

method of computing causal effects

p_method

character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test

either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test. Alternatively, "vcovHC" can improve results for zero-inflated date, i.e., from single cell RNAseq experiments.

lib_size

either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding

indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative

logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level

Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,X.mt)
#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  1  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   1   0   0
#> n2   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n3   0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n4   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n5   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n6   0  0  0  0  0  0  0  0  0   0   1   0   0   1   0   0   0   0   0   0   0
#> n7   0  0  0  0  0  0  0  1  0   1   0   1   1   0   0   0   0   0   0   0   0
#> n8   0  0  0  0  0  0  0  0  1   0   0   1   0   1   0   0   1   0   0   0   1
#> n9   0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0   0   1   0
#> n10  0  0  0  0  0  0  0  0  0   0   1   0   0   0   1   0   1   0   1   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   1   0   0   0   0   0
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   1   1   0   0   1
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   1   0   0   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   1   1   0
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    0   0   0   0   0   0   0   0   0
#> n2    1   0   1   0   0   0   0   1   0
#> n3    0   0   0   0   0   0   0   0   0
#> n4    0   0   0   0   0   0   0   0   0
#> n5    1   0   0   0   0   0   0   1   0
#> n6    0   1   0   0   1   0   0   0   0
#> n7    0   0   0   0   0   0   0   0   0
#> n8    1   1   0   0   1   0   0   0   0
#> n9    0   0   0   0   0   0   1   0   0
#> n10   0   0   0   0   0   0   0   0   0
#> n11   0   0   0   0   0   0   0   0   0
#> n12   0   0   0   1   0   1   0   1   0
#> n13   0   0   0   0   0   0   0   0   0
#> n14   0   0   0   0   0   0   0   0   0
#> n15   0   0   0   0   0   0   0   0   0
#> n16   0   0   1   0   0   0   0   0   0
#> n17   0   1   0   0   0   0   1   0   0
#> n18   0   0   0   0   0   0   0   0   0
#> n19   0   0   0   1   0   0   0   0   0
#> n20   0   0   0   0   1   0   0   0   0
#> n21   0   0   1   1   0   0   0   0   0
#> n22   0   0   0   0   0   1   1   0   0
#> n23   0   0   0   0   0   0   0   0   0
#> n24   0   0   0   0   0   0   0   0   0
#> n25   0   0   0   0   0   0   0   0   0
#> n26   0   0   0   0   0   0   0   0   1
#> n27   0   0   0   0   0   0   0   0   0
#> n28   0   0   0   0   0   0   0   0   1
#> n29   0   0   0   0   0   0   0   0   1
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1           n2 n3 n4 n5 n6 n7        n8        n9          n10         n11
#> n1  NA -0.005533955 NA NA NA NA NA        NA        NA           NA          NA
#> n2  NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n3  NA           NA NA NA NA NA NA 0.1711090        NA           NA          NA
#> n4  NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n5  NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n6  NA           NA NA NA NA NA NA        NA        NA           NA -0.02667937
#> n7  NA           NA NA NA NA NA NA 0.2456368        NA -0.008784167          NA
#> n8  NA           NA NA NA NA NA NA        NA 0.2413384           NA          NA
#> n9  NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n10 NA           NA NA NA NA NA NA        NA        NA           NA -0.07562367
#> n11 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n12 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n13 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n14 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n15 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n16 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n17 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n18 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n19 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n20 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n21 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n22 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n23 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n24 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n25 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n26 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n27 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n28 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n29 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#> n30 NA           NA NA NA NA NA NA        NA        NA           NA          NA
#>            n12        n13         n14         n15         n16         n17
#> n1          NA         NA          NA          NA          NA          NA
#> n2          NA         NA          NA          NA          NA          NA
#> n3          NA         NA          NA          NA          NA          NA
#> n4          NA         NA          NA          NA          NA          NA
#> n5          NA         NA          NA          NA          NA          NA
#> n6          NA         NA 0.004491339          NA          NA          NA
#> n7  0.09836086 0.01911909          NA          NA          NA          NA
#> n8  0.13515225         NA 0.072996949          NA          NA -0.17876388
#> n9          NA         NA          NA -0.22382571          NA          NA
#> n10         NA         NA          NA  0.04249155          NA -0.18304364
#> n11         NA         NA          NA          NA          NA          NA
#> n12         NA         NA 0.117706336          NA -0.31578355          NA
#> n13         NA         NA          NA          NA  0.01299381 -0.01511054
#> n14         NA         NA          NA          NA          NA          NA
#> n15         NA         NA          NA          NA          NA          NA
#> n16         NA         NA          NA          NA          NA          NA
#> n17         NA         NA          NA          NA          NA          NA
#> n18         NA         NA          NA          NA          NA          NA
#> n19         NA         NA          NA          NA          NA          NA
#> n20         NA         NA          NA          NA          NA          NA
#> n21         NA         NA          NA          NA          NA          NA
#> n22         NA         NA          NA          NA          NA          NA
#> n23         NA         NA          NA          NA          NA          NA
#> n24         NA         NA          NA          NA          NA          NA
#> n25         NA         NA          NA          NA          NA          NA
#> n26         NA         NA          NA          NA          NA          NA
#> n27         NA         NA          NA          NA          NA          NA
#> n28         NA         NA          NA          NA          NA          NA
#> n29         NA         NA          NA          NA          NA          NA
#> n30         NA         NA          NA          NA          NA          NA
#>           n18         n19        n20         n21         n22         n23
#> n1         NA -0.09589305         NA          NA          NA          NA
#> n2         NA          NA         NA          NA -0.26233921          NA
#> n3         NA          NA         NA          NA          NA          NA
#> n4         NA          NA         NA          NA          NA          NA
#> n5         NA          NA         NA          NA  0.02919431          NA
#> n6         NA          NA         NA          NA          NA -0.03706897
#> n7         NA          NA         NA          NA          NA          NA
#> n8         NA          NA         NA  0.01495956  0.12872407  0.80695256
#> n9         NA          NA  0.8272839          NA          NA          NA
#> n10        NA  0.50896610         NA          NA          NA          NA
#> n11        NA          NA         NA          NA          NA          NA
#> n12        NA          NA         NA          NA          NA          NA
#> n13 0.1082187          NA         NA  0.08122685          NA          NA
#> n14 0.3640612          NA         NA          NA          NA          NA
#> n15        NA          NA         NA          NA          NA          NA
#> n16        NA -0.48906929  0.5493404          NA          NA          NA
#> n17        NA          NA         NA -0.03722746          NA  0.20061383
#> n18        NA          NA -0.6893420          NA          NA          NA
#> n19        NA          NA         NA          NA          NA          NA
#> n20        NA          NA         NA          NA          NA          NA
#> n21        NA          NA         NA          NA          NA          NA
#> n22        NA          NA         NA          NA          NA          NA
#> n23        NA          NA         NA          NA          NA          NA
#> n24        NA          NA         NA          NA          NA          NA
#> n25        NA          NA         NA          NA          NA          NA
#> n26        NA          NA         NA          NA          NA          NA
#> n27        NA          NA         NA          NA          NA          NA
#> n28        NA          NA         NA          NA          NA          NA
#> n29        NA          NA         NA          NA          NA          NA
#> n30        NA          NA         NA          NA          NA          NA
#>            n24           n25        n26       n27        n28         n29
#> n1          NA            NA         NA        NA         NA          NA
#> n2   0.1029146            NA         NA        NA         NA  0.33538593
#> n3          NA            NA         NA        NA         NA          NA
#> n4          NA            NA         NA        NA         NA          NA
#> n5          NA            NA         NA        NA         NA -0.01770254
#> n6          NA            NA  0.0236040        NA         NA          NA
#> n7          NA            NA         NA        NA         NA          NA
#> n8          NA            NA -0.4035538        NA         NA          NA
#> n9          NA            NA         NA        NA -0.8333316          NA
#> n10         NA            NA         NA        NA         NA          NA
#> n11         NA            NA         NA        NA         NA          NA
#> n12         NA -0.7504644943         NA 0.2907813         NA  3.15044799
#> n13         NA            NA         NA        NA         NA          NA
#> n14         NA            NA         NA        NA         NA          NA
#> n15         NA            NA         NA        NA         NA          NA
#> n16  0.1384568            NA         NA        NA         NA          NA
#> n17         NA            NA         NA        NA -0.4307226          NA
#> n18         NA            NA         NA        NA         NA          NA
#> n19         NA  0.0008078723         NA        NA         NA          NA
#> n20         NA            NA  0.4945307        NA         NA          NA
#> n21 -0.2917472  0.4900906539         NA        NA         NA          NA
#> n22         NA            NA         NA 0.3222003 -0.4561933          NA
#> n23         NA            NA         NA        NA         NA          NA
#> n24         NA            NA         NA        NA         NA          NA
#> n25         NA            NA         NA        NA         NA          NA
#> n26         NA            NA         NA        NA         NA          NA
#> n27         NA            NA         NA        NA         NA          NA
#> n28         NA            NA         NA        NA         NA          NA
#> n29         NA            NA         NA        NA         NA          NA
#> n30         NA            NA         NA        NA         NA          NA
#>            n30
#> n1          NA
#> n2          NA
#> n3          NA
#> n4          NA
#> n5          NA
#> n6          NA
#> n7          NA
#> n8          NA
#> n9          NA
#> n10         NA
#> n11         NA
#> n12         NA
#> n13         NA
#> n14         NA
#> n15         NA
#> n16         NA
#> n17         NA
#> n18         NA
#> n19         NA
#> n20         NA
#> n21         NA
#> n22         NA
#> n23         NA
#> n24         NA
#> n25         NA
#> n26 0.70091365
#> n27         NA
#> n28 0.03346003
#> n29 0.75875039
#> n30         NA
#> 
#> $dce_stderr
#>     n1         n2 n3 n4 n5 n6 n7         n8        n9        n10       n11
#> n1  NA 0.01700439 NA NA NA NA NA         NA        NA         NA        NA
#> n2  NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n3  NA         NA NA NA NA NA NA 0.07171847        NA         NA        NA
#> n4  NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n5  NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n6  NA         NA NA NA NA NA NA         NA        NA         NA 0.1136636
#> n7  NA         NA NA NA NA NA NA 0.08935304        NA 0.02269994        NA
#> n8  NA         NA NA NA NA NA NA         NA 0.1226221         NA        NA
#> n9  NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n10 NA         NA NA NA NA NA NA         NA        NA         NA 0.3287959
#> n11 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n12 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n13 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n14 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n15 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n16 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n17 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n18 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n19 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n20 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n21 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n22 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n23 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n24 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n25 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n26 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n27 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n28 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n29 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#> n30 NA         NA NA NA NA NA NA         NA        NA         NA        NA
#>            n12        n13        n14       n15       n16       n17       n18
#> n1          NA         NA         NA        NA        NA        NA        NA
#> n2          NA         NA         NA        NA        NA        NA        NA
#> n3          NA         NA         NA        NA        NA        NA        NA
#> n4          NA         NA         NA        NA        NA        NA        NA
#> n5          NA         NA         NA        NA        NA        NA        NA
#> n6          NA         NA 0.03863234        NA        NA        NA        NA
#> n7  0.04290373 0.01997409         NA        NA        NA        NA        NA
#> n8  0.12157205         NA 0.09957120        NA        NA 0.2681802        NA
#> n9          NA         NA         NA 0.4756089        NA        NA        NA
#> n10         NA         NA         NA 0.4853241        NA 0.3910342        NA
#> n11         NA         NA         NA        NA        NA        NA        NA
#> n12         NA         NA 0.12976684        NA 0.1683907        NA        NA
#> n13         NA         NA         NA        NA 0.1713645 0.4454245 0.1600875
#> n14         NA         NA         NA        NA        NA        NA 0.5181311
#> n15         NA         NA         NA        NA        NA        NA        NA
#> n16         NA         NA         NA        NA        NA        NA        NA
#> n17         NA         NA         NA        NA        NA        NA        NA
#> n18         NA         NA         NA        NA        NA        NA        NA
#> n19         NA         NA         NA        NA        NA        NA        NA
#> n20         NA         NA         NA        NA        NA        NA        NA
#> n21         NA         NA         NA        NA        NA        NA        NA
#> n22         NA         NA         NA        NA        NA        NA        NA
#> n23         NA         NA         NA        NA        NA        NA        NA
#> n24         NA         NA         NA        NA        NA        NA        NA
#> n25         NA         NA         NA        NA        NA        NA        NA
#> n26         NA         NA         NA        NA        NA        NA        NA
#> n27         NA         NA         NA        NA        NA        NA        NA
#> n28         NA         NA         NA        NA        NA        NA        NA
#> n29         NA         NA         NA        NA        NA        NA        NA
#> n30         NA         NA         NA        NA        NA        NA        NA
#>           n19       n20        n21        n22        n23       n24       n25
#> n1  0.1233569        NA         NA         NA         NA        NA        NA
#> n2         NA        NA         NA 0.69155273         NA 0.4147581        NA
#> n3         NA        NA         NA         NA         NA        NA        NA
#> n4         NA        NA         NA         NA         NA        NA        NA
#> n5         NA        NA         NA 0.09629988         NA        NA        NA
#> n6         NA        NA         NA         NA 0.09522934        NA        NA
#> n7         NA        NA         NA         NA         NA        NA        NA
#> n8         NA        NA 0.15126943 0.53850367 0.79507856        NA        NA
#> n9         NA 1.0759087         NA         NA         NA        NA        NA
#> n10 0.6023502        NA         NA         NA         NA        NA        NA
#> n11        NA        NA         NA         NA         NA        NA        NA
#> n12        NA        NA         NA         NA         NA        NA 0.3318678
#> n13        NA        NA 0.14798892         NA         NA        NA        NA
#> n14        NA        NA         NA         NA         NA        NA        NA
#> n15        NA        NA         NA         NA         NA        NA        NA
#> n16 0.8769037 0.4118443         NA         NA         NA 0.4926213        NA
#> n17        NA        NA 0.07134177         NA 0.63611720        NA        NA
#> n18        NA 0.2728117         NA         NA         NA        NA        NA
#> n19        NA        NA         NA         NA         NA        NA 0.1802919
#> n20        NA        NA         NA         NA         NA        NA        NA
#> n21        NA        NA         NA         NA         NA 0.6536056 0.4852455
#> n22        NA        NA         NA         NA         NA        NA        NA
#> n23        NA        NA         NA         NA         NA        NA        NA
#> n24        NA        NA         NA         NA         NA        NA        NA
#> n25        NA        NA         NA         NA         NA        NA        NA
#> n26        NA        NA         NA         NA         NA        NA        NA
#> n27        NA        NA         NA         NA         NA        NA        NA
#> n28        NA        NA         NA         NA         NA        NA        NA
#> n29        NA        NA         NA         NA         NA        NA        NA
#> n30        NA        NA         NA         NA         NA        NA        NA
#>           n26       n27       n28        n29       n30
#> n1         NA        NA        NA         NA        NA
#> n2         NA        NA        NA 1.10774017        NA
#> n3         NA        NA        NA         NA        NA
#> n4         NA        NA        NA         NA        NA
#> n5         NA        NA        NA 0.06124827        NA
#> n6  0.1143774        NA        NA         NA        NA
#> n7         NA        NA        NA         NA        NA
#> n8  0.3873885        NA        NA         NA        NA
#> n9         NA        NA 1.1760675         NA        NA
#> n10        NA        NA        NA         NA        NA
#> n11        NA        NA        NA         NA        NA
#> n12        NA 0.4374662        NA 1.25741341        NA
#> n13        NA        NA        NA         NA        NA
#> n14        NA        NA        NA         NA        NA
#> n15        NA        NA        NA         NA        NA
#> n16        NA        NA        NA         NA        NA
#> n17        NA        NA 0.3164042         NA        NA
#> n18        NA        NA        NA         NA        NA
#> n19        NA        NA        NA         NA        NA
#> n20 0.4082011        NA        NA         NA        NA
#> n21        NA        NA        NA         NA        NA
#> n22        NA 0.5692940 1.1577544         NA        NA
#> n23        NA        NA        NA         NA        NA
#> n24        NA        NA        NA         NA        NA
#> n25        NA        NA        NA         NA        NA
#> n26        NA        NA        NA         NA 0.6008036
#> n27        NA        NA        NA         NA        NA
#> n28        NA        NA        NA         NA 0.5499341
#> n29        NA        NA        NA         NA 1.0449847
#> n30        NA        NA        NA         NA        NA
#> 
#> $dce_pvalue
#>     n1        n2 n3 n4 n5 n6 n7          n8         n9       n10       n11
#> n1  NA 0.7451932 NA NA NA NA NA          NA         NA        NA        NA
#> n2  NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n3  NA        NA NA NA NA NA NA 0.017992331         NA        NA        NA
#> n4  NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n5  NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n6  NA        NA NA NA NA NA NA          NA         NA        NA 0.8146694
#> n7  NA        NA NA NA NA NA NA 0.006535317         NA 0.6991991        NA
#> n8  NA        NA NA NA NA NA NA          NA 0.05049068        NA        NA
#> n9  NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n10 NA        NA NA NA NA NA NA          NA         NA        NA 0.8183327
#> n11 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n12 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n13 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n14 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n15 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n16 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n17 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n18 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n19 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n20 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n21 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n22 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n23 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n24 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n25 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n26 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n27 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n28 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n29 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#> n30 NA        NA NA NA NA NA NA          NA         NA        NA        NA
#>           n12       n13       n14       n15        n16       n17       n18
#> n1         NA        NA        NA        NA         NA        NA        NA
#> n2         NA        NA        NA        NA         NA        NA        NA
#> n3         NA        NA        NA        NA         NA        NA        NA
#> n4         NA        NA        NA        NA         NA        NA        NA
#> n5         NA        NA        NA        NA         NA        NA        NA
#> n6         NA        NA 0.9075667        NA         NA        NA        NA
#> n7  0.0229331 0.3396481        NA        NA         NA        NA        NA
#> n8  0.2676551        NA 0.4643836        NA         NA 0.5058399        NA
#> n9         NA        NA        NA 0.6384488         NA        NA        NA
#> n10        NA        NA        NA 0.9303223         NA 0.6402376        NA
#> n11        NA        NA        NA        NA         NA        NA        NA
#> n12        NA        NA 0.3655119        NA 0.06226917        NA        NA
#> n13        NA        NA        NA        NA 0.93963600 0.9729727 0.4998474
#> n14        NA        NA        NA        NA         NA        NA 0.4831382
#> n15        NA        NA        NA        NA         NA        NA        NA
#> n16        NA        NA        NA        NA         NA        NA        NA
#> n17        NA        NA        NA        NA         NA        NA        NA
#> n18        NA        NA        NA        NA         NA        NA        NA
#> n19        NA        NA        NA        NA         NA        NA        NA
#> n20        NA        NA        NA        NA         NA        NA        NA
#> n21        NA        NA        NA        NA         NA        NA        NA
#> n22        NA        NA        NA        NA         NA        NA        NA
#> n23        NA        NA        NA        NA         NA        NA        NA
#> n24        NA        NA        NA        NA         NA        NA        NA
#> n25        NA        NA        NA        NA         NA        NA        NA
#> n26        NA        NA        NA        NA         NA        NA        NA
#> n27        NA        NA        NA        NA         NA        NA        NA
#> n28        NA        NA        NA        NA         NA        NA        NA
#> n29        NA        NA        NA        NA         NA        NA        NA
#> n30        NA        NA        NA        NA         NA        NA        NA
#>           n19        n20       n21       n22       n23       n24        n25
#> n1  0.4378820         NA        NA        NA        NA        NA         NA
#> n2         NA         NA        NA 0.7048443        NA 0.8042951         NA
#> n3         NA         NA        NA        NA        NA        NA         NA
#> n4         NA         NA        NA        NA        NA        NA         NA
#> n5         NA         NA        NA 0.7620891        NA        NA         NA
#> n6         NA         NA        NA        NA 0.6975064        NA         NA
#> n7         NA         NA        NA        NA        NA        NA         NA
#> n8         NA         NA 0.9213260 0.8113293 0.3114144        NA         NA
#> n9         NA 0.44287801        NA        NA        NA        NA         NA
#> n10 0.3991700         NA        NA        NA        NA        NA         NA
#> n11        NA         NA        NA        NA        NA        NA         NA
#> n12        NA         NA        NA        NA        NA        NA 0.02485974
#> n13        NA         NA 0.5837250        NA        NA        NA         NA
#> n14        NA         NA        NA        NA        NA        NA         NA
#> n15        NA         NA        NA        NA        NA        NA         NA
#> n16 0.5776833 0.18383159        NA        NA        NA 0.7789660         NA
#> n17        NA         NA 0.6024044        NA 0.7528250        NA         NA
#> n18        NA 0.01231747        NA        NA        NA        NA         NA
#> n19        NA         NA        NA        NA        NA        NA 0.99642946
#> n20        NA         NA        NA        NA        NA        NA         NA
#> n21        NA         NA        NA        NA        NA 0.6558414 0.31378699
#> n22        NA         NA        NA        NA        NA        NA         NA
#> n23        NA         NA        NA        NA        NA        NA         NA
#> n24        NA         NA        NA        NA        NA        NA         NA
#> n25        NA         NA        NA        NA        NA        NA         NA
#> n26        NA         NA        NA        NA        NA        NA         NA
#> n27        NA         NA        NA        NA        NA        NA         NA
#> n28        NA         NA        NA        NA        NA        NA         NA
#> n29        NA         NA        NA        NA        NA        NA         NA
#> n30        NA         NA        NA        NA        NA        NA         NA
#>           n26       n27       n28        n29       n30
#> n1         NA        NA        NA         NA        NA
#> n2         NA        NA        NA 0.76239268        NA
#> n3         NA        NA        NA         NA        NA
#> n4         NA        NA        NA         NA        NA
#> n5         NA        NA        NA 0.77286447        NA
#> n6  0.8367166        NA        NA         NA        NA
#> n7         NA        NA        NA         NA        NA
#> n8  0.2988473        NA        NA         NA        NA
#> n9         NA        NA 0.4794390         NA        NA
#> n10        NA        NA        NA         NA        NA
#> n11        NA        NA        NA         NA        NA
#> n12        NA 0.5070436        NA 0.01305927        NA
#> n13        NA        NA        NA         NA        NA
#> n14        NA        NA        NA         NA        NA
#> n15        NA        NA        NA         NA        NA
#> n16        NA        NA        NA         NA        NA
#> n17        NA        NA 0.1750295         NA        NA
#> n18        NA        NA        NA         NA        NA
#> n19        NA        NA        NA         NA        NA
#> n20 0.2272130        NA        NA         NA        NA
#> n21        NA        NA        NA         NA        NA
#> n22        NA 0.5720856 0.6939986         NA        NA
#> n23        NA        NA        NA         NA        NA
#> n24        NA        NA        NA         NA        NA
#> n25        NA        NA        NA         NA        NA
#> n26        NA        NA        NA         NA 0.2448233
#> n27        NA        NA        NA         NA        NA
#> n28        NA        NA        NA         NA 0.9515476
#> n29        NA        NA        NA         NA 0.4686786
#> n30        NA        NA        NA         NA        NA
#> 
#> $pathway_pvalue
#> [1] 0.1991136
#> 
#> $deconfounding
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "dce"