Differential Causal Effects for negative binomial data

Shortcut for the main function to analyse negative binomial data

dce_nb(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver_args = list(method = "glm.dce.nb.fit", link = "identity"),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph

valid object defining a directed acyclic graph

df_expr_wt

data frame with wild type expression values

df_expr_mt

data from with mutation type expression values

solver_args

additional arguments for the solver function

adjustment_type

character string for the method to define the adjustment set Z for the regression

effect_type

method of computing causal effects

p_method

character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test

either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size

either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding

indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative

logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level

Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce_nb(dag,X.wt,X.mt)
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: alternation limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: alternation limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: alternation limit reached
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: iteration limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: iteration limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: iteration limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: iteration limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: iteration limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: alternation limit reached
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n2   0  0  1  0  0  0  0  0  0   0   1   0   0   1   0   1   1   1   0   0   0
#> n3   0  0  0  0  0  1  0  1  0   0   0   1   0   0   1   0   0   0   0   1   0
#> n4   0  0  0  0  0  1  0  0  0   0   1   0   0   0   0   0   0   0   0   0   0
#> n5   0  0  0  0  0  1  1  0  0   0   1   0   0   0   0   0   0   0   0   1   1
#> n6   0  0  0  0  0  0  1  0  0   0   0   1   0   0   0   0   0   0   1   0   0
#> n7   0  0  0  0  0  0  0  0  0   0   0   0   1   0   0   1   0   1   0   0   0
#> n8   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   0   1   0   0
#> n9   0  0  0  0  0  0  0  0  0   1   0   0   0   0   1   0   0   0   0   0   0
#> n10  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   1   0   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   1   0   0   0   0   0   0   0   0   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0   1   1   0
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    0   0   1   0   0   0   0   0   0
#> n2    0   1   0   0   0   0   0   0   0
#> n3    0   0   0   0   0   0   0   0   0
#> n4    1   0   0   0   0   0   0   0   0
#> n5    0   0   0   0   0   0   0   0   1
#> n6    0   0   1   0   0   0   0   0   0
#> n7    0   0   0   0   0   0   0   0   0
#> n8    0   0   0   0   0   0   0   0   0
#> n9    0   0   1   0   0   0   0   0   0
#> n10   0   0   0   1   0   1   0   0   0
#> n11   0   0   0   0   0   1   0   0   0
#> n12   1   1   0   0   0   0   1   0   0
#> n13   0   0   0   0   0   1   0   0   0
#> n14   0   0   0   0   0   0   0   0   0
#> n15   0   0   0   0   1   0   0   0   0
#> n16   0   0   0   0   0   0   0   0   0
#> n17   0   0   0   0   0   0   0   0   0
#> n18   0   0   0   0   0   0   1   0   0
#> n19   0   0   0   0   0   0   0   1   1
#> n20   0   0   0   0   0   0   0   0   0
#> n21   0   1   0   1   1   0   0   1   0
#> n22   0   0   0   0   0   0   0   0   0
#> n23   0   0   0   0   1   0   0   0   0
#> n24   0   0   0   1   0   0   0   0   0
#> n25   0   0   0   0   0   0   0   0   1
#> n26   0   0   0   0   0   0   1   0   0
#> n27   0   0   0   0   0   0   0   1   0
#> n28   0   0   0   0   0   0   0   0   0
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1 n2        n3 n4 n5          n6          n7          n8 n9           n10
#> n1  NA NA        NA NA NA          NA          NA -0.02738226 NA            NA
#> n2  NA NA 0.0088127 NA NA          NA          NA          NA NA            NA
#> n3  NA NA        NA NA NA -0.18881773          NA -0.67438966 NA            NA
#> n4  NA NA        NA NA NA -0.14728429          NA          NA NA            NA
#> n5  NA NA        NA NA NA  0.05726215  0.02761408          NA NA            NA
#> n6  NA NA        NA NA NA          NA -0.08496060          NA NA            NA
#> n7  NA NA        NA NA NA          NA          NA          NA NA            NA
#> n8  NA NA        NA NA NA          NA          NA          NA NA            NA
#> n9  NA NA        NA NA NA          NA          NA          NA NA -0.0008724893
#> n10 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n11 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n12 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n13 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n14 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n15 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n16 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n17 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n18 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n19 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n20 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n21 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n22 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n23 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n24 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n25 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n26 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n27 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n28 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n29 NA NA        NA NA NA          NA          NA          NA NA            NA
#> n30 NA NA        NA NA NA          NA          NA          NA NA            NA
#>             n11          n12         n13         n14         n15        n16
#> n1           NA           NA          NA          NA          NA         NA
#> n2   0.10907867           NA          NA 0.004303655          NA 0.01240363
#> n3           NA -0.131085918          NA          NA -0.80274514         NA
#> n4  -0.18989757           NA          NA          NA          NA         NA
#> n5  -0.09273651           NA          NA          NA          NA         NA
#> n6           NA -0.003502557          NA          NA          NA         NA
#> n7           NA           NA -0.04515884          NA          NA 0.06931141
#> n8           NA           NA          NA          NA          NA 0.31463169
#> n9           NA           NA          NA          NA -0.02481431         NA
#> n10          NA           NA          NA          NA          NA         NA
#> n11          NA -0.328290922          NA          NA          NA         NA
#> n12          NA           NA          NA          NA          NA         NA
#> n13          NA           NA          NA          NA -0.01231594         NA
#> n14          NA           NA          NA          NA          NA         NA
#> n15          NA           NA          NA          NA          NA         NA
#> n16          NA           NA          NA          NA          NA         NA
#> n17          NA           NA          NA          NA          NA         NA
#> n18          NA           NA          NA          NA          NA         NA
#> n19          NA           NA          NA          NA          NA         NA
#> n20          NA           NA          NA          NA          NA         NA
#> n21          NA           NA          NA          NA          NA         NA
#> n22          NA           NA          NA          NA          NA         NA
#> n23          NA           NA          NA          NA          NA         NA
#> n24          NA           NA          NA          NA          NA         NA
#> n25          NA           NA          NA          NA          NA         NA
#> n26          NA           NA          NA          NA          NA         NA
#> n27          NA           NA          NA          NA          NA         NA
#> n28          NA           NA          NA          NA          NA         NA
#> n29          NA           NA          NA          NA          NA         NA
#> n30          NA           NA          NA          NA          NA         NA
#>             n17          n18        n19        n20         n21        n22
#> n1           NA           NA         NA         NA          NA         NA
#> n2  0.003866256 -0.031799285         NA         NA          NA         NA
#> n3           NA           NA         NA  2.2085333          NA         NA
#> n4           NA           NA         NA         NA          NA  0.1127501
#> n5           NA           NA         NA -0.2231251  0.06013723         NA
#> n6           NA           NA  1.3128786         NA          NA         NA
#> n7           NA -0.005575709         NA         NA          NA         NA
#> n8           NA           NA -0.6470054         NA          NA         NA
#> n9           NA           NA         NA         NA          NA         NA
#> n10          NA -0.401774516         NA         NA          NA         NA
#> n11          NA           NA         NA         NA          NA         NA
#> n12          NA           NA         NA         NA          NA -0.2578272
#> n13          NA           NA -0.5793153  0.5775464          NA         NA
#> n14          NA           NA         NA         NA          NA         NA
#> n15          NA           NA         NA         NA          NA         NA
#> n16          NA           NA         NA         NA          NA         NA
#> n17          NA           NA         NA         NA          NA         NA
#> n18          NA           NA         NA         NA  0.30999814         NA
#> n19          NA           NA         NA         NA          NA         NA
#> n20          NA           NA         NA         NA -0.14794826         NA
#> n21          NA           NA         NA         NA          NA         NA
#> n22          NA           NA         NA         NA          NA         NA
#> n23          NA           NA         NA         NA          NA         NA
#> n24          NA           NA         NA         NA          NA         NA
#> n25          NA           NA         NA         NA          NA         NA
#> n26          NA           NA         NA         NA          NA         NA
#> n27          NA           NA         NA         NA          NA         NA
#> n28          NA           NA         NA         NA          NA         NA
#> n29          NA           NA         NA         NA          NA         NA
#> n30          NA           NA         NA         NA          NA         NA
#>             n23        n24        n25        n26        n27        n28
#> n1           NA 0.22365954         NA         NA         NA         NA
#> n2   0.04664931         NA         NA         NA         NA         NA
#> n3           NA         NA         NA         NA         NA         NA
#> n4           NA         NA         NA         NA         NA         NA
#> n5           NA         NA         NA         NA         NA         NA
#> n6           NA 4.48582967         NA         NA         NA         NA
#> n7           NA         NA         NA         NA         NA         NA
#> n8           NA         NA         NA         NA         NA         NA
#> n9           NA 0.05178065         NA         NA         NA         NA
#> n10          NA         NA  0.8759422         NA 0.11419378         NA
#> n11          NA         NA         NA         NA 0.02958922         NA
#> n12  0.06975122         NA         NA         NA         NA -0.7784104
#> n13          NA         NA         NA         NA 0.81681260         NA
#> n14          NA         NA         NA         NA         NA         NA
#> n15          NA         NA         NA  0.5290234         NA         NA
#> n16          NA         NA         NA         NA         NA         NA
#> n17          NA         NA         NA         NA         NA         NA
#> n18          NA         NA         NA         NA         NA -0.6720053
#> n19          NA         NA         NA         NA         NA         NA
#> n20          NA         NA         NA         NA         NA         NA
#> n21 -6.83328530         NA  0.0158291 -1.0238934         NA         NA
#> n22          NA         NA         NA         NA         NA         NA
#> n23          NA         NA         NA -0.8675064         NA         NA
#> n24          NA         NA -0.4617846         NA         NA         NA
#> n25          NA         NA         NA         NA         NA         NA
#> n26          NA         NA         NA         NA         NA  0.1037273
#> n27          NA         NA         NA         NA         NA         NA
#> n28          NA         NA         NA         NA         NA         NA
#> n29          NA         NA         NA         NA         NA         NA
#> n30          NA         NA         NA         NA         NA         NA
#>             n29         n30
#> n1           NA          NA
#> n2           NA          NA
#> n3           NA          NA
#> n4           NA          NA
#> n5           NA -0.03165255
#> n6           NA          NA
#> n7           NA          NA
#> n8           NA          NA
#> n9           NA          NA
#> n10          NA          NA
#> n11          NA          NA
#> n12          NA          NA
#> n13          NA          NA
#> n14          NA          NA
#> n15          NA          NA
#> n16          NA          NA
#> n17          NA          NA
#> n18          NA          NA
#> n19 -0.87991219 -0.97019266
#> n20          NA          NA
#> n21  0.09752226          NA
#> n22          NA          NA
#> n23          NA          NA
#> n24          NA          NA
#> n25          NA -0.02348865
#> n26          NA          NA
#> n27 -0.22011717          NA
#> n28          NA          NA
#> n29          NA          NA
#> n30          NA          NA
#> 
#> $dce_stderr
#>     n1 n2         n3 n4 n5        n6         n7        n8 n9         n10
#> n1  NA NA         NA NA NA        NA         NA 0.0222475 NA          NA
#> n2  NA NA 0.00960068 NA NA        NA         NA        NA NA          NA
#> n3  NA NA         NA NA NA 0.6785965         NA 0.7727634 NA          NA
#> n4  NA NA         NA NA NA 0.1664257         NA        NA NA          NA
#> n5  NA NA         NA NA NA 0.1839772 0.04467138        NA NA          NA
#> n6  NA NA         NA NA NA        NA 0.10916306        NA NA          NA
#> n7  NA NA         NA NA NA        NA         NA        NA NA          NA
#> n8  NA NA         NA NA NA        NA         NA        NA NA          NA
#> n9  NA NA         NA NA NA        NA         NA        NA NA 0.005388163
#> n10 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n11 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n12 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n13 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n14 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n15 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n16 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n17 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n18 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n19 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n20 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n21 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n22 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n23 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n24 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n25 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n26 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n27 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n28 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n29 NA NA         NA NA NA        NA         NA        NA NA          NA
#> n30 NA NA         NA NA NA        NA         NA        NA NA          NA
#>           n11       n12       n13         n14       n15       n16         n17
#> n1         NA        NA        NA          NA        NA        NA          NA
#> n2  0.2147199        NA        NA 0.008006556        NA 0.1053752 0.005026465
#> n3         NA 1.4682848        NA          NA 0.6589119        NA          NA
#> n4  0.1209145        NA        NA          NA        NA        NA          NA
#> n5  0.1486712        NA        NA          NA        NA        NA          NA
#> n6         NA 0.1240761        NA          NA        NA        NA          NA
#> n7         NA        NA 0.1107101          NA        NA 0.1231849          NA
#> n8         NA        NA        NA          NA        NA 0.7898200          NA
#> n9         NA        NA        NA          NA 0.1150088        NA          NA
#> n10        NA        NA        NA          NA        NA        NA          NA
#> n11        NA 0.1662758        NA          NA        NA        NA          NA
#> n12        NA        NA        NA          NA        NA        NA          NA
#> n13        NA        NA        NA          NA 0.1207264        NA          NA
#> n14        NA        NA        NA          NA        NA        NA          NA
#> n15        NA        NA        NA          NA        NA        NA          NA
#> n16        NA        NA        NA          NA        NA        NA          NA
#> n17        NA        NA        NA          NA        NA        NA          NA
#> n18        NA        NA        NA          NA        NA        NA          NA
#> n19        NA        NA        NA          NA        NA        NA          NA
#> n20        NA        NA        NA          NA        NA        NA          NA
#> n21        NA        NA        NA          NA        NA        NA          NA
#> n22        NA        NA        NA          NA        NA        NA          NA
#> n23        NA        NA        NA          NA        NA        NA          NA
#> n24        NA        NA        NA          NA        NA        NA          NA
#> n25        NA        NA        NA          NA        NA        NA          NA
#> n26        NA        NA        NA          NA        NA        NA          NA
#> n27        NA        NA        NA          NA        NA        NA          NA
#> n28        NA        NA        NA          NA        NA        NA          NA
#> n29        NA        NA        NA          NA        NA        NA          NA
#> n30        NA        NA        NA          NA        NA        NA          NA
#>            n18       n19        n20        n21        n22        n23       n24
#> n1          NA        NA         NA         NA         NA         NA 0.2447596
#> n2  0.07170379        NA         NA         NA         NA 0.08221553        NA
#> n3          NA        NA 1.49356880         NA         NA         NA        NA
#> n4          NA        NA         NA         NA 0.07383132         NA        NA
#> n5          NA        NA 0.06246175 0.04906267         NA         NA        NA
#> n6          NA 0.4478727         NA         NA         NA         NA 0.7600794
#> n7  0.19704723        NA         NA         NA         NA         NA        NA
#> n8          NA 0.8884917         NA         NA         NA         NA        NA
#> n9          NA        NA         NA         NA         NA         NA 0.2152666
#> n10 0.72110696        NA         NA         NA         NA         NA        NA
#> n11         NA        NA         NA         NA         NA         NA        NA
#> n12         NA        NA         NA         NA 0.23408763 0.25204528        NA
#> n13         NA 0.7694441 1.24787101         NA         NA         NA        NA
#> n14         NA        NA         NA         NA         NA         NA        NA
#> n15         NA        NA         NA         NA         NA         NA        NA
#> n16         NA        NA         NA         NA         NA         NA        NA
#> n17         NA        NA         NA         NA         NA         NA        NA
#> n18         NA        NA         NA 0.21864244         NA         NA        NA
#> n19         NA        NA         NA         NA         NA         NA        NA
#> n20         NA        NA         NA 0.21550399         NA         NA        NA
#> n21         NA        NA         NA         NA         NA 1.26547915        NA
#> n22         NA        NA         NA         NA         NA         NA        NA
#> n23         NA        NA         NA         NA         NA         NA        NA
#> n24         NA        NA         NA         NA         NA         NA        NA
#> n25         NA        NA         NA         NA         NA         NA        NA
#> n26         NA        NA         NA         NA         NA         NA        NA
#> n27         NA        NA         NA         NA         NA         NA        NA
#> n28         NA        NA         NA         NA         NA         NA        NA
#> n29         NA        NA         NA         NA         NA         NA        NA
#> n30         NA        NA         NA         NA         NA         NA        NA
#>           n25       n26       n27       n28       n29        n30
#> n1         NA        NA        NA        NA        NA         NA
#> n2         NA        NA        NA        NA        NA         NA
#> n3         NA        NA        NA        NA        NA         NA
#> n4         NA        NA        NA        NA        NA         NA
#> n5         NA        NA        NA        NA        NA 0.05094706
#> n6         NA        NA        NA        NA        NA         NA
#> n7         NA        NA        NA        NA        NA         NA
#> n8         NA        NA        NA        NA        NA         NA
#> n9         NA        NA        NA        NA        NA         NA
#> n10 0.6476525        NA 1.8860293        NA        NA         NA
#> n11        NA        NA 0.1632392        NA        NA         NA
#> n12        NA        NA        NA 0.1385985        NA         NA
#> n13        NA        NA 0.5276047        NA        NA         NA
#> n14        NA        NA        NA        NA        NA         NA
#> n15        NA 0.4063350        NA        NA        NA         NA
#> n16        NA        NA        NA        NA        NA         NA
#> n17        NA        NA        NA        NA        NA         NA
#> n18        NA        NA        NA 1.6042917        NA         NA
#> n19        NA        NA        NA        NA 0.5101231 0.45591394
#> n20        NA        NA        NA        NA        NA         NA
#> n21 0.4482861 0.5290770        NA        NA 0.8128469         NA
#> n22        NA        NA        NA        NA        NA         NA
#> n23        NA 0.5367777        NA        NA        NA         NA
#> n24 0.2414521        NA        NA        NA        NA         NA
#> n25        NA        NA        NA        NA        NA 0.44138909
#> n26        NA        NA        NA 0.2267238        NA         NA
#> n27        NA        NA        NA        NA 0.1240261         NA
#> n28        NA        NA        NA        NA        NA         NA
#> n29        NA        NA        NA        NA        NA         NA
#> n30        NA        NA        NA        NA        NA         NA
#> 
#> $dce_pvalue
#>     n1 n2        n3 n4 n5        n6        n7        n8 n9       n10       n11
#> n1  NA NA        NA NA NA        NA        NA 0.2183970 NA        NA        NA
#> n2  NA NA 0.3586584 NA NA        NA        NA        NA NA        NA 0.6114503
#> n3  NA NA        NA NA NA 0.7808224        NA 0.3828273 NA        NA        NA
#> n4  NA NA        NA NA NA 0.3761646        NA        NA NA        NA 0.1162963
#> n5  NA NA        NA NA NA 0.7556137 0.5364696        NA NA        NA 0.5327792
#> n6  NA NA        NA NA NA        NA 0.4363976        NA NA        NA        NA
#> n7  NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n8  NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n9  NA NA        NA NA NA        NA        NA        NA NA 0.8713633        NA
#> n10 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n11 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n12 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n13 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n14 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n15 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n16 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n17 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n18 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n19 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n20 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n21 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n22 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n23 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n24 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n25 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n26 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n27 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n28 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n29 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#> n30 NA NA        NA NA NA        NA        NA        NA NA        NA        NA
#>            n12       n13       n14       n15       n16       n17       n18
#> n1          NA        NA        NA        NA        NA        NA        NA
#> n2          NA        NA 0.5909109        NA 0.9062981 0.4417865 0.6574177
#> n3  0.92886077        NA        NA 0.2231142        NA        NA        NA
#> n4          NA        NA        NA        NA        NA        NA        NA
#> n5          NA        NA        NA        NA        NA        NA        NA
#> n6  0.97747943        NA        NA        NA        NA        NA        NA
#> n7          NA 0.6833459        NA        NA 0.5736654        NA 0.9774258
#> n8          NA        NA        NA        NA 0.6903658        NA        NA
#> n9          NA        NA        NA 0.8291747        NA        NA        NA
#> n10         NA        NA        NA        NA        NA        NA 0.5774157
#> n11 0.04833903        NA        NA        NA        NA        NA        NA
#> n12         NA        NA        NA        NA        NA        NA        NA
#> n13         NA        NA        NA 0.9187446        NA        NA        NA
#> n14         NA        NA        NA        NA        NA        NA        NA
#> n15         NA        NA        NA        NA        NA        NA        NA
#> n16         NA        NA        NA        NA        NA        NA        NA
#> n17         NA        NA        NA        NA        NA        NA        NA
#> n18         NA        NA        NA        NA        NA        NA        NA
#> n19         NA        NA        NA        NA        NA        NA        NA
#> n20         NA        NA        NA        NA        NA        NA        NA
#> n21         NA        NA        NA        NA        NA        NA        NA
#> n22         NA        NA        NA        NA        NA        NA        NA
#> n23         NA        NA        NA        NA        NA        NA        NA
#> n24         NA        NA        NA        NA        NA        NA        NA
#> n25         NA        NA        NA        NA        NA        NA        NA
#> n26         NA        NA        NA        NA        NA        NA        NA
#> n27         NA        NA        NA        NA        NA        NA        NA
#> n28         NA        NA        NA        NA        NA        NA        NA
#> n29         NA        NA        NA        NA        NA        NA        NA
#> n30         NA        NA        NA        NA        NA        NA        NA
#>            n19          n20       n21       n22          n23          n24
#> n1          NA           NA        NA        NA           NA 3.608257e-01
#> n2          NA           NA        NA        NA 5.704406e-01           NA
#> n3          NA 0.1392217515        NA        NA           NA           NA
#> n4          NA           NA        NA 0.1267285           NA           NA
#> n5          NA 0.0003540107 0.2203030        NA           NA           NA
#> n6  0.00337475           NA        NA        NA           NA 3.595764e-09
#> n7          NA           NA        NA        NA           NA           NA
#> n8  0.46648723           NA        NA        NA           NA           NA
#> n9          NA           NA        NA        NA           NA 8.099101e-01
#> n10         NA           NA        NA        NA           NA           NA
#> n11         NA           NA        NA        NA           NA           NA
#> n12         NA           NA        NA 0.2707170 7.819791e-01           NA
#> n13 0.45150933 0.6434895181        NA        NA           NA           NA
#> n14         NA           NA        NA        NA           NA           NA
#> n15         NA           NA        NA        NA           NA           NA
#> n16         NA           NA        NA        NA           NA           NA
#> n17         NA           NA        NA        NA           NA           NA
#> n18         NA           NA 0.1562400        NA           NA           NA
#> n19         NA           NA        NA        NA           NA           NA
#> n20         NA           NA 0.4923839        NA           NA           NA
#> n21         NA           NA        NA        NA 6.672963e-08           NA
#> n22         NA           NA        NA        NA           NA           NA
#> n23         NA           NA        NA        NA           NA           NA
#> n24         NA           NA        NA        NA           NA           NA
#> n25         NA           NA        NA        NA           NA           NA
#> n26         NA           NA        NA        NA           NA           NA
#> n27         NA           NA        NA        NA           NA           NA
#> n28         NA           NA        NA        NA           NA           NA
#> n29         NA           NA        NA        NA           NA           NA
#> n30         NA           NA        NA        NA           NA           NA
#>            n25        n26       n27          n28        n29        n30
#> n1          NA         NA        NA           NA         NA         NA
#> n2          NA         NA        NA           NA         NA         NA
#> n3          NA         NA        NA           NA         NA         NA
#> n4          NA         NA        NA           NA         NA         NA
#> n5          NA         NA        NA           NA         NA 0.53441336
#> n6          NA         NA        NA           NA         NA         NA
#> n7          NA         NA        NA           NA         NA         NA
#> n8          NA         NA        NA           NA         NA         NA
#> n9          NA         NA        NA           NA         NA         NA
#> n10 0.17621928         NA 0.9517198           NA         NA         NA
#> n11         NA         NA 0.8561612           NA         NA         NA
#> n12         NA         NA        NA 1.950917e-08         NA         NA
#> n13         NA         NA 0.1215856           NA         NA         NA
#> n14         NA         NA        NA           NA         NA         NA
#> n15         NA 0.19293724        NA           NA         NA         NA
#> n16         NA         NA        NA           NA         NA         NA
#> n17         NA         NA        NA           NA         NA         NA
#> n18         NA         NA        NA 6.753040e-01         NA         NA
#> n19         NA         NA        NA           NA 0.08454518 0.03333567
#> n20         NA         NA        NA           NA         NA         NA
#> n21 0.97183234 0.05296031        NA           NA 0.90450203         NA
#> n22         NA         NA        NA           NA         NA         NA
#> n23         NA 0.10606465        NA           NA         NA         NA
#> n24 0.05580811         NA        NA           NA         NA         NA
#> n25         NA         NA        NA           NA         NA 0.95756037
#> n26         NA         NA        NA 6.473080e-01         NA         NA
#> n27         NA         NA        NA           NA 0.07593678         NA
#> n28         NA         NA        NA           NA         NA         NA
#> n29         NA         NA        NA           NA         NA         NA
#> n30         NA         NA        NA           NA         NA         NA
#> 
#> $pathway_pvalue
#> [1] 1.800489e-07
#> 
#> $deconfounding
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "dce"