dce - main function

Main function to compute differential causal effects and its enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = "glm.dce.fit"),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  num_latent = "auto",
  conservative = FALSE,
  verbose = FALSE
)

Arguments

graph
valid object defining a directed acyclic graph

df_expr_wt
data frame with wild type expression values

df_expr_mt
data from with mutation type expression values

solver
character with name of solver function

solver_args
additional arguments for the solver function

adjustment_type
character string for the method to define the adjustment set Z for the regression

effect_type
method of computing causal effects

p_method
character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test
either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size
either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

verbose
logical for verbose output