Differential Causal Effects - main function

Main function to compute differential causal effects and the pathway enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for igraph
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for graphNEL
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for matrix
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph
valid object defining a directed acyclic graph

df_expr_wt
data frame with wild type expression values

df_expr_mt
data from with mutation type expression values

solver
character with name of solver function

solver_args
additional arguments for the solver function

adjustment_type
character string for the method to define the adjustment set Z for the regression

effect_type
method of computing causal effects

p_method
character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test
either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size
either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding
indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative
logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level
Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,X.mt)

#> Warning: Cycle(s) detected in network

#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  1  1  1  0  0  1  0  0   0   0   0   1   0   0   0   0   1   0   0   0
#> n2   0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n3   0  0  0  0  1  1  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n4   0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   0   1   0   0   0   0
#> n5   0  0  0  0  0  0  0  0  1   1   1   0   0   0   0   0   1   0   1   0   0
#> n6   0  0  0  0  0  0  1  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n7   0  0  0  0  0  0  0  0  0   1   1   0   0   0   1   0   0   0   0   0   1
#> n8   0  0  0  0  0  0  0  0  0   0   0   0   1   0   0   0   0   0   0   1   0
#> n9   0  0  0  0  0  0  0  0  0   0   0   1   0   0   1   0   0   0   0   0   0
#> n10  0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   0   0   0   0   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   1   0   0   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   1   0   0
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    0   0   0   1   0   0   0   0   0
#> n2    0   0   1   0   0   0   0   0   0
#> n3    0   0   0   0   0   0   0   0   0
#> n4    0   0   0   0   0   0   0   0   0
#> n5    0   0   1   0   0   0   0   0   0
#> n6    0   0   0   0   0   0   0   0   0
#> n7    1   0   0   0   0   0   0   0   0
#> n8    1   0   0   0   0   0   0   0   0
#> n9    0   0   0   0   0   0   0   0   0
#> n10   0   0   0   1   0   0   1   0   0
#> n11   0   0   0   0   1   0   0   0   0
#> n12   0   0   0   0   0   0   0   0   0
#> n13   0   1   0   0   0   0   0   0   0
#> n14   0   0   0   0   0   0   0   0   0
#> n15   0   0   0   0   0   0   0   0   0
#> n16   0   1   0   0   0   0   0   1   0
#> n17   0   0   0   0   0   0   0   0   0
#> n18   0   0   0   0   0   0   0   0   1
#> n19   0   0   0   0   0   0   0   0   0
#> n20   0   0   0   0   0   0   0   0   0
#> n21   0   0   0   0   0   0   0   0   0
#> n22   0   0   0   0   1   0   0   0   1
#> n23   0   0   0   0   0   0   0   1   0
#> n24   0   0   0   0   0   0   0   0   0
#> n25   0   0   0   0   0   1   0   0   0
#> n26   0   0   0   0   0   1   1   0   0
#> n27   0   0   0   0   0   0   0   0   0
#> n28   0   0   0   0   0   0   0   0   0
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1          n2         n3            n4         n5          n6
#> n1  NA 0.004801211 0.01888289 -0.0005098078         NA          NA
#> n2  NA          NA         NA            NA         NA          NA
#> n3  NA          NA         NA            NA -0.1859576 -0.01232321
#> n4  NA          NA         NA            NA         NA          NA
#> n5  NA          NA         NA            NA         NA          NA
#> n6  NA          NA         NA            NA         NA          NA
#> n7  NA          NA         NA            NA         NA          NA
#> n8  NA          NA         NA            NA         NA          NA
#> n9  NA          NA         NA            NA         NA          NA
#> n10 NA          NA         NA            NA         NA          NA
#> n11 NA          NA         NA            NA         NA          NA
#> n12 NA          NA         NA            NA         NA          NA
#> n13 NA          NA         NA            NA         NA          NA
#> n14 NA          NA         NA            NA         NA          NA
#> n15 NA          NA         NA            NA         NA          NA
#> n16 NA          NA         NA            NA         NA          NA
#> n17 NA          NA         NA            NA         NA          NA
#> n18 NA          NA         NA            NA         NA          NA
#> n19 NA          NA         NA            NA         NA          NA
#> n20 NA          NA         NA            NA         NA          NA
#> n21 NA          NA         NA            NA         NA          NA
#> n22 NA          NA         NA            NA         NA          NA
#> n23 NA          NA         NA            NA         NA          NA
#> n24 NA          NA         NA            NA         NA          NA
#> n25 NA          NA         NA            NA         NA          NA
#> n26 NA          NA         NA            NA         NA          NA
#> n27 NA          NA         NA            NA         NA          NA
#> n28 NA          NA         NA            NA         NA          NA
#> n29 NA          NA         NA            NA         NA          NA
#> n30 NA          NA         NA            NA         NA          NA
#>                n7         n8         n9        n10       n11         n12
#> n1   0.0005256392         NA         NA         NA        NA          NA
#> n2             NA 0.02063111         NA         NA        NA          NA
#> n3             NA         NA         NA         NA        NA          NA
#> n4             NA         NA         NA         NA        NA          NA
#> n5             NA         NA -0.1151341  0.4865857 0.2362119          NA
#> n6  -0.0902444229         NA         NA         NA        NA          NA
#> n7             NA         NA         NA -0.2317294 0.2205781          NA
#> n8             NA         NA         NA         NA        NA          NA
#> n9             NA         NA         NA         NA        NA -0.06881065
#> n10            NA         NA         NA         NA        NA          NA
#> n11            NA         NA         NA         NA        NA          NA
#> n12            NA         NA         NA         NA        NA          NA
#> n13            NA         NA         NA         NA        NA          NA
#> n14            NA         NA         NA         NA        NA          NA
#> n15            NA         NA         NA         NA        NA          NA
#> n16            NA         NA         NA         NA        NA          NA
#> n17            NA         NA         NA         NA        NA          NA
#> n18            NA         NA         NA         NA        NA          NA
#> n19            NA         NA         NA         NA        NA          NA
#> n20            NA         NA         NA         NA        NA          NA
#> n21            NA         NA         NA         NA        NA          NA
#> n22            NA         NA         NA         NA        NA          NA
#> n23            NA         NA         NA         NA        NA          NA
#> n24            NA         NA         NA         NA        NA          NA
#> n25            NA         NA         NA         NA        NA          NA
#> n26            NA         NA         NA         NA        NA          NA
#> n27            NA         NA         NA         NA        NA          NA
#> n28            NA         NA         NA         NA        NA          NA
#> n29            NA         NA         NA         NA        NA          NA
#> n30            NA         NA         NA         NA        NA          NA
#>             n13       n14          n15        n16        n17          n18
#> n1  0.001270465        NA           NA         NA         NA  0.000511614
#> n2           NA        NA           NA         NA         NA           NA
#> n3           NA        NA           NA         NA         NA           NA
#> n4           NA 0.3929691           NA         NA 0.04251925           NA
#> n5           NA        NA           NA         NA 0.10430518           NA
#> n6           NA        NA           NA         NA         NA           NA
#> n7           NA        NA -0.001863474         NA         NA           NA
#> n8  0.466912093        NA           NA         NA         NA           NA
#> n9           NA        NA  0.456829830         NA         NA           NA
#> n10          NA 0.0280132           NA         NA         NA           NA
#> n11          NA        NA           NA         NA         NA           NA
#> n12          NA        NA           NA         NA         NA           NA
#> n13          NA        NA           NA         NA         NA           NA
#> n14          NA        NA           NA         NA         NA           NA
#> n15          NA        NA           NA 0.08549674         NA -0.284899551
#> n16          NA        NA           NA         NA         NA           NA
#> n17          NA        NA           NA         NA         NA           NA
#> n18          NA        NA           NA         NA         NA           NA
#> n19          NA        NA           NA         NA         NA           NA
#> n20          NA        NA           NA         NA         NA           NA
#> n21          NA        NA           NA         NA         NA           NA
#> n22          NA        NA           NA         NA         NA           NA
#> n23          NA        NA           NA         NA         NA           NA
#> n24          NA        NA           NA         NA         NA           NA
#> n25          NA        NA           NA         NA         NA           NA
#> n26          NA        NA           NA         NA         NA           NA
#> n27          NA        NA           NA         NA         NA           NA
#> n28          NA        NA           NA         NA         NA           NA
#> n29          NA        NA           NA         NA         NA           NA
#> n30          NA        NA           NA         NA         NA           NA
#>            n19         n20        n21         n22       n23        n24
#> n1          NA          NA         NA          NA        NA         NA
#> n2          NA          NA         NA          NA        NA 0.14199810
#> n3          NA          NA         NA          NA        NA         NA
#> n4          NA          NA         NA          NA        NA         NA
#> n5  1.27931440          NA         NA          NA        NA 0.09047807
#> n6          NA -0.07140263         NA          NA        NA         NA
#> n7          NA          NA -0.2319674  0.01200753        NA         NA
#> n8          NA -0.13044601         NA -0.13906702        NA         NA
#> n9          NA          NA         NA          NA        NA         NA
#> n10         NA          NA         NA          NA        NA         NA
#> n11         NA          NA         NA          NA        NA         NA
#> n12         NA          NA -0.9922520          NA        NA         NA
#> n13         NA          NA         NA          NA 0.2413826         NA
#> n14         NA          NA         NA          NA        NA         NA
#> n15         NA          NA         NA          NA        NA         NA
#> n16         NA          NA         NA          NA 0.8824015         NA
#> n17         NA          NA         NA          NA        NA         NA
#> n18 0.05537496          NA         NA          NA        NA         NA
#> n19         NA          NA         NA          NA        NA         NA
#> n20         NA          NA         NA          NA        NA         NA
#> n21         NA          NA         NA          NA        NA         NA
#> n22         NA          NA         NA          NA        NA         NA
#> n23         NA          NA         NA          NA        NA         NA
#> n24         NA          NA         NA          NA        NA         NA
#> n25         NA          NA         NA          NA        NA         NA
#> n26         NA          NA         NA          NA        NA         NA
#> n27         NA          NA         NA          NA        NA         NA
#> n28         NA          NA         NA          NA        NA         NA
#> n29         NA          NA         NA          NA        NA         NA
#> n30         NA          NA         NA          NA        NA         NA
#>             n25        n26         n27         n28         n29       n30
#> n1  -0.01935107         NA          NA          NA          NA        NA
#> n2           NA         NA          NA          NA          NA        NA
#> n3           NA         NA          NA          NA          NA        NA
#> n4           NA         NA          NA          NA          NA        NA
#> n5           NA         NA          NA          NA          NA        NA
#> n6           NA         NA          NA          NA          NA        NA
#> n7           NA         NA          NA          NA          NA        NA
#> n8           NA         NA          NA          NA          NA        NA
#> n9           NA         NA          NA          NA          NA        NA
#> n10  0.04315021         NA          NA  0.02656044          NA        NA
#> n11          NA  0.3064765          NA          NA          NA        NA
#> n12          NA         NA          NA          NA          NA        NA
#> n13          NA         NA          NA          NA          NA        NA
#> n14          NA         NA          NA          NA          NA        NA
#> n15          NA         NA          NA          NA          NA        NA
#> n16          NA         NA          NA          NA -0.45554004        NA
#> n17          NA         NA          NA          NA          NA        NA
#> n18          NA         NA          NA          NA          NA 0.1453164
#> n19          NA         NA          NA          NA          NA        NA
#> n20          NA         NA          NA          NA          NA        NA
#> n21          NA         NA          NA          NA          NA        NA
#> n22          NA -0.7587991          NA          NA          NA 2.4838374
#> n23          NA         NA          NA          NA -0.05731215        NA
#> n24          NA         NA          NA          NA          NA        NA
#> n25          NA         NA -0.05082166          NA          NA        NA
#> n26          NA         NA -0.23440062 -0.73838467          NA        NA
#> n27          NA         NA          NA          NA          NA        NA
#> n28          NA         NA          NA          NA          NA        NA
#> n29          NA         NA          NA          NA          NA        NA
#> n30          NA         NA          NA          NA          NA        NA
#> 
#> $dce_stderr
#>     n1          n2         n3         n4        n5         n6         n7
#> n1  NA 0.009128126 0.01360551 0.01265482        NA         NA 0.01743288
#> n2  NA          NA         NA         NA        NA         NA         NA
#> n3  NA          NA         NA         NA 0.1110358 0.06212054         NA
#> n4  NA          NA         NA         NA        NA         NA         NA
#> n5  NA          NA         NA         NA        NA         NA         NA
#> n6  NA          NA         NA         NA        NA         NA 0.68876129
#> n7  NA          NA         NA         NA        NA         NA         NA
#> n8  NA          NA         NA         NA        NA         NA         NA
#> n9  NA          NA         NA         NA        NA         NA         NA
#> n10 NA          NA         NA         NA        NA         NA         NA
#> n11 NA          NA         NA         NA        NA         NA         NA
#> n12 NA          NA         NA         NA        NA         NA         NA
#> n13 NA          NA         NA         NA        NA         NA         NA
#> n14 NA          NA         NA         NA        NA         NA         NA
#> n15 NA          NA         NA         NA        NA         NA         NA
#> n16 NA          NA         NA         NA        NA         NA         NA
#> n17 NA          NA         NA         NA        NA         NA         NA
#> n18 NA          NA         NA         NA        NA         NA         NA
#> n19 NA          NA         NA         NA        NA         NA         NA
#> n20 NA          NA         NA         NA        NA         NA         NA
#> n21 NA          NA         NA         NA        NA         NA         NA
#> n22 NA          NA         NA         NA        NA         NA         NA
#> n23 NA          NA         NA         NA        NA         NA         NA
#> n24 NA          NA         NA         NA        NA         NA         NA
#> n25 NA          NA         NA         NA        NA         NA         NA
#> n26 NA          NA         NA         NA        NA         NA         NA
#> n27 NA          NA         NA         NA        NA         NA         NA
#> n28 NA          NA         NA         NA        NA         NA         NA
#> n29 NA          NA         NA         NA        NA         NA         NA
#> n30 NA          NA         NA         NA        NA         NA         NA
#>             n8         n9       n10        n11       n12        n13       n14
#> n1          NA         NA        NA         NA        NA 0.01426505        NA
#> n2  0.05241441         NA        NA         NA        NA         NA        NA
#> n3          NA         NA        NA         NA        NA         NA        NA
#> n4          NA         NA        NA         NA        NA         NA 0.2879054
#> n5          NA 0.08803854 0.4126977 0.24218826        NA         NA        NA
#> n6          NA         NA        NA         NA        NA         NA        NA
#> n7          NA         NA 0.1891828 0.09482059        NA         NA        NA
#> n8          NA         NA        NA         NA        NA 1.02842907        NA
#> n9          NA         NA        NA         NA 0.1177371         NA        NA
#> n10         NA         NA        NA         NA        NA         NA 0.1428439
#> n11         NA         NA        NA         NA        NA         NA        NA
#> n12         NA         NA        NA         NA        NA         NA        NA
#> n13         NA         NA        NA         NA        NA         NA        NA
#> n14         NA         NA        NA         NA        NA         NA        NA
#> n15         NA         NA        NA         NA        NA         NA        NA
#> n16         NA         NA        NA         NA        NA         NA        NA
#> n17         NA         NA        NA         NA        NA         NA        NA
#> n18         NA         NA        NA         NA        NA         NA        NA
#> n19         NA         NA        NA         NA        NA         NA        NA
#> n20         NA         NA        NA         NA        NA         NA        NA
#> n21         NA         NA        NA         NA        NA         NA        NA
#> n22         NA         NA        NA         NA        NA         NA        NA
#> n23         NA         NA        NA         NA        NA         NA        NA
#> n24         NA         NA        NA         NA        NA         NA        NA
#> n25         NA         NA        NA         NA        NA         NA        NA
#> n26         NA         NA        NA         NA        NA         NA        NA
#> n27         NA         NA        NA         NA        NA         NA        NA
#> n28         NA         NA        NA         NA        NA         NA        NA
#> n29         NA         NA        NA         NA        NA         NA        NA
#> n30         NA         NA        NA         NA        NA         NA        NA
#>           n15        n16       n17        n18       n19       n20       n21
#> n1         NA         NA        NA 0.01957965        NA        NA        NA
#> n2         NA         NA        NA         NA        NA        NA        NA
#> n3         NA         NA        NA         NA        NA        NA        NA
#> n4         NA         NA 0.1246039         NA        NA        NA        NA
#> n5         NA         NA 0.1353492         NA 0.4264707        NA        NA
#> n6         NA         NA        NA         NA        NA 0.1398330        NA
#> n7  0.1402729         NA        NA         NA        NA        NA 0.1359119
#> n8         NA         NA        NA         NA        NA 0.3434875        NA
#> n9  0.7017919         NA        NA         NA        NA        NA        NA
#> n10        NA         NA        NA         NA        NA        NA        NA
#> n11        NA         NA        NA         NA        NA        NA        NA
#> n12        NA         NA        NA         NA        NA        NA 2.1338490
#> n13        NA         NA        NA         NA        NA        NA        NA
#> n14        NA         NA        NA         NA        NA        NA        NA
#> n15        NA 0.05960254        NA 0.29713513        NA        NA        NA
#> n16        NA         NA        NA         NA        NA        NA        NA
#> n17        NA         NA        NA         NA        NA        NA        NA
#> n18        NA         NA        NA         NA 0.1628528        NA        NA
#> n19        NA         NA        NA         NA        NA        NA        NA
#> n20        NA         NA        NA         NA        NA        NA        NA
#> n21        NA         NA        NA         NA        NA        NA        NA
#> n22        NA         NA        NA         NA        NA        NA        NA
#> n23        NA         NA        NA         NA        NA        NA        NA
#> n24        NA         NA        NA         NA        NA        NA        NA
#> n25        NA         NA        NA         NA        NA        NA        NA
#> n26        NA         NA        NA         NA        NA        NA        NA
#> n27        NA         NA        NA         NA        NA        NA        NA
#> n28        NA         NA        NA         NA        NA        NA        NA
#> n29        NA         NA        NA         NA        NA        NA        NA
#> n30        NA         NA        NA         NA        NA        NA        NA
#>            n22       n23       n24       n25       n26        n27       n28
#> n1          NA        NA        NA 0.0250694        NA         NA        NA
#> n2          NA        NA 0.2105981        NA        NA         NA        NA
#> n3          NA        NA        NA        NA        NA         NA        NA
#> n4          NA        NA        NA        NA        NA         NA        NA
#> n5          NA        NA 0.1954341        NA        NA         NA        NA
#> n6          NA        NA        NA        NA        NA         NA        NA
#> n7  0.04038385        NA        NA        NA        NA         NA        NA
#> n8  0.19560356        NA        NA        NA        NA         NA        NA
#> n9          NA        NA        NA        NA        NA         NA        NA
#> n10         NA        NA        NA 0.2306218        NA         NA 0.1104180
#> n11         NA        NA        NA        NA 0.1466812         NA        NA
#> n12         NA        NA        NA        NA        NA         NA        NA
#> n13         NA 0.1736434        NA        NA        NA         NA        NA
#> n14         NA        NA        NA        NA        NA         NA        NA
#> n15         NA        NA        NA        NA        NA         NA        NA
#> n16         NA 0.5970325        NA        NA        NA         NA        NA
#> n17         NA        NA        NA        NA        NA         NA        NA
#> n18         NA        NA        NA        NA        NA         NA        NA
#> n19         NA        NA        NA        NA        NA         NA        NA
#> n20         NA        NA        NA        NA        NA         NA        NA
#> n21         NA        NA        NA        NA        NA         NA        NA
#> n22         NA        NA        NA        NA 0.4521893         NA        NA
#> n23         NA        NA        NA        NA        NA         NA        NA
#> n24         NA        NA        NA        NA        NA         NA        NA
#> n25         NA        NA        NA        NA        NA 0.05671833        NA
#> n26         NA        NA        NA        NA        NA 0.12416439 0.5201244
#> n27         NA        NA        NA        NA        NA         NA        NA
#> n28         NA        NA        NA        NA        NA         NA        NA
#> n29         NA        NA        NA        NA        NA         NA        NA
#> n30         NA        NA        NA        NA        NA         NA        NA
#>           n29       n30
#> n1         NA        NA
#> n2         NA        NA
#> n3         NA        NA
#> n4         NA        NA
#> n5         NA        NA
#> n6         NA        NA
#> n7         NA        NA
#> n8         NA        NA
#> n9         NA        NA
#> n10        NA        NA
#> n11        NA        NA
#> n12        NA        NA
#> n13        NA        NA
#> n14        NA        NA
#> n15        NA        NA
#> n16 0.4750409        NA
#> n17        NA        NA
#> n18        NA 0.1612585
#> n19        NA        NA
#> n20        NA        NA
#> n21        NA        NA
#> n22        NA 1.3869835
#> n23 0.1205905        NA
#> n24        NA        NA
#> n25        NA        NA
#> n26        NA        NA
#> n27        NA        NA
#> n28        NA        NA
#> n29        NA        NA
#> n30        NA        NA
#> 
#> $dce_pvalue
#>     n1        n2        n3        n4         n5        n6        n7        n8
#> n1  NA 0.5994968 0.1667476 0.9679064         NA        NA 0.9759764        NA
#> n2  NA        NA        NA        NA         NA        NA        NA 0.6942975
#> n3  NA        NA        NA        NA 0.09559393 0.8429588        NA        NA
#> n4  NA        NA        NA        NA         NA        NA        NA        NA
#> n5  NA        NA        NA        NA         NA        NA        NA        NA
#> n6  NA        NA        NA        NA         NA        NA 0.8958919        NA
#> n7  NA        NA        NA        NA         NA        NA        NA        NA
#> n8  NA        NA        NA        NA         NA        NA        NA        NA
#> n9  NA        NA        NA        NA         NA        NA        NA        NA
#> n10 NA        NA        NA        NA         NA        NA        NA        NA
#> n11 NA        NA        NA        NA         NA        NA        NA        NA
#> n12 NA        NA        NA        NA         NA        NA        NA        NA
#> n13 NA        NA        NA        NA         NA        NA        NA        NA
#> n14 NA        NA        NA        NA         NA        NA        NA        NA
#> n15 NA        NA        NA        NA         NA        NA        NA        NA
#> n16 NA        NA        NA        NA         NA        NA        NA        NA
#> n17 NA        NA        NA        NA         NA        NA        NA        NA
#> n18 NA        NA        NA        NA         NA        NA        NA        NA
#> n19 NA        NA        NA        NA         NA        NA        NA        NA
#> n20 NA        NA        NA        NA         NA        NA        NA        NA
#> n21 NA        NA        NA        NA         NA        NA        NA        NA
#> n22 NA        NA        NA        NA         NA        NA        NA        NA
#> n23 NA        NA        NA        NA         NA        NA        NA        NA
#> n24 NA        NA        NA        NA         NA        NA        NA        NA
#> n25 NA        NA        NA        NA         NA        NA        NA        NA
#> n26 NA        NA        NA        NA         NA        NA        NA        NA
#> n27 NA        NA        NA        NA         NA        NA        NA        NA
#> n28 NA        NA        NA        NA         NA        NA        NA        NA
#> n29 NA        NA        NA        NA         NA        NA        NA        NA
#> n30 NA        NA        NA        NA         NA        NA        NA        NA
#>            n9       n10        n11       n12       n13       n14       n15
#> n1         NA        NA         NA        NA 0.9291241        NA        NA
#> n2         NA        NA         NA        NA        NA        NA        NA
#> n3         NA        NA         NA        NA        NA        NA        NA
#> n4         NA        NA         NA        NA        NA 0.1738578        NA
#> n5  0.1924991 0.2398276 0.33061439        NA        NA        NA        NA
#> n6         NA        NA         NA        NA        NA        NA        NA
#> n7         NA 0.2221148 0.02104796        NA        NA        NA 0.9894145
#> n8         NA        NA         NA        NA 0.6503326        NA        NA
#> n9         NA        NA         NA 0.5596014        NA        NA 0.5158505
#> n10        NA        NA         NA        NA        NA 0.8447311        NA
#> n11        NA        NA         NA        NA        NA        NA        NA
#> n12        NA        NA         NA        NA        NA        NA        NA
#> n13        NA        NA         NA        NA        NA        NA        NA
#> n14        NA        NA         NA        NA        NA        NA        NA
#> n15        NA        NA         NA        NA        NA        NA        NA
#> n16        NA        NA         NA        NA        NA        NA        NA
#> n17        NA        NA         NA        NA        NA        NA        NA
#> n18        NA        NA         NA        NA        NA        NA        NA
#> n19        NA        NA         NA        NA        NA        NA        NA
#> n20        NA        NA         NA        NA        NA        NA        NA
#> n21        NA        NA         NA        NA        NA        NA        NA
#> n22        NA        NA         NA        NA        NA        NA        NA
#> n23        NA        NA         NA        NA        NA        NA        NA
#> n24        NA        NA         NA        NA        NA        NA        NA
#> n25        NA        NA         NA        NA        NA        NA        NA
#> n26        NA        NA         NA        NA        NA        NA        NA
#> n27        NA        NA         NA        NA        NA        NA        NA
#> n28        NA        NA         NA        NA        NA        NA        NA
#> n29        NA        NA         NA        NA        NA        NA        NA
#> n30        NA        NA         NA        NA        NA        NA        NA
#>           n16       n17       n18       n19       n20        n21       n22
#> n1         NA        NA 0.9791803        NA        NA         NA        NA
#> n2         NA        NA        NA        NA        NA         NA        NA
#> n3         NA        NA        NA        NA        NA         NA        NA
#> n4         NA 0.7332956        NA        NA        NA         NA        NA
#> n5         NA 0.4418591        NA 0.0030562        NA         NA        NA
#> n6         NA        NA        NA        NA 0.6101921         NA        NA
#> n7         NA        NA        NA        NA        NA 0.08948564 0.7665324
#> n8         NA        NA        NA        NA 0.7045321         NA 0.4779607
#> n9         NA        NA        NA        NA        NA         NA        NA
#> n10        NA        NA        NA        NA        NA         NA        NA
#> n11        NA        NA        NA        NA        NA         NA        NA
#> n12        NA        NA        NA        NA        NA 0.64244886        NA
#> n13        NA        NA        NA        NA        NA         NA        NA
#> n14        NA        NA        NA        NA        NA         NA        NA
#> n15 0.1530712        NA 0.3388545        NA        NA         NA        NA
#> n16        NA        NA        NA        NA        NA         NA        NA
#> n17        NA        NA        NA        NA        NA         NA        NA
#> n18        NA        NA        NA 0.7342051        NA         NA        NA
#> n19        NA        NA        NA        NA        NA         NA        NA
#> n20        NA        NA        NA        NA        NA         NA        NA
#> n21        NA        NA        NA        NA        NA         NA        NA
#> n22        NA        NA        NA        NA        NA         NA        NA
#> n23        NA        NA        NA        NA        NA         NA        NA
#> n24        NA        NA        NA        NA        NA         NA        NA
#> n25        NA        NA        NA        NA        NA         NA        NA
#> n26        NA        NA        NA        NA        NA         NA        NA
#> n27        NA        NA        NA        NA        NA         NA        NA
#> n28        NA        NA        NA        NA        NA         NA        NA
#> n29        NA        NA        NA        NA        NA         NA        NA
#> n30        NA        NA        NA        NA        NA         NA        NA
#>           n23       n24       n25        n26        n27       n28       n29
#> n1         NA        NA 0.4411032         NA         NA        NA        NA
#> n2         NA 0.5009481        NA         NA         NA        NA        NA
#> n3         NA        NA        NA         NA         NA        NA        NA
#> n4         NA        NA        NA         NA         NA        NA        NA
#> n5         NA 0.6439123        NA         NA         NA        NA        NA
#> n6         NA        NA        NA         NA         NA        NA        NA
#> n7         NA        NA        NA         NA         NA        NA        NA
#> n8         NA        NA        NA         NA         NA        NA        NA
#> n9         NA        NA        NA         NA         NA        NA        NA
#> n10        NA        NA 0.8517768         NA         NA 0.8101649        NA
#> n11        NA        NA        NA 0.03798899         NA        NA        NA
#> n12        NA        NA        NA         NA         NA        NA        NA
#> n13 0.1661065        NA        NA         NA         NA        NA        NA
#> n14        NA        NA        NA         NA         NA        NA        NA
#> n15        NA        NA        NA         NA         NA        NA        NA
#> n16 0.1410350        NA        NA         NA         NA        NA 0.3387780
#> n17        NA        NA        NA         NA         NA        NA        NA
#> n18        NA        NA        NA         NA         NA        NA        NA
#> n19        NA        NA        NA         NA         NA        NA        NA
#> n20        NA        NA        NA         NA         NA        NA        NA
#> n21        NA        NA        NA         NA         NA        NA        NA
#> n22        NA        NA        NA 0.09496310         NA        NA        NA
#> n23        NA        NA        NA         NA         NA        NA 0.6351401
#> n24        NA        NA        NA         NA         NA        NA        NA
#> n25        NA        NA        NA         NA 0.37135559        NA        NA
#> n26        NA        NA        NA         NA 0.06055786 0.1573367        NA
#> n27        NA        NA        NA         NA         NA        NA        NA
#> n28        NA        NA        NA         NA         NA        NA        NA
#> n29        NA        NA        NA         NA         NA        NA        NA
#> n30        NA        NA        NA         NA         NA        NA        NA
#>            n30
#> n1          NA
#> n2          NA
#> n3          NA
#> n4          NA
#> n5          NA
#> n6          NA
#> n7          NA
#> n8          NA
#> n9          NA
#> n10         NA
#> n11         NA
#> n12         NA
#> n13         NA
#> n14         NA
#> n15         NA
#> n16         NA
#> n17         NA
#> n18 0.36864310
#> n19         NA
#> n20         NA
#> n21         NA
#> n22 0.07489757
#> n23         NA
#> n24         NA
#> n25         NA
#> n26         NA
#> n27         NA
#> n28         NA
#> n29         NA
#> n30         NA
#> 
#> $pathway_pvalue
#> [1] 0.4750463
#> 
#> attr(,"class")
#> [1] "dce"