Differential Causal Effects - main function

Main function to compute differential causal effects and the pathway enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for igraph
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for graphNEL
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for matrix
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph
valid object defining a directed acyclic graph

df_expr_wt
data frame with wild type expression values

df_expr_mt
data from with mutation type expression values

solver
character with name of solver function

solver_args
additional arguments for the solver function

adjustment_type
character string for the method to define the adjustment set Z for the regression

effect_type
method of computing causal effects

p_method
character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test
either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size
either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding
indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative
logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level
Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,X.mt)

#> Warning: Cycle(s) detected in network

#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  0  0  1  0  1  0  0  1   0   1   0   0   0   0   0   1   0   0   0   0
#> n2   0  0  0  0  0  0  0  1  0   0   1   0   0   0   0   0   0   1   1   0   0
#> n3   0  0  0  1  1  0  0  0  0   0   0   0   0   0   0   1   0   0   0   1   1
#> n4   0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   1   0   0   0   0   1
#> n5   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0   0   0   0
#> n6   0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   0   0   1   0   0   0
#> n7   0  0  0  0  0  0  0  0  0   1   1   1   0   0   0   0   0   1   1   0   1
#> n8   0  0  0  0  0  0  0  0  0   0   0   0   1   0   0   1   1   0   0   0   0
#> n9   0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   0   0   0   0   0   0
#> n10  0  0  0  0  0  0  0  0  0   0   1   0   1   0   0   0   1   0   0   1   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   0   0   0   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   0   0   0   0   0   0
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0   1   1   0
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   1   0   0   1
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   1
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    1   0   1   0   0   0   0   0   0
#> n2    1   0   0   0   0   0   0   0   0
#> n3    0   0   1   0   0   0   0   0   1
#> n4    1   0   0   0   0   1   0   0   0
#> n5    0   0   0   0   0   1   0   0   0
#> n6    0   0   0   0   0   0   0   1   0
#> n7    0   0   0   0   0   0   0   0   1
#> n8    1   1   0   1   0   0   0   0   0
#> n9    1   0   0   0   0   0   0   0   0
#> n10   0   0   1   0   0   0   0   0   1
#> n11   1   0   0   1   1   1   1   0   0
#> n12   0   0   0   0   0   0   0   0   0
#> n13   0   0   0   0   0   0   0   0   0
#> n14   0   0   1   0   0   0   0   0   0
#> n15   0   1   0   0   0   0   1   1   0
#> n16   0   0   1   1   0   0   0   0   0
#> n17   0   0   0   0   0   0   1   0   0
#> n18   1   1   1   1   1   1   0   1   0
#> n19   0   0   1   0   0   0   1   0   0
#> n20   0   0   0   0   0   0   1   0   1
#> n21   0   0   0   0   0   1   0   1   0
#> n22   0   0   0   0   0   0   0   0   0
#> n23   0   0   0   0   0   0   0   0   0
#> n24   0   0   0   0   1   1   0   0   0
#> n25   0   0   0   0   0   0   1   0   0
#> n26   0   0   0   0   0   0   1   0   0
#> n27   0   0   0   0   0   0   1   0   0
#> n28   0   0   0   0   0   0   0   0   0
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1 n2 n3         n4         n5          n6 n7         n8           n9
#> n1  NA NA NA -0.1329204         NA -0.01992342 NA         NA -0.006933515
#> n2  NA NA NA         NA         NA          NA NA 0.00157553           NA
#> n3  NA NA NA -0.1866838 0.03035703          NA NA         NA           NA
#> n4  NA NA NA         NA         NA          NA NA         NA           NA
#> n5  NA NA NA         NA         NA          NA NA         NA           NA
#> n6  NA NA NA         NA         NA          NA NA 0.82547824           NA
#> n7  NA NA NA         NA         NA          NA NA         NA           NA
#> n8  NA NA NA         NA         NA          NA NA         NA           NA
#> n9  NA NA NA         NA         NA          NA NA         NA           NA
#> n10 NA NA NA         NA         NA          NA NA         NA           NA
#> n11 NA NA NA         NA         NA          NA NA         NA           NA
#> n12 NA NA NA         NA         NA          NA NA         NA           NA
#> n13 NA NA NA         NA         NA          NA NA         NA           NA
#> n14 NA NA NA         NA         NA          NA NA         NA           NA
#> n15 NA NA NA         NA         NA          NA NA         NA           NA
#> n16 NA NA NA         NA         NA          NA NA         NA           NA
#> n17 NA NA NA         NA         NA          NA NA         NA           NA
#> n18 NA NA NA         NA         NA          NA NA         NA           NA
#> n19 NA NA NA         NA         NA          NA NA         NA           NA
#> n20 NA NA NA         NA         NA          NA NA         NA           NA
#> n21 NA NA NA         NA         NA          NA NA         NA           NA
#> n22 NA NA NA         NA         NA          NA NA         NA           NA
#> n23 NA NA NA         NA         NA          NA NA         NA           NA
#> n24 NA NA NA         NA         NA          NA NA         NA           NA
#> n25 NA NA NA         NA         NA          NA NA         NA           NA
#> n26 NA NA NA         NA         NA          NA NA         NA           NA
#> n27 NA NA NA         NA         NA          NA NA         NA           NA
#> n28 NA NA NA         NA         NA          NA NA         NA           NA
#> n29 NA NA NA         NA         NA          NA NA         NA           NA
#> n30 NA NA NA         NA         NA          NA NA         NA           NA
#>             n10          n11        n12        n13        n14      n15
#> n1           NA -0.007242098         NA         NA         NA       NA
#> n2           NA  0.005771661         NA         NA         NA       NA
#> n3           NA           NA         NA         NA         NA       NA
#> n4           NA           NA         NA         NA         NA -0.13178
#> n5           NA           NA         NA         NA         NA       NA
#> n6           NA           NA         NA         NA         NA       NA
#> n7  -0.01480253 -0.005864668 -0.0253132         NA         NA       NA
#> n8           NA           NA         NA  0.7202383         NA       NA
#> n9           NA  0.376272649         NA         NA         NA       NA
#> n10          NA  0.208847092         NA -0.1699871         NA       NA
#> n11          NA           NA         NA         NA         NA       NA
#> n12          NA           NA         NA         NA         NA       NA
#> n13          NA           NA         NA         NA 0.08565511       NA
#> n14          NA           NA         NA         NA         NA       NA
#> n15          NA           NA         NA         NA         NA       NA
#> n16          NA           NA         NA         NA         NA       NA
#> n17          NA           NA         NA         NA         NA       NA
#> n18          NA           NA         NA         NA         NA       NA
#> n19          NA           NA         NA         NA         NA       NA
#> n20          NA           NA         NA         NA         NA       NA
#> n21          NA           NA         NA         NA         NA       NA
#> n22          NA           NA         NA         NA         NA       NA
#> n23          NA           NA         NA         NA         NA       NA
#> n24          NA           NA         NA         NA         NA       NA
#> n25          NA           NA         NA         NA         NA       NA
#> n26          NA           NA         NA         NA         NA       NA
#> n27          NA           NA         NA         NA         NA       NA
#> n28          NA           NA         NA         NA         NA       NA
#> n29          NA           NA         NA         NA         NA       NA
#> n30          NA           NA         NA         NA         NA       NA
#>             n16         n17         n18         n19        n20         n21
#> n1           NA -0.08306682          NA          NA         NA          NA
#> n2           NA          NA -0.06840168 -0.09546759         NA          NA
#> n3   0.15837313          NA          NA          NA 0.10271777  0.15039811
#> n4  -0.20962625          NA          NA          NA         NA -0.15543237
#> n5           NA -0.11627413          NA          NA         NA          NA
#> n6           NA          NA -1.74225095          NA         NA          NA
#> n7           NA          NA  0.06174148  0.03147051         NA -0.08856448
#> n8  -0.11917396 -0.23740482          NA          NA         NA          NA
#> n9           NA          NA          NA          NA         NA          NA
#> n10          NA  1.72556094          NA          NA 0.69638683          NA
#> n11  0.02798999          NA          NA          NA         NA          NA
#> n12          NA          NA          NA          NA         NA -0.23596896
#> n13          NA          NA          NA          NA         NA          NA
#> n14          NA          NA          NA          NA 0.02082481          NA
#> n15          NA          NA          NA          NA 0.44024948          NA
#> n16          NA -0.40564214          NA -0.43627484 0.37820204          NA
#> n17          NA          NA  0.12824457          NA         NA -0.69911848
#> n18          NA          NA          NA          NA 0.59769666 -0.04419662
#> n19          NA          NA          NA          NA         NA          NA
#> n20          NA          NA          NA          NA         NA          NA
#> n21          NA          NA          NA          NA         NA          NA
#> n22          NA          NA          NA          NA         NA          NA
#> n23          NA          NA          NA          NA         NA          NA
#> n24          NA          NA          NA          NA         NA          NA
#> n25          NA          NA          NA          NA         NA          NA
#> n26          NA          NA          NA          NA         NA          NA
#> n27          NA          NA          NA          NA         NA          NA
#> n28          NA          NA          NA          NA         NA          NA
#> n29          NA          NA          NA          NA         NA          NA
#> n30          NA          NA          NA          NA         NA          NA
#>             n22        n23         n24         n25         n26        n27
#> n1  -0.15156252         NA  0.05077664          NA          NA         NA
#> n2  -0.07871179         NA          NA          NA          NA         NA
#> n3           NA         NA -0.20134298          NA          NA         NA
#> n4   1.58526631         NA          NA          NA          NA  0.3077689
#> n5           NA         NA          NA          NA          NA  1.8547968
#> n6           NA         NA          NA          NA          NA         NA
#> n7           NA         NA          NA          NA          NA         NA
#> n8   0.10571399 1.28533455          NA  1.10433103          NA         NA
#> n9   1.65413789         NA          NA          NA          NA         NA
#> n10          NA         NA -0.84563046          NA          NA         NA
#> n11 -0.14125534         NA          NA -0.26932197  0.14963285 -0.1375581
#> n12          NA         NA          NA          NA          NA         NA
#> n13          NA         NA          NA          NA          NA         NA
#> n14          NA         NA  0.41565887          NA          NA         NA
#> n15          NA 0.01549135          NA          NA          NA         NA
#> n16          NA         NA  0.47605463  0.02192962          NA         NA
#> n17          NA         NA          NA          NA          NA         NA
#> n18  0.40896645 0.11425549 -0.10919782  0.03365184 -0.06337136  0.2802248
#> n19          NA         NA -1.12804308          NA          NA         NA
#> n20          NA         NA          NA          NA          NA         NA
#> n21          NA         NA          NA          NA          NA  0.3540549
#> n22          NA         NA          NA          NA          NA         NA
#> n23          NA         NA          NA          NA          NA         NA
#> n24          NA         NA          NA          NA -0.14073215 -0.1021437
#> n25          NA         NA          NA          NA          NA         NA
#> n26          NA         NA          NA          NA          NA         NA
#> n27          NA         NA          NA          NA          NA         NA
#> n28          NA         NA          NA          NA          NA         NA
#> n29          NA         NA          NA          NA          NA         NA
#> n30          NA         NA          NA          NA          NA         NA
#>               n28         n29          n30
#> n1             NA          NA           NA
#> n2             NA          NA           NA
#> n3             NA          NA  0.226347249
#> n4             NA          NA           NA
#> n5             NA          NA           NA
#> n6             NA -0.90072888           NA
#> n7             NA          NA  0.073202618
#> n8             NA          NA           NA
#> n9             NA          NA           NA
#> n10            NA          NA  0.077357482
#> n11 -0.0002938994          NA           NA
#> n12            NA          NA           NA
#> n13            NA          NA           NA
#> n14            NA          NA           NA
#> n15  0.3207889790 -0.11333189           NA
#> n16            NA          NA           NA
#> n17 -0.1699217884          NA           NA
#> n18            NA  0.04196184           NA
#> n19  1.5659284606          NA           NA
#> n20  0.0271595830          NA -0.008075212
#> n21            NA  0.14349754           NA
#> n22            NA          NA           NA
#> n23            NA          NA           NA
#> n24            NA          NA           NA
#> n25 -0.0243816135          NA           NA
#> n26 -0.5125658187          NA           NA
#> n27  0.0987926822          NA           NA
#> n28            NA          NA           NA
#> n29            NA          NA           NA
#> n30            NA          NA           NA
#> 
#> $dce_stderr
#>     n1 n2 n3         n4         n5         n6 n7         n8         n9
#> n1  NA NA NA 0.06526676         NA 0.01481962 NA         NA 0.01547317
#> n2  NA NA NA         NA         NA         NA NA 0.02543477         NA
#> n3  NA NA NA 0.11492626 0.01891123         NA NA         NA         NA
#> n4  NA NA NA         NA         NA         NA NA         NA         NA
#> n5  NA NA NA         NA         NA         NA NA         NA         NA
#> n6  NA NA NA         NA         NA         NA NA 0.81387215         NA
#> n7  NA NA NA         NA         NA         NA NA         NA         NA
#> n8  NA NA NA         NA         NA         NA NA         NA         NA
#> n9  NA NA NA         NA         NA         NA NA         NA         NA
#> n10 NA NA NA         NA         NA         NA NA         NA         NA
#> n11 NA NA NA         NA         NA         NA NA         NA         NA
#> n12 NA NA NA         NA         NA         NA NA         NA         NA
#> n13 NA NA NA         NA         NA         NA NA         NA         NA
#> n14 NA NA NA         NA         NA         NA NA         NA         NA
#> n15 NA NA NA         NA         NA         NA NA         NA         NA
#> n16 NA NA NA         NA         NA         NA NA         NA         NA
#> n17 NA NA NA         NA         NA         NA NA         NA         NA
#> n18 NA NA NA         NA         NA         NA NA         NA         NA
#> n19 NA NA NA         NA         NA         NA NA         NA         NA
#> n20 NA NA NA         NA         NA         NA NA         NA         NA
#> n21 NA NA NA         NA         NA         NA NA         NA         NA
#> n22 NA NA NA         NA         NA         NA NA         NA         NA
#> n23 NA NA NA         NA         NA         NA NA         NA         NA
#> n24 NA NA NA         NA         NA         NA NA         NA         NA
#> n25 NA NA NA         NA         NA         NA NA         NA         NA
#> n26 NA NA NA         NA         NA         NA NA         NA         NA
#> n27 NA NA NA         NA         NA         NA NA         NA         NA
#> n28 NA NA NA         NA         NA         NA NA         NA         NA
#> n29 NA NA NA         NA         NA         NA NA         NA         NA
#> n30 NA NA NA         NA         NA         NA NA         NA         NA
#>            n10        n11        n12       n13       n14       n15       n16
#> n1          NA 0.04526535         NA        NA        NA        NA        NA
#> n2          NA 0.04495937         NA        NA        NA        NA        NA
#> n3          NA         NA         NA        NA        NA        NA 0.1090576
#> n4          NA         NA         NA        NA        NA 0.1588604 0.2135160
#> n5          NA         NA         NA        NA        NA        NA        NA
#> n6          NA         NA         NA        NA        NA        NA        NA
#> n7  0.01285154 0.06187148 0.02116049        NA        NA        NA        NA
#> n8          NA         NA         NA 0.4113141        NA        NA 0.3584318
#> n9          NA 0.42649459         NA        NA        NA        NA        NA
#> n10         NA 0.73954940         NA 0.9441605        NA        NA        NA
#> n11         NA         NA         NA        NA        NA        NA 0.4157054
#> n12         NA         NA         NA        NA        NA        NA        NA
#> n13         NA         NA         NA        NA 0.1183327        NA        NA
#> n14         NA         NA         NA        NA        NA        NA        NA
#> n15         NA         NA         NA        NA        NA        NA        NA
#> n16         NA         NA         NA        NA        NA        NA        NA
#> n17         NA         NA         NA        NA        NA        NA        NA
#> n18         NA         NA         NA        NA        NA        NA        NA
#> n19         NA         NA         NA        NA        NA        NA        NA
#> n20         NA         NA         NA        NA        NA        NA        NA
#> n21         NA         NA         NA        NA        NA        NA        NA
#> n22         NA         NA         NA        NA        NA        NA        NA
#> n23         NA         NA         NA        NA        NA        NA        NA
#> n24         NA         NA         NA        NA        NA        NA        NA
#> n25         NA         NA         NA        NA        NA        NA        NA
#> n26         NA         NA         NA        NA        NA        NA        NA
#> n27         NA         NA         NA        NA        NA        NA        NA
#> n28         NA         NA         NA        NA        NA        NA        NA
#> n29         NA         NA         NA        NA        NA        NA        NA
#> n30         NA         NA         NA        NA        NA        NA        NA
#>            n17        n18        n19       n20       n21       n22       n23
#> n1  0.07844767         NA         NA        NA        NA 0.1646239        NA
#> n2          NA 0.13472258 0.11668886        NA        NA 0.0679749        NA
#> n3          NA         NA         NA 0.1938164 0.2190194        NA        NA
#> n4          NA         NA         NA        NA 0.7523873 1.0871495        NA
#> n5  1.51638539         NA         NA        NA        NA        NA        NA
#> n6          NA 1.20483581         NA        NA        NA        NA        NA
#> n7          NA 0.05939398 0.06704606        NA 0.1844073        NA        NA
#> n8  0.46816579         NA         NA        NA        NA 0.4531854 0.6091704
#> n9          NA         NA         NA        NA        NA 1.5864270        NA
#> n10 2.20539971         NA         NA 2.0695810        NA        NA        NA
#> n11         NA         NA         NA        NA        NA 0.5949044        NA
#> n12         NA         NA         NA        NA 1.3596614        NA        NA
#> n13         NA         NA         NA        NA        NA        NA        NA
#> n14         NA         NA         NA 2.4332141        NA        NA        NA
#> n15         NA         NA         NA 1.3691508        NA        NA 0.6997514
#> n16 0.32495923         NA 0.73225690 0.7937795        NA        NA        NA
#> n17         NA 0.24644546         NA        NA 0.2517343        NA        NA
#> n18         NA         NA         NA 0.4580330 0.4547438 0.3769232 0.2186669
#> n19         NA         NA         NA        NA        NA        NA        NA
#> n20         NA         NA         NA        NA        NA        NA        NA
#> n21         NA         NA         NA        NA        NA        NA        NA
#> n22         NA         NA         NA        NA        NA        NA        NA
#> n23         NA         NA         NA        NA        NA        NA        NA
#> n24         NA         NA         NA        NA        NA        NA        NA
#> n25         NA         NA         NA        NA        NA        NA        NA
#> n26         NA         NA         NA        NA        NA        NA        NA
#> n27         NA         NA         NA        NA        NA        NA        NA
#> n28         NA         NA         NA        NA        NA        NA        NA
#> n29         NA         NA         NA        NA        NA        NA        NA
#> n30         NA         NA         NA        NA        NA        NA        NA
#>            n24       n25       n26       n27       n28       n29        n30
#> n1  0.08499437        NA        NA        NA        NA        NA         NA
#> n2          NA        NA        NA        NA        NA        NA         NA
#> n3  0.11437144        NA        NA        NA        NA        NA 0.14484866
#> n4          NA        NA        NA 0.7584542        NA        NA         NA
#> n5          NA        NA        NA 0.9740267        NA        NA         NA
#> n6          NA        NA        NA        NA        NA 0.5071446         NA
#> n7          NA        NA        NA        NA        NA        NA 0.06584265
#> n8          NA 0.4471581        NA        NA        NA        NA         NA
#> n9          NA        NA        NA        NA        NA        NA         NA
#> n10 1.37513187        NA        NA        NA        NA        NA 0.79442267
#> n11         NA 0.2882776 0.8560998 1.0796745 0.8928069        NA         NA
#> n12         NA        NA        NA        NA        NA        NA         NA
#> n13         NA        NA        NA        NA        NA        NA         NA
#> n14 1.55474132        NA        NA        NA        NA        NA         NA
#> n15         NA        NA        NA        NA 1.9114767 0.4763018         NA
#> n16 0.36441020 0.3568487        NA        NA        NA        NA         NA
#> n17         NA        NA        NA        NA 0.4001872        NA         NA
#> n18 0.35106339 0.1412422 0.5492167 0.4253056        NA 0.2446747         NA
#> n19 0.66998644        NA        NA        NA 1.1265886        NA         NA
#> n20         NA        NA        NA        NA 0.3352507        NA 0.17647320
#> n21         NA        NA        NA 0.2574048        NA 0.1314076         NA
#> n22         NA        NA        NA        NA        NA        NA         NA
#> n23         NA        NA        NA        NA        NA        NA         NA
#> n24         NA        NA 0.1853888 0.2796913        NA        NA         NA
#> n25         NA        NA        NA        NA 0.7680770        NA         NA
#> n26         NA        NA        NA        NA 0.5501005        NA         NA
#> n27         NA        NA        NA        NA 0.3023137        NA         NA
#> n28         NA        NA        NA        NA        NA        NA         NA
#> n29         NA        NA        NA        NA        NA        NA         NA
#> n30         NA        NA        NA        NA        NA        NA         NA
#> 
#> $dce_pvalue
#>     n1 n2 n3         n4        n5        n6 n7        n8        n9       n10
#> n1  NA NA NA 0.04303809        NA 0.1803746 NA        NA 0.6545764        NA
#> n2  NA NA NA         NA        NA        NA NA 0.9506706        NA        NA
#> n3  NA NA NA 0.10590198 0.1100513        NA NA        NA        NA        NA
#> n4  NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n5  NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n6  NA NA NA         NA        NA        NA NA 0.3117218        NA        NA
#> n7  NA NA NA         NA        NA        NA NA        NA        NA 0.2508023
#> n8  NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n9  NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n10 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n11 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n12 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n13 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n14 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n15 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n16 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n17 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n18 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n19 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n20 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n21 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n22 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n23 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n24 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n25 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n26 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n27 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n28 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n29 NA NA NA         NA        NA        NA NA        NA        NA        NA
#> n30 NA NA NA         NA        NA        NA NA        NA        NA        NA
#>           n11       n12        n13       n14       n15       n16       n17
#> n1  0.8730521        NA         NA        NA        NA        NA 0.2909568
#> n2  0.8979838        NA         NA        NA        NA        NA        NA
#> n3         NA        NA         NA        NA        NA 0.1480457        NA
#> n4         NA        NA         NA        NA 0.4078328 0.3274425        NA
#> n5         NA        NA         NA        NA        NA        NA 0.9389583
#> n6         NA        NA         NA        NA        NA        NA        NA
#> n7  0.9245801 0.2330448         NA        NA        NA        NA        NA
#> n8         NA        NA 0.08153141        NA        NA 0.7398842 0.6126696
#> n9  0.3787363        NA         NA        NA        NA        NA        NA
#> n10 0.7779396        NA 0.85730873        NA        NA        NA 0.4349185
#> n11        NA        NA         NA        NA        NA 0.9463903        NA
#> n12        NA        NA         NA        NA        NA        NA        NA
#> n13        NA        NA         NA 0.4700388        NA        NA        NA
#> n14        NA        NA         NA        NA        NA        NA        NA
#> n15        NA        NA         NA        NA        NA        NA        NA
#> n16        NA        NA         NA        NA        NA        NA 0.2134786
#> n17        NA        NA         NA        NA        NA        NA        NA
#> n18        NA        NA         NA        NA        NA        NA        NA
#> n19        NA        NA         NA        NA        NA        NA        NA
#> n20        NA        NA         NA        NA        NA        NA        NA
#> n21        NA        NA         NA        NA        NA        NA        NA
#> n22        NA        NA         NA        NA        NA        NA        NA
#> n23        NA        NA         NA        NA        NA        NA        NA
#> n24        NA        NA         NA        NA        NA        NA        NA
#> n25        NA        NA         NA        NA        NA        NA        NA
#> n26        NA        NA         NA        NA        NA        NA        NA
#> n27        NA        NA         NA        NA        NA        NA        NA
#> n28        NA        NA         NA        NA        NA        NA        NA
#> n29        NA        NA         NA        NA        NA        NA        NA
#> n30        NA        NA         NA        NA        NA        NA        NA
#>           n18       n19       n20         n21       n22        n23        n24
#> n1         NA        NA        NA          NA 0.3583601         NA 0.55092197
#> n2  0.6122188 0.4142719        NA          NA 0.2482920         NA         NA
#> n3         NA        NA 0.5967294 0.493090562        NA         NA 0.07989452
#> n4         NA        NA        NA 0.836552440 0.1464231         NA         NA
#> n5         NA        NA        NA          NA        NA         NA         NA
#> n6  0.1497772        NA        NA          NA        NA         NA         NA
#> n7  0.2998416 0.6393151        NA 0.631574122        NA         NA         NA
#> n8         NA        NA        NA          NA 0.8158012 0.03615581         NA
#> n9         NA        NA        NA          NA 0.2983935         NA         NA
#> n10        NA        NA 0.7368674          NA        NA         NA 0.53931167
#> n11        NA        NA        NA          NA 0.8125752         NA         NA
#> n12        NA        NA        NA 0.862400242        NA         NA         NA
#> n13        NA        NA        NA          NA        NA         NA         NA
#> n14        NA        NA 0.9931801          NA        NA         NA 0.78948430
#> n15        NA        NA 0.7481403          NA        NA 0.98236034         NA
#> n16        NA 0.5520289 0.6343019          NA        NA         NA 0.19302266
#> n17 0.6034196        NA        NA 0.006044328        NA         NA         NA
#> n18        NA        NA 0.1935149 0.922678921 0.2793055 0.60193101 0.75610819
#> n19        NA        NA        NA          NA        NA         NA 0.09388565
#> n20        NA        NA        NA          NA        NA         NA         NA
#> n21        NA        NA        NA          NA        NA         NA         NA
#> n22        NA        NA        NA          NA        NA         NA         NA
#> n23        NA        NA        NA          NA        NA         NA         NA
#> n24        NA        NA        NA          NA        NA         NA         NA
#> n25        NA        NA        NA          NA        NA         NA         NA
#> n26        NA        NA        NA          NA        NA         NA         NA
#> n27        NA        NA        NA          NA        NA         NA         NA
#> n28        NA        NA        NA          NA        NA         NA         NA
#> n29        NA        NA        NA          NA        NA         NA         NA
#> n30        NA        NA        NA          NA        NA         NA         NA
#>           n25       n26        n27       n28        n29       n30
#> n1         NA        NA         NA        NA         NA        NA
#> n2         NA        NA         NA        NA         NA        NA
#> n3         NA        NA         NA        NA         NA 0.1197489
#> n4         NA        NA 0.68535288        NA         NA        NA
#> n5         NA        NA 0.05835655        NA         NA        NA
#> n6         NA        NA         NA        NA 0.07728737        NA
#> n7         NA        NA         NA        NA         NA 0.2675941
#> n8  0.0143970        NA         NA        NA         NA        NA
#> n9         NA        NA         NA        NA         NA        NA
#> n10        NA        NA         NA        NA         NA 0.9225286
#> n11 0.3513887 0.8614392 0.89875588 0.9997377         NA        NA
#> n12        NA        NA         NA        NA         NA        NA
#> n13        NA        NA         NA        NA         NA        NA
#> n14        NA        NA         NA        NA         NA        NA
#> n15        NA        NA         NA 0.8668977 0.81217769        NA
#> n16 0.9510633        NA         NA        NA         NA        NA
#> n17        NA        NA         NA 0.6716148         NA        NA
#> n18 0.8119418 0.9082631 0.51077976        NA 0.86401466        NA
#> n19        NA        NA         NA 0.1661632         NA        NA
#> n20        NA        NA         NA 0.9355199         NA 0.9635521
#> n21        NA        NA 0.17065520        NA 0.27625939        NA
#> n22        NA        NA         NA        NA         NA        NA
#> n23        NA        NA         NA        NA         NA        NA
#> n24        NA 0.4487634 0.71538505        NA         NA        NA
#> n25        NA        NA         NA 0.9747101         NA        NA
#> n26        NA        NA         NA 0.3526391         NA        NA
#> n27        NA        NA         NA 0.7441992         NA        NA
#> n28        NA        NA         NA        NA         NA        NA
#> n29        NA        NA         NA        NA         NA        NA
#> n30        NA        NA         NA        NA         NA        NA
#> 
#> $pathway_pvalue
#> [1] 0.5333836
#> 
#> attr(,"class")
#> [1] "dce"