Differential Causal Effects - main function

Main function to compute differential causal effects and the pathway enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  verbose = FALSE
)

# S4 method for igraph
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  verbose = FALSE
)

# S4 method for graphNEL
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  verbose = FALSE
)

# S4 method for matrix
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  verbose = FALSE
)

Arguments

graph
valid object defining a directed acyclic graph

df_expr_wt
data frame with wild type expression values

df_expr_mt
data from with mutation type expression values

solver
character with name of solver function

solver_args
additional arguments for the solver function

adjustment_type
character string for the method to define the adjustment set Z for the regression

effect_type
method of computing causal effects

p_method
character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test
either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size
either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding
indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative
logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

verbose
logical for verbose output

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,X.mt)

#> Warning: Cycle(s) detected in network

#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   1   1   1   0   0   0
#> n2   0  0  1  0  0  0  0  0  1   0   1   0   0   0   0   1   0   0   0   0   0
#> n3   0  0  0  0  0  0  1  0  1   1   0   0   0   0   0   0   0   1   0   0   0
#> n4   0  0  0  0  1  1  1  0  0   0   0   0   0   0   1   0   0   1   1   0   0
#> n5   0  0  0  0  0  1  0  0  0   0   1   0   0   0   1   0   0   0   0   0   0
#> n6   0  0  0  0  0  0  0  1  0   0   1   1   0   0   0   0   0   0   0   0   0
#> n7   0  0  0  0  0  0  0  0  0   0   0   0   1   1   1   0   0   1   0   0   0
#> n8   0  0  0  0  0  0  0  0  1   0   0   0   0   0   0   0   1   0   0   1   0
#> n9   0  0  0  0  0  0  0  0  0   0   0   0   0   1   1   0   0   0   1   0   0
#> n10  0  0  0  0  0  0  0  0  0   0   0   1   0   0   0   0   0   0   0   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   1   1   0   0   0   1   0   0   0   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0   0   0   1
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   1   0   0   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    0   0   0   0   0   0   0   0   0
#> n2    1   0   0   1   0   1   0   0   1
#> n3    0   0   0   0   0   0   0   0   0
#> n4    0   0   0   0   0   1   0   0   0
#> n5    0   0   0   0   0   0   0   0   0
#> n6    0   0   0   0   0   1   0   0   1
#> n7    0   1   0   0   0   0   1   0   0
#> n8    0   0   0   0   0   0   0   0   0
#> n9    0   1   0   0   0   0   0   1   0
#> n10   0   0   0   0   0   0   0   0   0
#> n11   0   1   0   0   0   0   0   1   0
#> n12   0   1   0   0   0   0   1   0   0
#> n13   0   0   0   1   0   1   0   0   0
#> n14   0   0   0   0   0   0   0   0   0
#> n15   0   0   0   1   0   0   0   1   0
#> n16   0   0   1   0   0   0   0   0   0
#> n17   0   0   0   0   0   1   0   0   0
#> n18   0   0   0   1   0   0   0   0   0
#> n19   0   1   1   0   0   0   0   0   1
#> n20   0   0   0   0   0   1   0   0   0
#> n21   0   0   0   0   0   0   0   0   0
#> n22   0   1   0   0   0   0   0   0   0
#> n23   0   0   1   0   0   0   0   0   0
#> n24   0   0   0   0   0   0   0   1   0
#> n25   0   0   0   0   0   0   0   1   0
#> n26   0   0   0   0   0   0   0   0   1
#> n27   0   0   0   0   0   0   0   0   0
#> n28   0   0   0   0   0   0   0   0   1
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1 n2         n3 n4           n5          n6         n7         n8
#> n1  NA NA         NA NA           NA          NA         NA         NA
#> n2  NA NA -0.0158564 NA           NA          NA         NA         NA
#> n3  NA NA         NA NA           NA          NA 0.74425469         NA
#> n4  NA NA         NA NA -0.003142513 -0.01226063 0.09042489         NA
#> n5  NA NA         NA NA           NA -0.07489950         NA         NA
#> n6  NA NA         NA NA           NA          NA         NA 0.05350847
#> n7  NA NA         NA NA           NA          NA         NA         NA
#> n8  NA NA         NA NA           NA          NA         NA         NA
#> n9  NA NA         NA NA           NA          NA         NA         NA
#> n10 NA NA         NA NA           NA          NA         NA         NA
#> n11 NA NA         NA NA           NA          NA         NA         NA
#> n12 NA NA         NA NA           NA          NA         NA         NA
#> n13 NA NA         NA NA           NA          NA         NA         NA
#> n14 NA NA         NA NA           NA          NA         NA         NA
#> n15 NA NA         NA NA           NA          NA         NA         NA
#> n16 NA NA         NA NA           NA          NA         NA         NA
#> n17 NA NA         NA NA           NA          NA         NA         NA
#> n18 NA NA         NA NA           NA          NA         NA         NA
#> n19 NA NA         NA NA           NA          NA         NA         NA
#> n20 NA NA         NA NA           NA          NA         NA         NA
#> n21 NA NA         NA NA           NA          NA         NA         NA
#> n22 NA NA         NA NA           NA          NA         NA         NA
#> n23 NA NA         NA NA           NA          NA         NA         NA
#> n24 NA NA         NA NA           NA          NA         NA         NA
#> n25 NA NA         NA NA           NA          NA         NA         NA
#> n26 NA NA         NA NA           NA          NA         NA         NA
#> n27 NA NA         NA NA           NA          NA         NA         NA
#> n28 NA NA         NA NA           NA          NA         NA         NA
#> n29 NA NA         NA NA           NA          NA         NA         NA
#> n30 NA NA         NA NA           NA          NA         NA         NA
#>              n9         n10         n11         n12         n13        n14
#> n1           NA          NA          NA          NA          NA 0.02506077
#> n2  -0.03662647          NA -0.06537389          NA          NA         NA
#> n3   0.18124232 -0.06916058          NA          NA          NA         NA
#> n4           NA          NA          NA          NA          NA         NA
#> n5           NA          NA -0.25614272          NA          NA         NA
#> n6           NA          NA -0.18264319 -0.20553018          NA         NA
#> n7           NA          NA          NA          NA -0.24397227 0.20112085
#> n8  -0.20396607          NA          NA          NA          NA         NA
#> n9           NA          NA          NA          NA          NA 0.64427649
#> n10          NA          NA          NA  0.30119876          NA         NA
#> n11          NA          NA          NA -0.06676994 -0.02565627         NA
#> n12          NA          NA          NA          NA          NA         NA
#> n13          NA          NA          NA          NA          NA         NA
#> n14          NA          NA          NA          NA          NA         NA
#> n15          NA          NA          NA          NA          NA         NA
#> n16          NA          NA          NA          NA          NA         NA
#> n17          NA          NA          NA          NA          NA         NA
#> n18          NA          NA          NA          NA          NA         NA
#> n19          NA          NA          NA          NA          NA         NA
#> n20          NA          NA          NA          NA          NA         NA
#> n21          NA          NA          NA          NA          NA         NA
#> n22          NA          NA          NA          NA          NA         NA
#> n23          NA          NA          NA          NA          NA         NA
#> n24          NA          NA          NA          NA          NA         NA
#> n25          NA          NA          NA          NA          NA         NA
#> n26          NA          NA          NA          NA          NA         NA
#> n27          NA          NA          NA          NA          NA         NA
#> n28          NA          NA          NA          NA          NA         NA
#> n29          NA          NA          NA          NA          NA         NA
#> n30          NA          NA          NA          NA          NA         NA
#>             n15         n16         n17         n18         n19        n20
#> n1           NA -0.03883634 -0.07587297 -0.21586431          NA         NA
#> n2           NA -0.06337460          NA          NA          NA         NA
#> n3           NA          NA          NA  1.64332088          NA         NA
#> n4   0.07002761          NA          NA  0.03457117  0.01377727         NA
#> n5  -0.43413099          NA          NA          NA          NA         NA
#> n6           NA          NA          NA          NA          NA         NA
#> n7  -0.42991014          NA          NA  0.11434543          NA         NA
#> n8           NA          NA  0.30421112          NA          NA  0.2274981
#> n9  -0.25302795          NA          NA          NA -0.02671571         NA
#> n10          NA          NA          NA          NA          NA         NA
#> n11          NA          NA -0.29437197          NA          NA         NA
#> n12          NA          NA          NA          NA          NA -0.1076579
#> n13 -0.36217060          NA          NA          NA          NA         NA
#> n14          NA          NA          NA -0.33811426          NA         NA
#> n15          NA          NA          NA          NA          NA         NA
#> n16          NA          NA          NA          NA          NA         NA
#> n17          NA          NA          NA          NA          NA  0.3695955
#> n18          NA          NA          NA          NA          NA         NA
#> n19          NA          NA          NA          NA          NA         NA
#> n20          NA          NA          NA          NA          NA         NA
#> n21          NA          NA          NA          NA          NA         NA
#> n22          NA          NA          NA          NA          NA         NA
#> n23          NA          NA          NA          NA          NA         NA
#> n24          NA          NA          NA          NA          NA         NA
#> n25          NA          NA          NA          NA          NA         NA
#> n26          NA          NA          NA          NA          NA         NA
#> n27          NA          NA          NA          NA          NA         NA
#> n28          NA          NA          NA          NA          NA         NA
#> n29          NA          NA          NA          NA          NA         NA
#> n30          NA          NA          NA          NA          NA         NA
#>            n21         n22         n23         n24         n25 n26        n27
#> n1          NA          NA          NA          NA          NA  NA         NA
#> n2          NA 0.008611816          NA          NA -0.23742552  NA -0.1303073
#> n3          NA          NA          NA          NA          NA  NA         NA
#> n4          NA          NA          NA          NA          NA  NA -0.1553150
#> n5          NA          NA          NA          NA          NA  NA         NA
#> n6          NA          NA          NA          NA          NA  NA  0.3000018
#> n7          NA          NA  0.16840516          NA          NA  NA         NA
#> n8          NA          NA          NA          NA          NA  NA         NA
#> n9          NA          NA -0.37513942          NA          NA  NA         NA
#> n10         NA          NA          NA          NA          NA  NA         NA
#> n11         NA          NA  0.02377339          NA          NA  NA         NA
#> n12         NA          NA  0.08387154          NA          NA  NA         NA
#> n13 -1.9201007          NA          NA          NA -0.35339868  NA -0.7947446
#> n14         NA          NA          NA          NA          NA  NA         NA
#> n15  0.1140282          NA          NA          NA -0.81569956  NA         NA
#> n16  1.5288836          NA          NA  0.36552440          NA  NA         NA
#> n17         NA          NA          NA          NA          NA  NA -0.5344875
#> n18         NA          NA          NA          NA -0.06860646  NA         NA
#> n19  3.1532717          NA -1.77881561 -0.05722452          NA  NA         NA
#> n20 -0.2423641          NA          NA          NA          NA  NA -1.3105215
#> n21         NA          NA          NA          NA          NA  NA         NA
#> n22         NA          NA  0.83860146          NA          NA  NA         NA
#> n23         NA          NA          NA -0.21603175          NA  NA         NA
#> n24         NA          NA          NA          NA          NA  NA         NA
#> n25         NA          NA          NA          NA          NA  NA         NA
#> n26         NA          NA          NA          NA          NA  NA         NA
#> n27         NA          NA          NA          NA          NA  NA         NA
#> n28         NA          NA          NA          NA          NA  NA         NA
#> n29         NA          NA          NA          NA          NA  NA         NA
#> n30         NA          NA          NA          NA          NA  NA         NA
#>            n28         n29        n30
#> n1          NA          NA         NA
#> n2          NA          NA -0.1424501
#> n3          NA          NA         NA
#> n4          NA          NA         NA
#> n5          NA          NA         NA
#> n6          NA          NA  0.6550903
#> n7  0.07210089          NA         NA
#> n8          NA          NA         NA
#> n9          NA -0.22025406         NA
#> n10         NA          NA         NA
#> n11         NA -0.05285491         NA
#> n12 0.02699409          NA         NA
#> n13         NA          NA         NA
#> n14         NA          NA         NA
#> n15         NA -0.20980642         NA
#> n16         NA          NA         NA
#> n17         NA          NA         NA
#> n18         NA          NA         NA
#> n19         NA          NA -0.3439059
#> n20         NA          NA         NA
#> n21         NA          NA         NA
#> n22         NA          NA         NA
#> n23         NA          NA         NA
#> n24         NA -0.21634598         NA
#> n25         NA  0.19171475         NA
#> n26         NA          NA -0.0146057
#> n27         NA          NA         NA
#> n28         NA          NA  0.4817357
#> n29         NA          NA         NA
#> n30         NA          NA         NA
#> 
#> $dce_stderr
#>     n1 n2         n3 n4         n5         n6         n7         n8         n9
#> n1  NA NA         NA NA         NA         NA         NA         NA         NA
#> n2  NA NA 0.02017282 NA         NA         NA         NA         NA 0.02864433
#> n3  NA NA         NA NA         NA         NA 0.39433200         NA 0.19374341
#> n4  NA NA         NA NA 0.01065462 0.03004638 0.05971195         NA         NA
#> n5  NA NA         NA NA         NA 0.26875479         NA         NA         NA
#> n6  NA NA         NA NA         NA         NA         NA 0.08006033         NA
#> n7  NA NA         NA NA         NA         NA         NA         NA         NA
#> n8  NA NA         NA NA         NA         NA         NA         NA 0.43736877
#> n9  NA NA         NA NA         NA         NA         NA         NA         NA
#> n10 NA NA         NA NA         NA         NA         NA         NA         NA
#> n11 NA NA         NA NA         NA         NA         NA         NA         NA
#> n12 NA NA         NA NA         NA         NA         NA         NA         NA
#> n13 NA NA         NA NA         NA         NA         NA         NA         NA
#> n14 NA NA         NA NA         NA         NA         NA         NA         NA
#> n15 NA NA         NA NA         NA         NA         NA         NA         NA
#> n16 NA NA         NA NA         NA         NA         NA         NA         NA
#> n17 NA NA         NA NA         NA         NA         NA         NA         NA
#> n18 NA NA         NA NA         NA         NA         NA         NA         NA
#> n19 NA NA         NA NA         NA         NA         NA         NA         NA
#> n20 NA NA         NA NA         NA         NA         NA         NA         NA
#> n21 NA NA         NA NA         NA         NA         NA         NA         NA
#> n22 NA NA         NA NA         NA         NA         NA         NA         NA
#> n23 NA NA         NA NA         NA         NA         NA         NA         NA
#> n24 NA NA         NA NA         NA         NA         NA         NA         NA
#> n25 NA NA         NA NA         NA         NA         NA         NA         NA
#> n26 NA NA         NA NA         NA         NA         NA         NA         NA
#> n27 NA NA         NA NA         NA         NA         NA         NA         NA
#> n28 NA NA         NA NA         NA         NA         NA         NA         NA
#> n29 NA NA         NA NA         NA         NA         NA         NA         NA
#> n30 NA NA         NA NA         NA         NA         NA         NA         NA
#>            n10        n11       n12       n13        n14        n15        n16
#> n1          NA         NA        NA        NA 0.06586027         NA 0.08536513
#> n2          NA 0.04334838        NA        NA         NA         NA 0.11754822
#> n3  0.08420809         NA        NA        NA         NA         NA         NA
#> n4          NA         NA        NA        NA         NA 0.03767462         NA
#> n5          NA 0.78583560        NA        NA         NA 0.50477886         NA
#> n6          NA 0.42343156 0.1287147        NA         NA         NA         NA
#> n7          NA         NA        NA 0.2021605 0.69537020 0.29280131         NA
#> n8          NA         NA        NA        NA         NA         NA         NA
#> n9          NA         NA        NA        NA 1.10319514 0.81987168         NA
#> n10         NA         NA 0.9110356        NA         NA         NA         NA
#> n11         NA         NA 0.1650745 0.1876833         NA         NA         NA
#> n12         NA         NA        NA        NA         NA         NA         NA
#> n13         NA         NA        NA        NA         NA 0.47123039         NA
#> n14         NA         NA        NA        NA         NA         NA         NA
#> n15         NA         NA        NA        NA         NA         NA         NA
#> n16         NA         NA        NA        NA         NA         NA         NA
#> n17         NA         NA        NA        NA         NA         NA         NA
#> n18         NA         NA        NA        NA         NA         NA         NA
#> n19         NA         NA        NA        NA         NA         NA         NA
#> n20         NA         NA        NA        NA         NA         NA         NA
#> n21         NA         NA        NA        NA         NA         NA         NA
#> n22         NA         NA        NA        NA         NA         NA         NA
#> n23         NA         NA        NA        NA         NA         NA         NA
#> n24         NA         NA        NA        NA         NA         NA         NA
#> n25         NA         NA        NA        NA         NA         NA         NA
#> n26         NA         NA        NA        NA         NA         NA         NA
#> n27         NA         NA        NA        NA         NA         NA         NA
#> n28         NA         NA        NA        NA         NA         NA         NA
#> n29         NA         NA        NA        NA         NA         NA         NA
#> n30         NA         NA        NA        NA         NA         NA         NA
#>            n17       n18         n19       n20       n21     n22       n23
#> n1  0.05756333 0.2039966          NA        NA        NA      NA        NA
#> n2          NA        NA          NA        NA        NA 0.01905        NA
#> n3          NA 1.4280957          NA        NA        NA      NA        NA
#> n4          NA 0.1124043 0.008770311        NA        NA      NA        NA
#> n5          NA        NA          NA        NA        NA      NA        NA
#> n6          NA        NA          NA        NA        NA      NA        NA
#> n7          NA 0.9498349          NA        NA        NA      NA 0.1976981
#> n8  1.17734417        NA          NA 1.0514686        NA      NA        NA
#> n9          NA        NA 0.167948562        NA        NA      NA 0.3582329
#> n10         NA        NA          NA        NA        NA      NA        NA
#> n11 0.67523752        NA          NA        NA        NA      NA 0.2584100
#> n12         NA        NA          NA 0.5465507        NA      NA 0.2213246
#> n13         NA        NA          NA        NA 1.2649244      NA        NA
#> n14         NA 0.5145981          NA        NA        NA      NA        NA
#> n15         NA        NA          NA        NA 1.0518330      NA        NA
#> n16         NA        NA          NA        NA 1.2371790      NA        NA
#> n17         NA        NA          NA 0.2676147        NA      NA        NA
#> n18         NA        NA          NA        NA        NA      NA        NA
#> n19         NA        NA          NA        NA 2.6921851      NA 1.0872331
#> n20         NA        NA          NA        NA 0.8713372      NA        NA
#> n21         NA        NA          NA        NA        NA      NA        NA
#> n22         NA        NA          NA        NA        NA      NA 0.3172236
#> n23         NA        NA          NA        NA        NA      NA        NA
#> n24         NA        NA          NA        NA        NA      NA        NA
#> n25         NA        NA          NA        NA        NA      NA        NA
#> n26         NA        NA          NA        NA        NA      NA        NA
#> n27         NA        NA          NA        NA        NA      NA        NA
#> n28         NA        NA          NA        NA        NA      NA        NA
#> n29         NA        NA          NA        NA        NA      NA        NA
#> n30         NA        NA          NA        NA        NA      NA        NA
#>           n24       n25 n26        n27       n28       n29       n30
#> n1         NA        NA  NA         NA        NA        NA        NA
#> n2         NA 0.1877239  NA 0.07872916        NA        NA 0.1091344
#> n3         NA        NA  NA         NA        NA        NA        NA
#> n4         NA        NA  NA 0.10662001        NA        NA        NA
#> n5         NA        NA  NA         NA        NA        NA        NA
#> n6         NA        NA  NA 0.77828948        NA        NA 0.3217528
#> n7         NA        NA  NA         NA 0.1353677        NA        NA
#> n8         NA        NA  NA         NA        NA        NA        NA
#> n9         NA        NA  NA         NA        NA 1.3205182        NA
#> n10        NA        NA  NA         NA        NA        NA        NA
#> n11        NA        NA  NA         NA        NA 0.9042982        NA
#> n12        NA        NA  NA         NA 0.1087531        NA        NA
#> n13        NA 1.4988099  NA 0.95774335        NA        NA        NA
#> n14        NA        NA  NA         NA        NA        NA        NA
#> n15        NA 0.7461832  NA         NA        NA 0.7094819        NA
#> n16 0.2385014        NA  NA         NA        NA        NA        NA
#> n17        NA        NA  NA 0.63179367        NA        NA        NA
#> n18        NA 0.1641781  NA         NA        NA        NA        NA
#> n19 2.1288816        NA  NA         NA        NA        NA 0.7372308
#> n20        NA        NA  NA 1.14938377        NA        NA        NA
#> n21        NA        NA  NA         NA        NA        NA        NA
#> n22        NA        NA  NA         NA        NA        NA        NA
#> n23 1.0210735        NA  NA         NA        NA        NA        NA
#> n24        NA        NA  NA         NA        NA 0.2566951        NA
#> n25        NA        NA  NA         NA        NA 0.3654268        NA
#> n26        NA        NA  NA         NA        NA        NA 0.1154464
#> n27        NA        NA  NA         NA        NA        NA        NA
#> n28        NA        NA  NA         NA        NA        NA 0.5389159
#> n29        NA        NA  NA         NA        NA        NA        NA
#> n30        NA        NA  NA         NA        NA        NA        NA
#> 
#> $dce_pvalue
#>     n1 n2        n3 n4       n5        n6         n7        n8        n9
#> n1  NA NA        NA NA       NA        NA         NA        NA        NA
#> n2  NA NA 0.4328003 NA       NA        NA         NA        NA 0.2025274
#> n3  NA NA        NA NA       NA        NA 0.06060227        NA 0.3507060
#> n4  NA NA        NA NA 0.768349 0.6836774 0.13154762        NA        NA
#> n5  NA NA        NA NA       NA 0.7807789         NA        NA        NA
#> n6  NA NA        NA NA       NA        NA         NA 0.5047117        NA
#> n7  NA NA        NA NA       NA        NA         NA        NA        NA
#> n8  NA NA        NA NA       NA        NA         NA        NA 0.6414896
#> n9  NA NA        NA NA       NA        NA         NA        NA        NA
#> n10 NA NA        NA NA       NA        NA         NA        NA        NA
#> n11 NA NA        NA NA       NA        NA         NA        NA        NA
#> n12 NA NA        NA NA       NA        NA         NA        NA        NA
#> n13 NA NA        NA NA       NA        NA         NA        NA        NA
#> n14 NA NA        NA NA       NA        NA         NA        NA        NA
#> n15 NA NA        NA NA       NA        NA         NA        NA        NA
#> n16 NA NA        NA NA       NA        NA         NA        NA        NA
#> n17 NA NA        NA NA       NA        NA         NA        NA        NA
#> n18 NA NA        NA NA       NA        NA         NA        NA        NA
#> n19 NA NA        NA NA       NA        NA         NA        NA        NA
#> n20 NA NA        NA NA       NA        NA         NA        NA        NA
#> n21 NA NA        NA NA       NA        NA         NA        NA        NA
#> n22 NA NA        NA NA       NA        NA         NA        NA        NA
#> n23 NA NA        NA NA       NA        NA         NA        NA        NA
#> n24 NA NA        NA NA       NA        NA         NA        NA        NA
#> n25 NA NA        NA NA       NA        NA         NA        NA        NA
#> n26 NA NA        NA NA       NA        NA         NA        NA        NA
#> n27 NA NA        NA NA       NA        NA         NA        NA        NA
#> n28 NA NA        NA NA       NA        NA         NA        NA        NA
#> n29 NA NA        NA NA       NA        NA         NA        NA        NA
#> n30 NA NA        NA NA       NA        NA         NA        NA        NA
#>           n10       n11       n12       n13       n14        n15       n16
#> n1         NA        NA        NA        NA 0.7039758         NA 0.6496532
#> n2         NA 0.1331380        NA        NA        NA         NA 0.5904042
#> n3  0.4124803        NA        NA        NA        NA         NA        NA
#> n4         NA        NA        NA        NA        NA 0.06456211        NA
#> n5         NA 0.7448139        NA        NA        NA 0.39082750        NA
#> n6         NA 0.6667047 0.1119573        NA        NA         NA        NA
#> n7         NA        NA        NA 0.2289845 0.7727186 0.14366840        NA
#> n8         NA        NA        NA        NA        NA         NA        NA
#> n9         NA        NA        NA        NA 0.5599066 0.75794951        NA
#> n10        NA        NA 0.7412948        NA        NA         NA        NA
#> n11        NA        NA 0.6863120 0.8914128        NA         NA        NA
#> n12        NA        NA        NA        NA        NA         NA        NA
#> n13        NA        NA        NA        NA        NA 0.44309664        NA
#> n14        NA        NA        NA        NA        NA         NA        NA
#> n15        NA        NA        NA        NA        NA         NA        NA
#> n16        NA        NA        NA        NA        NA         NA        NA
#> n17        NA        NA        NA        NA        NA         NA        NA
#> n18        NA        NA        NA        NA        NA         NA        NA
#> n19        NA        NA        NA        NA        NA         NA        NA
#> n20        NA        NA        NA        NA        NA         NA        NA
#> n21        NA        NA        NA        NA        NA         NA        NA
#> n22        NA        NA        NA        NA        NA         NA        NA
#> n23        NA        NA        NA        NA        NA         NA        NA
#> n24        NA        NA        NA        NA        NA         NA        NA
#> n25        NA        NA        NA        NA        NA         NA        NA
#> n26        NA        NA        NA        NA        NA         NA        NA
#> n27        NA        NA        NA        NA        NA         NA        NA
#> n28        NA        NA        NA        NA        NA         NA        NA
#> n29        NA        NA        NA        NA        NA         NA        NA
#> n30        NA        NA        NA        NA        NA         NA        NA
#>           n17       n18       n19       n20       n21       n22         n23
#> n1  0.1890158 0.2912775        NA        NA        NA        NA          NA
#> n2         NA        NA        NA        NA        NA 0.6517229          NA
#> n3         NA 0.2512688        NA        NA        NA        NA          NA
#> n4         NA 0.7587429 0.1178190        NA        NA        NA          NA
#> n5         NA        NA        NA        NA        NA        NA          NA
#> n6         NA        NA        NA        NA        NA        NA          NA
#> n7         NA 0.9043044        NA        NA        NA        NA 0.395369633
#> n8  0.7963817        NA        NA 0.8289327        NA        NA          NA
#> n9         NA        NA 0.8737821        NA        NA        NA 0.296340356
#> n10        NA        NA        NA        NA        NA        NA          NA
#> n11 0.6633656        NA        NA        NA        NA        NA 0.926795982
#> n12        NA        NA        NA 0.8440561        NA        NA 0.705146141
#> n13        NA        NA        NA        NA 0.1306701        NA          NA
#> n14        NA 0.5119469        NA        NA        NA        NA          NA
#> n15        NA        NA        NA        NA 0.9137881        NA          NA
#> n16        NA        NA        NA        NA 0.2180494        NA          NA
#> n17        NA        NA        NA 0.1688786        NA        NA          NA
#> n18        NA        NA        NA        NA        NA        NA          NA
#> n19        NA        NA        NA        NA 0.2429424        NA 0.103458079
#> n20        NA        NA        NA        NA 0.7811982        NA          NA
#> n21        NA        NA        NA        NA        NA        NA          NA
#> n22        NA        NA        NA        NA        NA        NA 0.008874592
#> n23        NA        NA        NA        NA        NA        NA          NA
#> n24        NA        NA        NA        NA        NA        NA          NA
#> n25        NA        NA        NA        NA        NA        NA          NA
#> n26        NA        NA        NA        NA        NA        NA          NA
#> n27        NA        NA        NA        NA        NA        NA          NA
#> n28        NA        NA        NA        NA        NA        NA          NA
#> n29        NA        NA        NA        NA        NA        NA          NA
#> n30        NA        NA        NA        NA        NA        NA          NA
#>           n24       n25 n26        n27       n28       n29        n30
#> n1         NA        NA  NA         NA        NA        NA         NA
#> n2         NA 0.2074595  NA 0.09949757        NA        NA 0.19333063
#> n3         NA        NA  NA         NA        NA        NA         NA
#> n4         NA        NA  NA 0.14679501        NA        NA         NA
#> n5         NA        NA  NA         NA        NA        NA         NA
#> n6         NA        NA  NA 0.70032114        NA        NA 0.04312352
#> n7         NA        NA  NA         NA 0.5949056        NA         NA
#> n8         NA        NA  NA         NA        NA        NA         NA
#> n9         NA        NA  NA         NA        NA 0.8677098         NA
#> n10        NA        NA  NA         NA        NA        NA         NA
#> n11        NA        NA  NA         NA        NA 0.9534528         NA
#> n12        NA        NA  NA         NA 0.8042367        NA         NA
#> n13        NA 0.8138499  NA 0.40767702        NA        NA         NA
#> n14        NA        NA  NA         NA        NA        NA         NA
#> n15        NA 0.2757362  NA         NA        NA 0.7677755         NA
#> n16 0.1270235        NA  NA         NA        NA        NA         NA
#> n17        NA        NA  NA 0.39862573        NA        NA         NA
#> n18        NA 0.6765181  NA         NA        NA        NA         NA
#> n19 0.9785833        NA  NA         NA        NA        NA 0.64139839
#> n20        NA        NA  NA 0.25564038        NA        NA         NA
#> n21        NA        NA  NA         NA        NA        NA         NA
#> n22        NA        NA  NA         NA        NA        NA         NA
#> n23 0.8326742        NA  NA         NA        NA        NA         NA
#> n24        NA        NA  NA         NA        NA 0.4003927         NA
#> n25        NA        NA  NA         NA        NA 0.6004568         NA
#> n26        NA        NA  NA         NA        NA        NA 0.89945399
#> n27        NA        NA  NA         NA        NA        NA         NA
#> n28        NA        NA  NA         NA        NA        NA 0.37249578
#> n29        NA        NA  NA         NA        NA        NA         NA
#> n30        NA        NA  NA         NA        NA        NA         NA
#> 
#> $pathway_pvalue
#> [1] 0.5048647
#> 
#> attr(,"class")
#> [1] "dce"