Differential Causal Effects - main function

Main function to compute differential causal effects and the pathway enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for igraph
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for graphNEL
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for matrix
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph
valid object defining a directed acyclic graph

df_expr_wt
data frame with wild type expression values

df_expr_mt
data from with mutation type expression values

solver
character with name of solver function

solver_args
additional arguments for the solver function. please adress this argument, if you use your own solver function. the default argument works with glm functions in the packages MASS, stats and glm2

adjustment_type
character string for the method to define the adjustment set Z for the regression

effect_type
method of computing causal effects

p_method
character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test
either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test. Alternatively, "vcovHC" can improve results for zero-inflated date, i.e., from single cell RNAseq experiments.

lib_size
either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding
indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative
logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level
Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,X.mt)

#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  0  1  0  0  0  0  0  0   0   0   0   0   1   1   1   1   0   0   0   0
#> n2   0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   0   0   1   0   0   0
#> n3   0  0  0  0  0  1  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n4   0  0  0  0  1  0  0  0  0   0   0   0   0   0   1   0   1   0   1   0   0
#> n5   0  0  0  0  0  1  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n6   0  0  0  0  0  0  0  0  0   1   0   0   0   0   1   0   0   0   0   0   1
#> n7   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n8   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n9   0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   1   0   0   0   1   0
#> n10  0  0  0  0  0  0  0  0  0   0   0   1   0   1   0   0   0   0   0   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   1   0   0   0   0   1
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   1   0   0   0
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0   0   1   0
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   1   0   0
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   1   1   0   0
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    0   0   1   1   1   0   1   1   0
#> n2    0   0   0   0   0   0   1   0   0
#> n3    1   0   0   0   0   0   0   0   0
#> n4    0   1   0   0   0   0   0   0   0
#> n5    0   0   0   1   0   0   0   0   0
#> n6    1   0   0   0   0   0   0   0   0
#> n7    0   0   0   0   0   0   0   0   1
#> n8    0   0   0   0   0   0   0   0   0
#> n9    0   0   0   1   0   0   0   0   0
#> n10   0   0   0   0   0   1   0   0   0
#> n11   0   0   0   0   0   0   0   0   1
#> n12   0   0   0   0   0   0   0   0   0
#> n13   0   0   0   0   0   0   0   0   0
#> n14   0   0   0   0   0   0   0   0   0
#> n15   1   0   0   0   0   0   0   0   0
#> n16   0   1   1   0   0   1   0   0   0
#> n17   0   0   1   0   0   0   0   0   1
#> n18   0   0   0   0   0   0   0   0   0
#> n19   0   0   0   0   0   0   0   0   0
#> n20   0   0   0   0   0   0   0   0   0
#> n21   0   0   0   0   1   0   0   0   0
#> n22   0   0   0   0   0   0   0   0   0
#> n23   0   0   0   0   0   1   0   0   0
#> n24   0   0   0   0   0   0   1   1   0
#> n25   0   0   0   0   0   0   0   0   0
#> n26   0   0   0   0   0   0   0   0   0
#> n27   0   0   0   0   0   0   0   0   0
#> n28   0   0   0   0   0   0   0   0   0
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1 n2          n3 n4         n5          n6 n7         n8 n9        n10
#> n1  NA NA 0.004446906 NA         NA          NA NA         NA NA         NA
#> n2  NA NA          NA NA         NA          NA NA 0.02195047 NA         NA
#> n3  NA NA          NA NA         NA -0.10882355 NA         NA NA         NA
#> n4  NA NA          NA NA 0.01202434          NA NA         NA NA         NA
#> n5  NA NA          NA NA         NA  0.08412434 NA         NA NA         NA
#> n6  NA NA          NA NA         NA          NA NA         NA NA -0.1592655
#> n7  NA NA          NA NA         NA          NA NA         NA NA         NA
#> n8  NA NA          NA NA         NA          NA NA         NA NA         NA
#> n9  NA NA          NA NA         NA          NA NA         NA NA         NA
#> n10 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n11 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n12 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n13 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n14 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n15 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n16 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n17 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n18 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n19 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n20 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n21 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n22 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n23 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n24 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n25 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n26 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n27 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n28 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n29 NA NA          NA NA         NA          NA NA         NA NA         NA
#> n30 NA NA          NA NA         NA          NA NA         NA NA         NA
#>             n11        n12 n13        n14          n15         n16         n17
#> n1           NA         NA  NA -0.0230885  0.009880227 -0.02258093 -0.17154210
#> n2           NA         NA  NA         NA           NA          NA          NA
#> n3           NA         NA  NA         NA           NA          NA          NA
#> n4           NA         NA  NA         NA -0.009786146          NA  0.08005795
#> n5           NA         NA  NA         NA           NA          NA          NA
#> n6           NA         NA  NA         NA  0.421949948          NA          NA
#> n7           NA         NA  NA         NA           NA          NA          NA
#> n8           NA         NA  NA         NA           NA          NA          NA
#> n9  -0.01931399         NA  NA         NA           NA  0.04194682          NA
#> n10          NA 0.01750228  NA -0.7103709           NA          NA          NA
#> n11          NA         NA  NA  0.5623870           NA -0.32408682          NA
#> n12          NA         NA  NA         NA           NA          NA          NA
#> n13          NA         NA  NA         NA           NA          NA  0.06782724
#> n14          NA         NA  NA         NA           NA          NA          NA
#> n15          NA         NA  NA         NA           NA          NA          NA
#> n16          NA         NA  NA         NA           NA          NA          NA
#> n17          NA         NA  NA         NA           NA          NA          NA
#> n18          NA         NA  NA         NA           NA          NA          NA
#> n19          NA         NA  NA         NA           NA          NA          NA
#> n20          NA         NA  NA         NA           NA          NA          NA
#> n21          NA         NA  NA         NA           NA          NA          NA
#> n22          NA         NA  NA         NA           NA          NA          NA
#> n23          NA         NA  NA         NA           NA          NA          NA
#> n24          NA         NA  NA         NA           NA          NA          NA
#> n25          NA         NA  NA         NA           NA          NA          NA
#> n26          NA         NA  NA         NA           NA          NA          NA
#> n27          NA         NA  NA         NA           NA          NA          NA
#> n28          NA         NA  NA         NA           NA          NA          NA
#> n29          NA         NA  NA         NA           NA          NA          NA
#> n30          NA         NA  NA         NA           NA          NA          NA
#>            n18        n19         n20         n21         n22         n23
#> n1          NA         NA          NA          NA          NA          NA
#> n2  0.10480319         NA          NA          NA          NA          NA
#> n3          NA         NA          NA          NA  0.11229107          NA
#> n4          NA 0.03918013          NA          NA          NA -0.01403817
#> n5          NA         NA          NA          NA          NA          NA
#> n6          NA         NA          NA -2.17031005 -0.16382283          NA
#> n7          NA         NA          NA  0.06284203          NA          NA
#> n8          NA         NA          NA          NA          NA          NA
#> n9          NA         NA -0.03351826          NA          NA          NA
#> n10         NA         NA          NA          NA          NA          NA
#> n11         NA         NA          NA -0.39607492          NA          NA
#> n12 0.35360501         NA          NA          NA          NA          NA
#> n13         NA         NA  0.01611799          NA          NA          NA
#> n14         NA         NA  1.86109513          NA          NA          NA
#> n15         NA         NA          NA          NA  0.04729577          NA
#> n16         NA 0.14752279          NA          NA          NA -0.08867837
#> n17 0.04397883 0.26253121          NA          NA          NA          NA
#> n18         NA         NA          NA          NA          NA          NA
#> n19         NA         NA          NA          NA          NA          NA
#> n20         NA         NA          NA          NA          NA          NA
#> n21         NA         NA          NA          NA          NA          NA
#> n22         NA         NA          NA          NA          NA          NA
#> n23         NA         NA          NA          NA          NA          NA
#> n24         NA         NA          NA          NA          NA          NA
#> n25         NA         NA          NA          NA          NA          NA
#> n26         NA         NA          NA          NA          NA          NA
#> n27         NA         NA          NA          NA          NA          NA
#> n28         NA         NA          NA          NA          NA          NA
#> n29         NA         NA          NA          NA          NA          NA
#> n30         NA         NA          NA          NA          NA          NA
#>             n24         n25       n26        n27        n28          n29
#> n1  -0.05487977  0.02395073 0.1172733         NA -0.1941715 -0.002244606
#> n2           NA          NA        NA         NA -0.3520006           NA
#> n3           NA          NA        NA         NA         NA           NA
#> n4           NA          NA        NA         NA         NA           NA
#> n5           NA -1.23990020        NA         NA         NA           NA
#> n6           NA          NA        NA         NA         NA           NA
#> n7           NA          NA        NA         NA         NA           NA
#> n8           NA          NA        NA         NA         NA           NA
#> n9           NA -0.04150957        NA         NA         NA           NA
#> n10          NA          NA        NA 0.55994927         NA           NA
#> n11          NA          NA        NA         NA         NA           NA
#> n12          NA          NA        NA         NA         NA           NA
#> n13          NA          NA        NA         NA         NA           NA
#> n14          NA          NA        NA         NA         NA           NA
#> n15          NA          NA        NA         NA         NA           NA
#> n16  0.11145093          NA        NA 0.08157446         NA           NA
#> n17  0.25553743          NA        NA         NA         NA           NA
#> n18          NA          NA        NA         NA         NA           NA
#> n19          NA          NA        NA         NA         NA           NA
#> n20          NA          NA        NA         NA         NA           NA
#> n21          NA          NA 0.3221807         NA         NA           NA
#> n22          NA          NA        NA         NA         NA           NA
#> n23          NA          NA        NA 0.13149386         NA           NA
#> n24          NA          NA        NA         NA  0.4075590  0.064234698
#> n25          NA          NA        NA         NA         NA           NA
#> n26          NA          NA        NA         NA         NA           NA
#> n27          NA          NA        NA         NA         NA           NA
#> n28          NA          NA        NA         NA         NA           NA
#> n29          NA          NA        NA         NA         NA           NA
#> n30          NA          NA        NA         NA         NA           NA
#>             n30
#> n1           NA
#> n2           NA
#> n3           NA
#> n4           NA
#> n5           NA
#> n6           NA
#> n7   0.06752828
#> n8           NA
#> n9           NA
#> n10          NA
#> n11 -0.06687664
#> n12          NA
#> n13          NA
#> n14          NA
#> n15          NA
#> n16          NA
#> n17 -0.48293346
#> n18          NA
#> n19          NA
#> n20          NA
#> n21          NA
#> n22          NA
#> n23          NA
#> n24          NA
#> n25          NA
#> n26          NA
#> n27          NA
#> n28          NA
#> n29          NA
#> n30          NA
#> 
#> $dce_stderr
#>     n1 n2          n3 n4         n5        n6 n7         n8 n9       n10
#> n1  NA NA 0.009209754 NA         NA        NA NA         NA NA        NA
#> n2  NA NA          NA NA         NA        NA NA 0.01315102 NA        NA
#> n3  NA NA          NA NA         NA 0.4523304 NA         NA NA        NA
#> n4  NA NA          NA NA 0.01525628        NA NA         NA NA        NA
#> n5  NA NA          NA NA         NA 0.2234551 NA         NA NA        NA
#> n6  NA NA          NA NA         NA        NA NA         NA NA 0.1391678
#> n7  NA NA          NA NA         NA        NA NA         NA NA        NA
#> n8  NA NA          NA NA         NA        NA NA         NA NA        NA
#> n9  NA NA          NA NA         NA        NA NA         NA NA        NA
#> n10 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n11 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n12 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n13 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n14 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n15 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n16 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n17 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n18 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n19 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n20 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n21 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n22 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n23 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n24 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n25 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n26 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n27 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n28 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n29 NA NA          NA NA         NA        NA NA         NA NA        NA
#> n30 NA NA          NA NA         NA        NA NA         NA NA        NA
#>            n11       n12 n13       n14        n15       n16       n17
#> n1          NA        NA  NA 0.0833231 0.08868819 0.0598536 0.1153415
#> n2          NA        NA  NA        NA         NA        NA        NA
#> n3          NA        NA  NA        NA         NA        NA        NA
#> n4          NA        NA  NA        NA 0.03187977        NA 0.1228826
#> n5          NA        NA  NA        NA         NA        NA        NA
#> n6          NA        NA  NA        NA 0.29594859        NA        NA
#> n7          NA        NA  NA        NA         NA        NA        NA
#> n8          NA        NA  NA        NA         NA        NA        NA
#> n9  0.01837896        NA  NA        NA         NA 0.1333229        NA
#> n10         NA 0.1163578  NA 1.3413981         NA        NA        NA
#> n11         NA        NA  NA 0.8284598         NA 1.0400524        NA
#> n12         NA        NA  NA        NA         NA        NA        NA
#> n13         NA        NA  NA        NA         NA        NA 0.0577193
#> n14         NA        NA  NA        NA         NA        NA        NA
#> n15         NA        NA  NA        NA         NA        NA        NA
#> n16         NA        NA  NA        NA         NA        NA        NA
#> n17         NA        NA  NA        NA         NA        NA        NA
#> n18         NA        NA  NA        NA         NA        NA        NA
#> n19         NA        NA  NA        NA         NA        NA        NA
#> n20         NA        NA  NA        NA         NA        NA        NA
#> n21         NA        NA  NA        NA         NA        NA        NA
#> n22         NA        NA  NA        NA         NA        NA        NA
#> n23         NA        NA  NA        NA         NA        NA        NA
#> n24         NA        NA  NA        NA         NA        NA        NA
#> n25         NA        NA  NA        NA         NA        NA        NA
#> n26         NA        NA  NA        NA         NA        NA        NA
#> n27         NA        NA  NA        NA         NA        NA        NA
#> n28         NA        NA  NA        NA         NA        NA        NA
#> n29         NA        NA  NA        NA         NA        NA        NA
#> n30         NA        NA  NA        NA         NA        NA        NA
#>            n18       n19        n20        n21        n22       n23        n24
#> n1          NA        NA         NA         NA         NA        NA 0.07468137
#> n2  0.04614738        NA         NA         NA         NA        NA         NA
#> n3          NA        NA         NA         NA 0.27833017        NA         NA
#> n4          NA 0.1932808         NA         NA         NA 0.0464228         NA
#> n5          NA        NA         NA         NA         NA        NA         NA
#> n6          NA        NA         NA 1.35468001 0.12373419        NA         NA
#> n7          NA        NA         NA 0.05368548         NA        NA         NA
#> n8          NA        NA         NA         NA         NA        NA         NA
#> n9          NA        NA 0.12081547         NA         NA        NA         NA
#> n10         NA        NA         NA         NA         NA        NA         NA
#> n11         NA        NA         NA 0.93711459         NA        NA         NA
#> n12 1.26595074        NA         NA         NA         NA        NA         NA
#> n13         NA        NA 0.06756581         NA         NA        NA         NA
#> n14         NA        NA 0.65279421         NA         NA        NA         NA
#> n15         NA        NA         NA         NA 0.08545895        NA         NA
#> n16         NA 0.7051035         NA         NA         NA 0.3534113 0.34195927
#> n17 0.44737266 0.4308435         NA         NA         NA        NA 0.37279557
#> n18         NA        NA         NA         NA         NA        NA         NA
#> n19         NA        NA         NA         NA         NA        NA         NA
#> n20         NA        NA         NA         NA         NA        NA         NA
#> n21         NA        NA         NA         NA         NA        NA         NA
#> n22         NA        NA         NA         NA         NA        NA         NA
#> n23         NA        NA         NA         NA         NA        NA         NA
#> n24         NA        NA         NA         NA         NA        NA         NA
#> n25         NA        NA         NA         NA         NA        NA         NA
#> n26         NA        NA         NA         NA         NA        NA         NA
#> n27         NA        NA         NA         NA         NA        NA         NA
#> n28         NA        NA         NA         NA         NA        NA         NA
#> n29         NA        NA         NA         NA         NA        NA         NA
#> n30         NA        NA         NA         NA         NA        NA         NA
#>            n25       n26       n27       n28        n29       n30
#> n1  0.15048931 0.0845978        NA 0.1316451 0.01492463        NA
#> n2          NA        NA        NA 0.1346937         NA        NA
#> n3          NA        NA        NA        NA         NA        NA
#> n4          NA        NA        NA        NA         NA        NA
#> n5  1.39727660        NA        NA        NA         NA        NA
#> n6          NA        NA        NA        NA         NA        NA
#> n7          NA        NA        NA        NA         NA 0.1126504
#> n8          NA        NA        NA        NA         NA        NA
#> n9  0.07132867        NA        NA        NA         NA        NA
#> n10         NA        NA 1.5075671        NA         NA        NA
#> n11         NA        NA        NA        NA         NA 1.3820861
#> n12         NA        NA        NA        NA         NA        NA
#> n13         NA        NA        NA        NA         NA        NA
#> n14         NA        NA        NA        NA         NA        NA
#> n15         NA        NA        NA        NA         NA        NA
#> n16         NA        NA 0.5158197        NA         NA        NA
#> n17         NA        NA        NA        NA         NA 0.8407294
#> n18         NA        NA        NA        NA         NA        NA
#> n19         NA        NA        NA        NA         NA        NA
#> n20         NA        NA        NA        NA         NA        NA
#> n21         NA 0.3518717        NA        NA         NA        NA
#> n22         NA        NA        NA        NA         NA        NA
#> n23         NA        NA 0.2800887        NA         NA        NA
#> n24         NA        NA        NA 0.9336488 0.05774465        NA
#> n25         NA        NA        NA        NA         NA        NA
#> n26         NA        NA        NA        NA         NA        NA
#> n27         NA        NA        NA        NA         NA        NA
#> n28         NA        NA        NA        NA         NA        NA
#> n29         NA        NA        NA        NA         NA        NA
#> n30         NA        NA        NA        NA         NA        NA
#> 
#> $dce_pvalue
#>     n1 n2        n3 n4        n5        n6 n7         n8 n9       n10       n11
#> n1  NA NA 0.6297429 NA        NA        NA NA         NA NA        NA        NA
#> n2  NA NA        NA NA        NA        NA NA 0.09669235 NA        NA        NA
#> n3  NA NA        NA NA        NA 0.8101314 NA         NA NA        NA        NA
#> n4  NA NA        NA NA 0.4315571        NA NA         NA NA        NA        NA
#> n5  NA NA        NA NA        NA 0.7069782 NA         NA NA        NA        NA
#> n6  NA NA        NA NA        NA        NA NA         NA NA 0.2538768        NA
#> n7  NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n8  NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n9  NA NA        NA NA        NA        NA NA         NA NA        NA 0.2946099
#> n10 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n11 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n12 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n13 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n14 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n15 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n16 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n17 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n18 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n19 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n20 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n21 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n22 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n23 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n24 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n25 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n26 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n27 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n28 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n29 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#> n30 NA NA        NA NA        NA        NA NA         NA NA        NA        NA
#>           n12 n13       n14       n15       n16       n17        n18       n19
#> n1         NA  NA 0.7819984 0.9114099 0.7063816 0.1385547         NA        NA
#> n2         NA  NA        NA        NA        NA        NA 0.02423061        NA
#> n3         NA  NA        NA        NA        NA        NA         NA        NA
#> n4         NA  NA        NA 0.7591918        NA 0.5154871         NA 0.8395712
#> n5         NA  NA        NA        NA        NA        NA         NA        NA
#> n6         NA  NA        NA 0.1555631        NA        NA         NA        NA
#> n7         NA  NA        NA        NA        NA        NA         NA        NA
#> n8         NA  NA        NA        NA        NA        NA         NA        NA
#> n9         NA  NA        NA        NA 0.7533804        NA         NA        NA
#> n10 0.8805912  NA 0.5970122        NA        NA        NA         NA        NA
#> n11        NA  NA 0.4980518        NA 0.7556744        NA         NA        NA
#> n12        NA  NA        NA        NA        NA        NA 0.78029703        NA
#> n13        NA  NA        NA        NA        NA 0.2413714         NA        NA
#> n14        NA  NA        NA        NA        NA        NA         NA        NA
#> n15        NA  NA        NA        NA        NA        NA         NA        NA
#> n16        NA  NA        NA        NA        NA        NA         NA 0.8344995
#> n17        NA  NA        NA        NA        NA        NA 0.92179398 0.5430254
#> n18        NA  NA        NA        NA        NA        NA         NA        NA
#> n19        NA  NA        NA        NA        NA        NA         NA        NA
#> n20        NA  NA        NA        NA        NA        NA         NA        NA
#> n21        NA  NA        NA        NA        NA        NA         NA        NA
#> n22        NA  NA        NA        NA        NA        NA         NA        NA
#> n23        NA  NA        NA        NA        NA        NA         NA        NA
#> n24        NA  NA        NA        NA        NA        NA         NA        NA
#> n25        NA  NA        NA        NA        NA        NA         NA        NA
#> n26        NA  NA        NA        NA        NA        NA         NA        NA
#> n27        NA  NA        NA        NA        NA        NA         NA        NA
#> n28        NA  NA        NA        NA        NA        NA         NA        NA
#> n29        NA  NA        NA        NA        NA        NA         NA        NA
#> n30        NA  NA        NA        NA        NA        NA         NA        NA
#>             n20       n21       n22       n23       n24       n25      n26
#> n1           NA        NA        NA        NA 0.4633078 0.8737127 0.167247
#> n2           NA        NA        NA        NA        NA        NA       NA
#> n3           NA        NA 0.6870648        NA        NA        NA       NA
#> n4           NA        NA        NA 0.7626691        NA        NA       NA
#> n5           NA        NA        NA        NA        NA 0.3759789       NA
#> n6           NA 0.1107811 0.1870804        NA        NA        NA       NA
#> n7           NA 0.2431973        NA        NA        NA        NA       NA
#> n8           NA        NA        NA        NA        NA        NA       NA
#> n9  0.781739633        NA        NA        NA        NA 0.5612705       NA
#> n10          NA        NA        NA        NA        NA        NA       NA
#> n11          NA 0.6730160        NA        NA        NA        NA       NA
#> n12          NA        NA        NA        NA        NA        NA       NA
#> n13 0.811701825        NA        NA        NA        NA        NA       NA
#> n14 0.004840403        NA        NA        NA        NA        NA       NA
#> n15          NA        NA 0.5806178        NA        NA        NA       NA
#> n16          NA        NA        NA 0.8021462 0.7448445        NA       NA
#> n17          NA        NA        NA        NA 0.4938875        NA       NA
#> n18          NA        NA        NA        NA        NA        NA       NA
#> n19          NA        NA        NA        NA        NA        NA       NA
#> n20          NA        NA        NA        NA        NA        NA       NA
#> n21          NA        NA        NA        NA        NA        NA 0.361027
#> n22          NA        NA        NA        NA        NA        NA       NA
#> n23          NA        NA        NA        NA        NA        NA       NA
#> n24          NA        NA        NA        NA        NA        NA       NA
#> n25          NA        NA        NA        NA        NA        NA       NA
#> n26          NA        NA        NA        NA        NA        NA       NA
#> n27          NA        NA        NA        NA        NA        NA       NA
#> n28          NA        NA        NA        NA        NA        NA       NA
#> n29          NA        NA        NA        NA        NA        NA       NA
#> n30          NA        NA        NA        NA        NA        NA       NA
#>           n27         n28       n29       n30
#> n1         NA 0.141827199 0.8806068        NA
#> n2         NA 0.009661838        NA        NA
#> n3         NA          NA        NA        NA
#> n4         NA          NA        NA        NA
#> n5         NA          NA        NA        NA
#> n6         NA          NA        NA        NA
#> n7         NA          NA        NA 0.5495653
#> n8         NA          NA        NA        NA
#> n9         NA          NA        NA        NA
#> n10 0.7107258          NA        NA        NA
#> n11        NA          NA        NA 0.9614567
#> n12        NA          NA        NA        NA
#> n13        NA          NA        NA        NA
#> n14        NA          NA        NA        NA
#> n15        NA          NA        NA        NA
#> n16 0.8745103          NA        NA        NA
#> n17        NA          NA        NA 0.5663617
#> n18        NA          NA        NA        NA
#> n19        NA          NA        NA        NA
#> n20        NA          NA        NA        NA
#> n21        NA          NA        NA        NA
#> n22        NA          NA        NA        NA
#> n23 0.6392649          NA        NA        NA
#> n24        NA 0.662953132 0.2673751        NA
#> n25        NA          NA        NA        NA
#> n26        NA          NA        NA        NA
#> n27        NA          NA        NA        NA
#> n28        NA          NA        NA        NA
#> n29        NA          NA        NA        NA
#> n30        NA          NA        NA        NA
#> 
#> $pathway_pvalue
#> [1] 0.2620021
#> 
#> attr(,"class")
#> [1] "dce"