Differential Causal Effects - main function

Main function to compute differential causal effects and the pathway enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "lm",
  solver_args = list(),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for igraph
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "lm",
  solver_args = list(),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for graphNEL
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "lm",
  solver_args = list(),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for matrix
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "lm",
  solver_args = list(),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph

valid object defining a directed acyclic graph

df_expr_wt

data frame with wild type expression values

df_expr_mt

data from with mutation type expression values

solver

character with name of solver function

solver_args

additional arguments for the solver function. please adress this argument, if you use your own solver function. the default argument works with glm functions in the packages MASS, stats and glm2

adjustment_type

character string for the method to define the adjustment set Z for the regression

effect_type

method of computing causal effects

p_method

character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test

either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test. Alternatively, "vcovHC" can improve results for zero-inflated date, i.e., from single cell RNAseq experiments.

lib_size

either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding

indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative

logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level

Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,X.mt)
#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  0  0  0  1  1  0  0  0   1   0   0   0   1   0   0   0   0   0   0   0
#> n2   0  0  0  1  0  0  0  0  0   0   0   0   0   1   0   1   0   0   0   0   1
#> n3   0  0  0  0  0  0  0  0  0   0   0   1   0   0   0   0   0   0   0   0   0
#> n4   0  0  0  0  1  0  0  1  0   0   0   1   0   0   1   0   0   0   0   0   0
#> n5   0  0  0  0  0  1  0  1  1   0   1   0   1   0   0   1   1   0   0   0   0
#> n6   0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n7   0  0  0  0  0  0  0  0  0   0   0   0   1   0   0   0   1   0   1   1   0
#> n8   0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   0   0   0   0   0   1
#> n9   0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   0   0   0   0   0   0
#> n10  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0   0   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   1   0   0   0   0   0   1   0   0   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   1   0   0   1   1   1   0   0   0
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   1   0   0   0
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   1   0   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    0   0   0   0   0   0   0   1   0
#> n2    0   0   0   0   0   0   1   0   0
#> n3    0   0   0   0   0   0   0   0   0
#> n4    0   0   0   0   0   0   0   0   0
#> n5    1   0   0   0   0   0   1   0   0
#> n6    0   0   0   0   0   0   0   0   0
#> n7    0   0   0   0   0   0   0   0   0
#> n8    0   1   0   0   0   0   0   1   0
#> n9    0   1   0   0   1   0   0   0   0
#> n10   0   0   0   0   0   0   0   0   0
#> n11   1   1   1   0   0   1   0   0   0
#> n12   0   0   0   0   0   0   0   0   0
#> n13   0   0   0   0   0   1   0   0   0
#> n14   1   0   0   0   0   1   0   0   0
#> n15   0   0   0   0   0   0   0   0   0
#> n16   0   0   0   0   0   0   0   0   0
#> n17   0   0   0   1   0   0   0   0   1
#> n18   0   0   0   0   0   0   1   0   0
#> n19   0   0   0   1   0   0   0   0   0
#> n20   0   0   1   0   0   0   0   0   0
#> n21   0   0   0   0   0   0   0   0   0
#> n22   0   0   0   0   0   0   0   0   0
#> n23   0   0   1   0   1   0   0   0   0
#> n24   0   0   0   1   1   0   0   0   0
#> n25   0   0   0   0   0   0   0   0   1
#> n26   0   0   0   0   0   0   0   0   0
#> n27   0   0   0   0   0   0   0   1   0
#> n28   0   0   0   0   0   0   0   0   1
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1 n2 n3          n4          n5          n6 n7          n8           n9
#> n1  NA NA NA          NA -0.01073367 0.003057568 NA          NA           NA
#> n2  NA NA NA 0.006786702          NA          NA NA          NA           NA
#> n3  NA NA NA          NA          NA          NA NA          NA           NA
#> n4  NA NA NA          NA  1.68969285          NA NA -2.11413387           NA
#> n5  NA NA NA          NA          NA 0.139043124 NA  0.06381703 -0.002973463
#> n6  NA NA NA          NA          NA          NA NA -0.21228935           NA
#> n7  NA NA NA          NA          NA          NA NA          NA           NA
#> n8  NA NA NA          NA          NA          NA NA          NA           NA
#> n9  NA NA NA          NA          NA          NA NA          NA           NA
#> n10 NA NA NA          NA          NA          NA NA          NA           NA
#> n11 NA NA NA          NA          NA          NA NA          NA           NA
#> n12 NA NA NA          NA          NA          NA NA          NA           NA
#> n13 NA NA NA          NA          NA          NA NA          NA           NA
#> n14 NA NA NA          NA          NA          NA NA          NA           NA
#> n15 NA NA NA          NA          NA          NA NA          NA           NA
#> n16 NA NA NA          NA          NA          NA NA          NA           NA
#> n17 NA NA NA          NA          NA          NA NA          NA           NA
#> n18 NA NA NA          NA          NA          NA NA          NA           NA
#> n19 NA NA NA          NA          NA          NA NA          NA           NA
#> n20 NA NA NA          NA          NA          NA NA          NA           NA
#> n21 NA NA NA          NA          NA          NA NA          NA           NA
#> n22 NA NA NA          NA          NA          NA NA          NA           NA
#> n23 NA NA NA          NA          NA          NA NA          NA           NA
#> n24 NA NA NA          NA          NA          NA NA          NA           NA
#> n25 NA NA NA          NA          NA          NA NA          NA           NA
#> n26 NA NA NA          NA          NA          NA NA          NA           NA
#> n27 NA NA NA          NA          NA          NA NA          NA           NA
#> n28 NA NA NA          NA          NA          NA NA          NA           NA
#> n29 NA NA NA          NA          NA          NA NA          NA           NA
#> n30 NA NA NA          NA          NA          NA NA          NA           NA
#>              n10         n11         n12         n13         n14       n15
#> n1  -0.008938252          NA          NA          NA 0.039319115        NA
#> n2            NA          NA          NA          NA 0.007655821        NA
#> n3            NA          NA  0.02693081          NA          NA        NA
#> n4            NA          NA -1.00038837          NA          NA 3.6124662
#> n5            NA 0.008637367          NA -0.13772265          NA        NA
#> n6            NA          NA          NA          NA          NA        NA
#> n7            NA          NA          NA  0.04338822          NA        NA
#> n8            NA 0.012933741          NA          NA          NA        NA
#> n9            NA          NA          NA          NA 2.125693823        NA
#> n10           NA          NA          NA          NA          NA 0.5183903
#> n11           NA          NA  1.64957189          NA          NA        NA
#> n12           NA          NA          NA -0.83623324          NA        NA
#> n13           NA          NA          NA          NA          NA 0.2217633
#> n14           NA          NA          NA          NA          NA        NA
#> n15           NA          NA          NA          NA          NA        NA
#> n16           NA          NA          NA          NA          NA        NA
#> n17           NA          NA          NA          NA          NA        NA
#> n18           NA          NA          NA          NA          NA        NA
#> n19           NA          NA          NA          NA          NA        NA
#> n20           NA          NA          NA          NA          NA        NA
#> n21           NA          NA          NA          NA          NA        NA
#> n22           NA          NA          NA          NA          NA        NA
#> n23           NA          NA          NA          NA          NA        NA
#> n24           NA          NA          NA          NA          NA        NA
#> n25           NA          NA          NA          NA          NA        NA
#> n26           NA          NA          NA          NA          NA        NA
#> n27           NA          NA          NA          NA          NA        NA
#> n28           NA          NA          NA          NA          NA        NA
#> n29           NA          NA          NA          NA          NA        NA
#> n30           NA          NA          NA          NA          NA        NA
#>            n16         n17         n18         n19          n20          n21
#> n1          NA          NA          NA          NA           NA           NA
#> n2   0.0711836          NA          NA          NA           NA  0.006358489
#> n3          NA          NA          NA          NA           NA           NA
#> n4          NA          NA          NA          NA           NA           NA
#> n5  -0.8581543  0.57662496          NA          NA           NA           NA
#> n6          NA          NA          NA          NA           NA           NA
#> n7          NA  0.05084262          NA -0.04662114 -0.003099715           NA
#> n8          NA          NA          NA          NA           NA -0.305659403
#> n9          NA          NA          NA          NA           NA           NA
#> n10         NA          NA          NA          NA           NA           NA
#> n11         NA          NA  3.04399416          NA           NA           NA
#> n12 -0.1494670 -0.07933935 -0.42046155          NA           NA           NA
#> n13         NA          NA  0.07871658          NA           NA           NA
#> n14         NA          NA          NA          NA           NA -0.491417731
#> n15         NA          NA          NA  0.35510209           NA           NA
#> n16         NA          NA          NA          NA           NA           NA
#> n17         NA          NA          NA          NA           NA           NA
#> n18         NA          NA          NA          NA           NA           NA
#> n19         NA          NA          NA          NA -0.026579583           NA
#> n20         NA          NA          NA          NA           NA           NA
#> n21         NA          NA          NA          NA           NA           NA
#> n22         NA          NA          NA          NA           NA           NA
#> n23         NA          NA          NA          NA           NA           NA
#> n24         NA          NA          NA          NA           NA           NA
#> n25         NA          NA          NA          NA           NA           NA
#> n26         NA          NA          NA          NA           NA           NA
#> n27         NA          NA          NA          NA           NA           NA
#> n28         NA          NA          NA          NA           NA           NA
#> n29         NA          NA          NA          NA           NA           NA
#> n30         NA          NA          NA          NA           NA           NA
#>             n22         n23        n24        n25         n26        n27
#> n1           NA          NA         NA         NA          NA         NA
#> n2           NA          NA         NA         NA          NA         NA
#> n3           NA          NA         NA         NA          NA         NA
#> n4           NA          NA         NA         NA          NA         NA
#> n5   0.32344302          NA         NA         NA          NA         NA
#> n6           NA          NA         NA         NA          NA         NA
#> n7           NA          NA         NA         NA          NA         NA
#> n8           NA  0.02092164         NA         NA          NA         NA
#> n9           NA  0.11936586         NA         NA -0.17822268         NA
#> n10          NA          NA         NA         NA          NA         NA
#> n11 -0.49912605 -0.08001357 0.05532222         NA          NA  3.0369722
#> n12          NA          NA         NA         NA          NA         NA
#> n13          NA          NA         NA         NA          NA  0.1257878
#> n14  0.07247266          NA         NA         NA          NA -0.9671473
#> n15          NA          NA         NA         NA          NA         NA
#> n16          NA          NA         NA         NA          NA         NA
#> n17          NA          NA         NA -0.4176647          NA         NA
#> n18          NA          NA         NA         NA          NA         NA
#> n19          NA          NA         NA -0.3051228          NA         NA
#> n20          NA          NA 0.03678291         NA          NA         NA
#> n21          NA          NA         NA         NA          NA         NA
#> n22          NA          NA         NA         NA          NA         NA
#> n23          NA          NA 0.27859183         NA -0.12475973         NA
#> n24          NA          NA         NA  5.6246622  0.02960317         NA
#> n25          NA          NA         NA         NA          NA         NA
#> n26          NA          NA         NA         NA          NA         NA
#> n27          NA          NA         NA         NA          NA         NA
#> n28          NA          NA         NA         NA          NA         NA
#> n29          NA          NA         NA         NA          NA         NA
#> n30          NA          NA         NA         NA          NA         NA
#>             n28         n29          n30
#> n1           NA -0.03140676           NA
#> n2   0.05056073          NA           NA
#> n3           NA          NA           NA
#> n4           NA          NA           NA
#> n5  -0.90532838          NA           NA
#> n6           NA          NA           NA
#> n7           NA          NA           NA
#> n8           NA  0.44658560           NA
#> n9           NA          NA           NA
#> n10          NA          NA           NA
#> n11          NA          NA           NA
#> n12          NA          NA           NA
#> n13          NA          NA           NA
#> n14          NA          NA           NA
#> n15          NA          NA           NA
#> n16          NA          NA           NA
#> n17          NA          NA -0.210000366
#> n18  0.02424096          NA           NA
#> n19          NA          NA           NA
#> n20          NA          NA           NA
#> n21          NA          NA           NA
#> n22          NA          NA           NA
#> n23          NA          NA           NA
#> n24          NA          NA           NA
#> n25          NA          NA  0.005612923
#> n26          NA          NA           NA
#> n27          NA -0.63418785           NA
#> n28          NA          NA -0.001917071
#> n29          NA          NA           NA
#> n30          NA          NA           NA
#> 
#> $dce_stderr
#>     n1 n2 n3          n4         n5         n6 n7        n8         n9
#> n1  NA NA NA          NA 0.01767034 0.01958585 NA        NA         NA
#> n2  NA NA NA 0.008091904         NA         NA NA        NA         NA
#> n3  NA NA NA          NA         NA         NA NA        NA         NA
#> n4  NA NA NA          NA 1.08308766         NA NA 1.6541779         NA
#> n5  NA NA NA          NA         NA 0.19834833 NA 0.2604775 0.03398234
#> n6  NA NA NA          NA         NA         NA NA 0.1533401         NA
#> n7  NA NA NA          NA         NA         NA NA        NA         NA
#> n8  NA NA NA          NA         NA         NA NA        NA         NA
#> n9  NA NA NA          NA         NA         NA NA        NA         NA
#> n10 NA NA NA          NA         NA         NA NA        NA         NA
#> n11 NA NA NA          NA         NA         NA NA        NA         NA
#> n12 NA NA NA          NA         NA         NA NA        NA         NA
#> n13 NA NA NA          NA         NA         NA NA        NA         NA
#> n14 NA NA NA          NA         NA         NA NA        NA         NA
#> n15 NA NA NA          NA         NA         NA NA        NA         NA
#> n16 NA NA NA          NA         NA         NA NA        NA         NA
#> n17 NA NA NA          NA         NA         NA NA        NA         NA
#> n18 NA NA NA          NA         NA         NA NA        NA         NA
#> n19 NA NA NA          NA         NA         NA NA        NA         NA
#> n20 NA NA NA          NA         NA         NA NA        NA         NA
#> n21 NA NA NA          NA         NA         NA NA        NA         NA
#> n22 NA NA NA          NA         NA         NA NA        NA         NA
#> n23 NA NA NA          NA         NA         NA NA        NA         NA
#> n24 NA NA NA          NA         NA         NA NA        NA         NA
#> n25 NA NA NA          NA         NA         NA NA        NA         NA
#> n26 NA NA NA          NA         NA         NA NA        NA         NA
#> n27 NA NA NA          NA         NA         NA NA        NA         NA
#> n28 NA NA NA          NA         NA         NA NA        NA         NA
#> n29 NA NA NA          NA         NA         NA NA        NA         NA
#> n30 NA NA NA          NA         NA         NA NA        NA         NA
#>           n10        n11        n12        n13        n14       n15       n16
#> n1  0.0209575         NA         NA         NA 0.10094215        NA        NA
#> n2         NA         NA         NA         NA 0.03419051        NA 0.1048539
#> n3         NA         NA 0.03430213         NA         NA        NA        NA
#> n4         NA         NA 2.68565579         NA         NA 2.0488130        NA
#> n5         NA 0.18385305         NA 1.49948525         NA        NA 0.9960194
#> n6         NA         NA         NA         NA         NA        NA        NA
#> n7         NA         NA         NA 0.05140395         NA        NA        NA
#> n8         NA 0.07271024         NA         NA         NA        NA        NA
#> n9         NA         NA         NA         NA 4.86773651        NA        NA
#> n10        NA         NA         NA         NA         NA 0.3589182        NA
#> n11        NA         NA 2.60174212         NA         NA        NA        NA
#> n12        NA         NA         NA 0.97458323         NA        NA 0.3170045
#> n13        NA         NA         NA         NA         NA 0.1952536        NA
#> n14        NA         NA         NA         NA         NA        NA        NA
#> n15        NA         NA         NA         NA         NA        NA        NA
#> n16        NA         NA         NA         NA         NA        NA        NA
#> n17        NA         NA         NA         NA         NA        NA        NA
#> n18        NA         NA         NA         NA         NA        NA        NA
#> n19        NA         NA         NA         NA         NA        NA        NA
#> n20        NA         NA         NA         NA         NA        NA        NA
#> n21        NA         NA         NA         NA         NA        NA        NA
#> n22        NA         NA         NA         NA         NA        NA        NA
#> n23        NA         NA         NA         NA         NA        NA        NA
#> n24        NA         NA         NA         NA         NA        NA        NA
#> n25        NA         NA         NA         NA         NA        NA        NA
#> n26        NA         NA         NA         NA         NA        NA        NA
#> n27        NA         NA         NA         NA         NA        NA        NA
#> n28        NA         NA         NA         NA         NA        NA        NA
#> n29        NA         NA         NA         NA         NA        NA        NA
#> n30        NA         NA         NA         NA         NA        NA        NA
#>            n17       n18        n19        n20       n21       n22        n23
#> n1          NA        NA         NA         NA        NA        NA         NA
#> n2          NA        NA         NA         NA 0.1078789        NA         NA
#> n3          NA        NA         NA         NA        NA        NA         NA
#> n4          NA        NA         NA         NA        NA        NA         NA
#> n5  0.86396175        NA         NA         NA        NA 0.2691272         NA
#> n6          NA        NA         NA         NA        NA        NA         NA
#> n7  0.06605873        NA 0.08442258 0.04938139        NA        NA         NA
#> n8          NA        NA         NA         NA 0.2790149        NA 0.05775513
#> n9          NA        NA         NA         NA        NA        NA 0.34788686
#> n10         NA        NA         NA         NA        NA        NA         NA
#> n11         NA 1.8873451         NA         NA        NA 0.6071818 0.10157152
#> n12 0.32475876 0.3911539         NA         NA        NA        NA         NA
#> n13         NA 0.1258606         NA         NA        NA        NA         NA
#> n14         NA        NA         NA         NA 0.2261940 0.1505149         NA
#> n15         NA        NA 0.26140854         NA        NA        NA         NA
#> n16         NA        NA         NA         NA        NA        NA         NA
#> n17         NA        NA         NA         NA        NA        NA         NA
#> n18         NA        NA         NA         NA        NA        NA         NA
#> n19         NA        NA         NA 0.08716191        NA        NA         NA
#> n20         NA        NA         NA         NA        NA        NA         NA
#> n21         NA        NA         NA         NA        NA        NA         NA
#> n22         NA        NA         NA         NA        NA        NA         NA
#> n23         NA        NA         NA         NA        NA        NA         NA
#> n24         NA        NA         NA         NA        NA        NA         NA
#> n25         NA        NA         NA         NA        NA        NA         NA
#> n26         NA        NA         NA         NA        NA        NA         NA
#> n27         NA        NA         NA         NA        NA        NA         NA
#> n28         NA        NA         NA         NA        NA        NA         NA
#> n29         NA        NA         NA         NA        NA        NA         NA
#> n30         NA        NA         NA         NA        NA        NA         NA
#>            n24       n25       n26       n27        n28        n29       n30
#> n1          NA        NA        NA        NA         NA 0.06176139        NA
#> n2          NA        NA        NA        NA 0.09852105         NA        NA
#> n3          NA        NA        NA        NA         NA         NA        NA
#> n4          NA        NA        NA        NA         NA         NA        NA
#> n5          NA        NA        NA        NA 0.83273222         NA        NA
#> n6          NA        NA        NA        NA         NA         NA        NA
#> n7          NA        NA        NA        NA         NA         NA        NA
#> n8          NA        NA        NA        NA         NA 0.45756497        NA
#> n9          NA        NA 0.3325846        NA         NA         NA        NA
#> n10         NA        NA        NA        NA         NA         NA        NA
#> n11 0.09136681        NA        NA 1.3960604         NA         NA        NA
#> n12         NA        NA        NA        NA         NA         NA        NA
#> n13         NA        NA        NA 0.1492839         NA         NA        NA
#> n14         NA        NA        NA 0.6340642         NA         NA        NA
#> n15         NA        NA        NA        NA         NA         NA        NA
#> n16         NA        NA        NA        NA         NA         NA        NA
#> n17         NA 0.3574099        NA        NA         NA         NA 0.1894028
#> n18         NA        NA        NA        NA 0.40097829         NA        NA
#> n19         NA 0.3094946        NA        NA         NA         NA        NA
#> n20 0.09527745        NA        NA        NA         NA         NA        NA
#> n21         NA        NA        NA        NA         NA         NA        NA
#> n22         NA        NA        NA        NA         NA         NA        NA
#> n23 0.12831205        NA 0.1439453        NA         NA         NA        NA
#> n24         NA 4.3873125 0.1522385        NA         NA         NA        NA
#> n25         NA        NA        NA        NA         NA         NA 0.0636206
#> n26         NA        NA        NA        NA         NA         NA        NA
#> n27         NA        NA        NA        NA         NA 0.37524363        NA
#> n28         NA        NA        NA        NA         NA         NA 0.2585592
#> n29         NA        NA        NA        NA         NA         NA        NA
#> n30         NA        NA        NA        NA         NA         NA        NA
#> 
#> $dce_pvalue
#>     n1 n2 n3        n4        n5        n6 n7        n8        n9       n10
#> n1  NA NA NA        NA 0.5442621 0.8761062 NA        NA        NA 0.6702158
#> n2  NA NA NA 0.4026579        NA        NA NA        NA        NA        NA
#> n3  NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n4  NA NA NA        NA 0.1203731        NA NA 0.2027564        NA        NA
#> n5  NA NA NA        NA        NA 0.4841487 NA 0.8067180 0.9303651        NA
#> n6  NA NA NA        NA        NA        NA NA 0.1678323        NA        NA
#> n7  NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n8  NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n9  NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n10 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n11 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n12 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n13 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n14 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n15 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n16 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n17 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n18 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n19 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n20 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n21 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n22 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n23 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n24 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n25 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n26 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n27 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n28 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n29 NA NA NA        NA        NA        NA NA        NA        NA        NA
#> n30 NA NA NA        NA        NA        NA NA        NA        NA        NA
#>           n11       n12       n13       n14        n15       n16       n17
#> n1         NA        NA        NA 0.6973135         NA        NA        NA
#> n2         NA        NA        NA 0.8230555         NA 0.4980123        NA
#> n3         NA 0.4333393        NA        NA         NA        NA        NA
#> n4         NA 0.7099323        NA        NA 0.07944118        NA        NA
#> n5  0.9625782        NA 0.9269157        NA         NA 0.3899913 0.5053056
#> n6         NA        NA        NA        NA         NA        NA        NA
#> n7         NA        NA 0.3996627        NA         NA        NA 0.4424294
#> n8  0.8590062        NA        NA        NA         NA        NA        NA
#> n9         NA        NA        NA 0.6628215         NA        NA        NA
#> n10        NA        NA        NA        NA 0.15026395        NA        NA
#> n11        NA 0.5268186        NA        NA         NA        NA        NA
#> n12        NA        NA 0.3919504        NA         NA 0.6378262 0.8072602
#> n13        NA        NA        NA        NA 0.25748320        NA        NA
#> n14        NA        NA        NA        NA         NA        NA        NA
#> n15        NA        NA        NA        NA         NA        NA        NA
#> n16        NA        NA        NA        NA         NA        NA        NA
#> n17        NA        NA        NA        NA         NA        NA        NA
#> n18        NA        NA        NA        NA         NA        NA        NA
#> n19        NA        NA        NA        NA         NA        NA        NA
#> n20        NA        NA        NA        NA         NA        NA        NA
#> n21        NA        NA        NA        NA         NA        NA        NA
#> n22        NA        NA        NA        NA         NA        NA        NA
#> n23        NA        NA        NA        NA         NA        NA        NA
#> n24        NA        NA        NA        NA         NA        NA        NA
#> n25        NA        NA        NA        NA         NA        NA        NA
#> n26        NA        NA        NA        NA         NA        NA        NA
#> n27        NA        NA        NA        NA         NA        NA        NA
#> n28        NA        NA        NA        NA         NA        NA        NA
#> n29        NA        NA        NA        NA         NA        NA        NA
#> n30        NA        NA        NA        NA         NA        NA        NA
#>           n18       n19       n20        n21       n22       n23        n24
#> n1         NA        NA        NA         NA        NA        NA         NA
#> n2         NA        NA        NA 0.95305918        NA        NA         NA
#> n3         NA        NA        NA         NA        NA        NA         NA
#> n4         NA        NA        NA         NA        NA        NA         NA
#> n5         NA        NA        NA         NA 0.2309126        NA         NA
#> n6         NA        NA        NA         NA        NA        NA         NA
#> n7         NA 0.5814162 0.9500129         NA        NA        NA         NA
#> n8         NA        NA        NA 0.27468663        NA 0.7175699         NA
#> n9         NA        NA        NA         NA        NA 0.7318820         NA
#> n10        NA        NA        NA         NA        NA        NA         NA
#> n11 0.1084207        NA        NA         NA 0.4120754 0.4318109 0.54556467
#> n12 0.2837710        NA        NA         NA        NA        NA         NA
#> n13 0.5324415        NA        NA         NA        NA        NA         NA
#> n14        NA        NA        NA 0.03105229 0.6307166        NA         NA
#> n15        NA 0.1759417        NA         NA        NA        NA         NA
#> n16        NA        NA        NA         NA        NA        NA         NA
#> n17        NA        NA        NA         NA        NA        NA         NA
#> n18        NA        NA        NA         NA        NA        NA         NA
#> n19        NA        NA 0.7607384         NA        NA        NA         NA
#> n20        NA        NA        NA         NA        NA        NA 0.69987893
#> n21        NA        NA        NA         NA        NA        NA         NA
#> n22        NA        NA        NA         NA        NA        NA         NA
#> n23        NA        NA        NA         NA        NA        NA 0.03115523
#> n24        NA        NA        NA         NA        NA        NA         NA
#> n25        NA        NA        NA         NA        NA        NA         NA
#> n26        NA        NA        NA         NA        NA        NA         NA
#> n27        NA        NA        NA         NA        NA        NA         NA
#> n28        NA        NA        NA         NA        NA        NA         NA
#> n29        NA        NA        NA         NA        NA        NA         NA
#> n30        NA        NA        NA         NA        NA        NA         NA
#>           n25       n26        n27       n28        n29       n30
#> n1         NA        NA         NA        NA 0.61166226        NA
#> n2         NA        NA         NA 0.6083913         NA        NA
#> n3         NA        NA         NA        NA         NA        NA
#> n4         NA        NA         NA        NA         NA        NA
#> n5         NA        NA         NA 0.2783210         NA        NA
#> n6         NA        NA         NA        NA         NA        NA
#> n7         NA        NA         NA        NA         NA        NA
#> n8         NA        NA         NA        NA 0.33030310        NA
#> n9         NA 0.5926610         NA        NA         NA        NA
#> n10        NA        NA         NA        NA         NA        NA
#> n11        NA        NA 0.03082292        NA         NA        NA
#> n12        NA        NA         NA        NA         NA        NA
#> n13        NA        NA 0.40050740        NA         NA        NA
#> n14        NA        NA 0.12884352        NA         NA        NA
#> n15        NA        NA         NA        NA         NA        NA
#> n16        NA        NA         NA        NA         NA        NA
#> n17 0.2440341        NA         NA        NA         NA 0.2689395
#> n18        NA        NA         NA 0.9518572         NA        NA
#> n19 0.3254355        NA         NA        NA         NA        NA
#> n20        NA        NA         NA        NA         NA        NA
#> n21        NA        NA         NA        NA         NA        NA
#> n22        NA        NA         NA        NA         NA        NA
#> n23        NA 0.3871906         NA        NA         NA        NA
#> n24 0.2013945 0.8460293         NA        NA         NA        NA
#> n25        NA        NA         NA        NA         NA 0.9297908
#> n26        NA        NA         NA        NA         NA        NA
#> n27        NA        NA         NA        NA 0.09265464        NA
#> n28        NA        NA         NA        NA         NA 0.9940920
#> n29        NA        NA         NA        NA         NA        NA
#> n30        NA        NA         NA        NA         NA        NA
#> 
#> $pathway_pvalue
#> [1] 0.790449
#> 
#> $deconfounding
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "dce"