Differential Causal Effects - main function

Main function to compute differential causal effects and the pathway enrichment

dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for igraph
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for graphNEL
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

# S4 method for matrix
dce(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver = "glm2",
  solver_args = list(method = glm.dce.fit),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph
valid object defining a directed acyclic graph

df_expr_wt
data frame with wild type expression values

df_expr_mt
data from with mutation type expression values

solver
character with name of solver function

solver_args
additional arguments for the solver function. please adress this argument, if you use your own solver function. the default argument works with glm functions in the packages MASS, stats and glm2

adjustment_type
character string for the method to define the adjustment set Z for the regression

effect_type
method of computing causal effects

p_method
character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test
either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size
either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding
indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative
logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level
Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce(dag,X.wt,X.mt)

#> Warning: Cycle(s) detected in network

#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  0  0  1  1  0  0  0  0   0   0   0   0   0   1   0   1   1   1   0   0
#> n2   0  0  0  0  0  1  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n3   0  0  0  1  0  0  0  0  0   1   0   0   0   0   1   0   0   0   0   0   0
#> n4   0  0  0  0  1  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n5   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n6   0  0  0  0  0  0  0  0  0   0   1   0   0   1   0   1   0   0   0   1   0
#> n7   0  0  0  0  0  0  0  1  0   0   0   1   0   0   0   1   0   0   0   0   0
#> n8   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   1   0   0   0   0
#> n9   0  0  0  0  0  0  0  0  0   1   0   1   0   0   1   0   0   1   1   0   0
#> n10  0  0  0  0  0  0  0  0  0   0   0   0   0   1   0   0   1   0   0   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   1   0   1   0   0   0   0   0   0   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   1   0   0   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   1   1   0
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    1   1   0   0   0   0   0   0   0
#> n2    0   0   1   0   0   0   0   0   0
#> n3    1   0   0   0   0   0   0   0   0
#> n4    0   0   0   0   0   0   0   0   1
#> n5    0   0   0   0   0   1   0   0   0
#> n6    0   0   0   0   0   0   0   0   0
#> n7    0   0   1   0   0   0   0   0   0
#> n8    0   0   0   0   0   0   0   0   1
#> n9    0   0   0   0   0   1   0   0   1
#> n10   0   0   0   0   0   0   0   0   0
#> n11   0   0   0   0   0   1   0   0   0
#> n12   1   0   0   1   0   0   0   0   0
#> n13   0   1   0   0   0   0   0   0   0
#> n14   0   0   0   1   0   0   0   0   0
#> n15   0   0   0   0   0   0   0   0   0
#> n16   0   0   0   0   0   0   0   0   0
#> n17   0   0   0   0   0   0   0   0   0
#> n18   0   0   0   0   1   0   1   0   0
#> n19   0   0   0   0   0   0   0   0   0
#> n20   0   1   0   0   0   0   0   0   0
#> n21   0   0   0   1   0   0   1   1   0
#> n22   0   0   1   0   0   0   0   0   0
#> n23   0   0   0   0   0   0   0   0   0
#> n24   0   0   0   0   0   0   0   0   0
#> n25   0   0   0   0   1   0   0   0   0
#> n26   0   0   0   0   0   0   1   0   0
#> n27   0   0   0   0   0   0   0   0   0
#> n28   0   0   0   0   0   0   0   1   0
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1 n2 n3         n4          n5         n6 n7          n8 n9          n10
#> n1  NA NA NA 0.03488675 -0.03626563         NA NA          NA NA           NA
#> n2  NA NA NA         NA          NA 0.01382682 NA          NA NA           NA
#> n3  NA NA NA 0.01474419          NA         NA NA          NA NA  0.008636008
#> n4  NA NA NA         NA -0.07481054         NA NA          NA NA           NA
#> n5  NA NA NA         NA          NA         NA NA          NA NA           NA
#> n6  NA NA NA         NA          NA         NA NA          NA NA           NA
#> n7  NA NA NA         NA          NA         NA NA -0.01162355 NA           NA
#> n8  NA NA NA         NA          NA         NA NA          NA NA           NA
#> n9  NA NA NA         NA          NA         NA NA          NA NA -0.005817926
#> n10 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n11 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n12 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n13 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n14 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n15 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n16 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n17 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n18 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n19 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n20 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n21 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n22 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n23 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n24 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n25 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n26 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n27 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n28 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n29 NA NA NA         NA          NA         NA NA          NA NA           NA
#> n30 NA NA NA         NA          NA         NA NA          NA NA           NA
#>             n11         n12 n13        n14         n15         n16         n17
#> n1           NA          NA  NA         NA -0.02654237          NA -0.02437838
#> n2           NA          NA  NA         NA          NA          NA          NA
#> n3           NA          NA  NA         NA -0.01014807          NA          NA
#> n4           NA          NA  NA         NA          NA          NA          NA
#> n5           NA          NA  NA         NA          NA          NA          NA
#> n6  -0.02419093          NA  NA -0.1308290          NA -0.60815999          NA
#> n7           NA -0.06463582  NA         NA          NA -0.01859214          NA
#> n8           NA          NA  NA         NA          NA -0.42049788  0.79568901
#> n9           NA  0.14982231  NA         NA -0.05018880          NA          NA
#> n10          NA          NA  NA -0.2051495          NA          NA -0.66549338
#> n11          NA -7.12713682  NA -0.4063843          NA          NA          NA
#> n12          NA          NA  NA         NA          NA          NA          NA
#> n13          NA          NA  NA         NA          NA          NA          NA
#> n14          NA          NA  NA         NA          NA          NA          NA
#> n15          NA          NA  NA         NA          NA          NA          NA
#> n16          NA          NA  NA         NA          NA          NA          NA
#> n17          NA          NA  NA         NA          NA          NA          NA
#> n18          NA          NA  NA         NA          NA          NA          NA
#> n19          NA          NA  NA         NA          NA          NA          NA
#> n20          NA          NA  NA         NA          NA          NA          NA
#> n21          NA          NA  NA         NA          NA          NA          NA
#> n22          NA          NA  NA         NA          NA          NA          NA
#> n23          NA          NA  NA         NA          NA          NA          NA
#> n24          NA          NA  NA         NA          NA          NA          NA
#> n25          NA          NA  NA         NA          NA          NA          NA
#> n26          NA          NA  NA         NA          NA          NA          NA
#> n27          NA          NA  NA         NA          NA          NA          NA
#> n28          NA          NA  NA         NA          NA          NA          NA
#> n29          NA          NA  NA         NA          NA          NA          NA
#> n30          NA          NA  NA         NA          NA          NA          NA
#>             n18         n19        n20         n21          n22        n23
#> n1  -0.02721095 -0.02244108         NA          NA  0.001981338 -0.2137190
#> n2           NA          NA         NA          NA           NA         NA
#> n3           NA          NA         NA          NA -0.027088627         NA
#> n4           NA          NA         NA -0.03594796           NA         NA
#> n5           NA          NA         NA          NA           NA         NA
#> n6           NA          NA -0.2284926          NA           NA         NA
#> n7           NA          NA         NA          NA           NA         NA
#> n8           NA          NA         NA          NA           NA         NA
#> n9   0.01306927 -0.01819117         NA          NA           NA         NA
#> n10          NA          NA         NA          NA           NA         NA
#> n11          NA          NA         NA          NA           NA         NA
#> n12          NA          NA         NA          NA  1.358905788         NA
#> n13          NA          NA         NA          NA           NA -0.1365796
#> n14 -3.66154670          NA         NA          NA           NA         NA
#> n15          NA          NA -0.2952312          NA           NA         NA
#> n16          NA          NA         NA  0.73885885           NA         NA
#> n17          NA          NA         NA  0.37496071           NA         NA
#> n18          NA  0.23145304  0.0505696          NA           NA         NA
#> n19          NA          NA         NA          NA           NA         NA
#> n20          NA          NA         NA          NA           NA  1.3000372
#> n21          NA          NA         NA          NA           NA         NA
#> n22          NA          NA         NA          NA           NA         NA
#> n23          NA          NA         NA          NA           NA         NA
#> n24          NA          NA         NA          NA           NA         NA
#> n25          NA          NA         NA          NA           NA         NA
#> n26          NA          NA         NA          NA           NA         NA
#> n27          NA          NA         NA          NA           NA         NA
#> n28          NA          NA         NA          NA           NA         NA
#> n29          NA          NA         NA          NA           NA         NA
#> n30          NA          NA         NA          NA           NA         NA
#>             n24         n25        n26         n27        n28        n29
#> n1           NA          NA         NA          NA         NA         NA
#> n2   0.06241717          NA         NA          NA         NA         NA
#> n3           NA          NA         NA          NA         NA         NA
#> n4           NA          NA         NA          NA         NA         NA
#> n5           NA          NA         NA -0.05793702         NA         NA
#> n6           NA          NA         NA          NA         NA         NA
#> n7   0.04159746          NA         NA          NA         NA         NA
#> n8           NA          NA         NA          NA         NA         NA
#> n9           NA          NA         NA -0.03599575         NA         NA
#> n10          NA          NA         NA          NA         NA         NA
#> n11          NA          NA         NA  2.33969143         NA         NA
#> n12          NA -0.34148965         NA          NA         NA         NA
#> n13          NA          NA         NA          NA         NA         NA
#> n14          NA -0.38896745         NA          NA         NA         NA
#> n15          NA          NA         NA          NA         NA         NA
#> n16          NA          NA         NA          NA         NA         NA
#> n17          NA          NA         NA          NA         NA         NA
#> n18          NA          NA -0.1491484          NA  0.2012927         NA
#> n19          NA          NA         NA          NA         NA         NA
#> n20          NA          NA         NA          NA         NA         NA
#> n21          NA -0.01710321         NA          NA -0.2896411  0.6388044
#> n22 -0.09538800          NA         NA          NA         NA         NA
#> n23          NA          NA         NA          NA         NA         NA
#> n24          NA          NA         NA          NA         NA         NA
#> n25          NA          NA -0.2868027          NA         NA         NA
#> n26          NA          NA         NA          NA  0.3122134         NA
#> n27          NA          NA         NA          NA         NA         NA
#> n28          NA          NA         NA          NA         NA -0.2075903
#> n29          NA          NA         NA          NA         NA         NA
#> n30          NA          NA         NA          NA         NA         NA
#>             n30
#> n1           NA
#> n2           NA
#> n3           NA
#> n4   0.57754573
#> n5           NA
#> n6           NA
#> n7           NA
#> n8  -0.51325038
#> n9  -0.06223313
#> n10          NA
#> n11          NA
#> n12          NA
#> n13          NA
#> n14          NA
#> n15          NA
#> n16          NA
#> n17          NA
#> n18          NA
#> n19          NA
#> n20          NA
#> n21          NA
#> n22          NA
#> n23          NA
#> n24          NA
#> n25          NA
#> n26          NA
#> n27          NA
#> n28          NA
#> n29          NA
#> n30          NA
#> 
#> $dce_stderr
#>     n1 n2 n3         n4        n5         n6 n7         n8 n9         n10
#> n1  NA NA NA 0.13033146 0.1238930         NA NA         NA NA          NA
#> n2  NA NA NA         NA        NA 0.01790668 NA         NA NA          NA
#> n3  NA NA NA 0.09831981        NA         NA NA         NA NA 0.016573349
#> n4  NA NA NA         NA 0.1489495         NA NA         NA NA          NA
#> n5  NA NA NA         NA        NA         NA NA         NA NA          NA
#> n6  NA NA NA         NA        NA         NA NA         NA NA          NA
#> n7  NA NA NA         NA        NA         NA NA 0.01041465 NA          NA
#> n8  NA NA NA         NA        NA         NA NA         NA NA          NA
#> n9  NA NA NA         NA        NA         NA NA         NA NA 0.007396727
#> n10 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n11 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n12 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n13 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n14 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n15 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n16 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n17 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n18 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n19 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n20 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n21 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n22 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n23 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n24 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n25 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n26 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n27 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n28 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n29 NA NA NA         NA        NA         NA NA         NA NA          NA
#> n30 NA NA NA         NA        NA         NA NA         NA NA          NA
#>            n11        n12 n13       n14       n15        n16        n17
#> n1          NA         NA  NA        NA 0.1674944         NA 0.02871891
#> n2          NA         NA  NA        NA        NA         NA         NA
#> n3          NA         NA  NA        NA 0.1166415         NA         NA
#> n4          NA         NA  NA        NA        NA         NA         NA
#> n5          NA         NA  NA        NA        NA         NA         NA
#> n6  0.03384966         NA  NA 0.1554555        NA 0.88399671         NA
#> n7          NA 0.04369356  NA        NA        NA 0.06622581         NA
#> n8          NA         NA  NA        NA        NA 1.23981357 2.41656711
#> n9          NA 0.09327373  NA        NA 0.1555272         NA         NA
#> n10         NA         NA  NA 0.9627026        NA         NA 2.99118822
#> n11         NA 3.90126008  NA 0.6997811        NA         NA         NA
#> n12         NA         NA  NA        NA        NA         NA         NA
#> n13         NA         NA  NA        NA        NA         NA         NA
#> n14         NA         NA  NA        NA        NA         NA         NA
#> n15         NA         NA  NA        NA        NA         NA         NA
#> n16         NA         NA  NA        NA        NA         NA         NA
#> n17         NA         NA  NA        NA        NA         NA         NA
#> n18         NA         NA  NA        NA        NA         NA         NA
#> n19         NA         NA  NA        NA        NA         NA         NA
#> n20         NA         NA  NA        NA        NA         NA         NA
#> n21         NA         NA  NA        NA        NA         NA         NA
#> n22         NA         NA  NA        NA        NA         NA         NA
#> n23         NA         NA  NA        NA        NA         NA         NA
#> n24         NA         NA  NA        NA        NA         NA         NA
#> n25         NA         NA  NA        NA        NA         NA         NA
#> n26         NA         NA  NA        NA        NA         NA         NA
#> n27         NA         NA  NA        NA        NA         NA         NA
#> n28         NA         NA  NA        NA        NA         NA         NA
#> n29         NA         NA  NA        NA        NA         NA         NA
#> n30         NA         NA  NA        NA        NA         NA         NA
#>           n18        n19       n20       n21        n22       n23        n24
#> n1  0.1354774 0.07419213        NA        NA 0.08648424 0.1217421         NA
#> n2         NA         NA        NA        NA         NA        NA 0.09814272
#> n3         NA         NA        NA        NA 0.13780305        NA         NA
#> n4         NA         NA        NA 0.2585920         NA        NA         NA
#> n5         NA         NA        NA        NA         NA        NA         NA
#> n6         NA         NA 0.9054720        NA         NA        NA         NA
#> n7         NA         NA        NA        NA         NA        NA 0.06975648
#> n8         NA         NA        NA        NA         NA        NA         NA
#> n9  0.1259127 0.10027849        NA        NA         NA        NA         NA
#> n10        NA         NA        NA        NA         NA        NA         NA
#> n11        NA         NA        NA        NA         NA        NA         NA
#> n12        NA         NA        NA        NA 0.89229032        NA         NA
#> n13        NA         NA        NA        NA         NA 0.1226555         NA
#> n14 2.0234830         NA        NA        NA         NA        NA         NA
#> n15        NA         NA 0.3970378        NA         NA        NA         NA
#> n16        NA         NA        NA 0.7631083         NA        NA         NA
#> n17        NA         NA        NA 0.3446947         NA        NA         NA
#> n18        NA 0.12508782 0.4209874        NA         NA        NA         NA
#> n19        NA         NA        NA        NA         NA        NA         NA
#> n20        NA         NA        NA        NA         NA 0.9858372         NA
#> n21        NA         NA        NA        NA         NA        NA         NA
#> n22        NA         NA        NA        NA         NA        NA 0.55301110
#> n23        NA         NA        NA        NA         NA        NA         NA
#> n24        NA         NA        NA        NA         NA        NA         NA
#> n25        NA         NA        NA        NA         NA        NA         NA
#> n26        NA         NA        NA        NA         NA        NA         NA
#> n27        NA         NA        NA        NA         NA        NA         NA
#> n28        NA         NA        NA        NA         NA        NA         NA
#> n29        NA         NA        NA        NA         NA        NA         NA
#> n30        NA         NA        NA        NA         NA        NA         NA
#>           n25       n26        n27      n28       n29       n30
#> n1         NA        NA         NA       NA        NA        NA
#> n2         NA        NA         NA       NA        NA        NA
#> n3         NA        NA         NA       NA        NA        NA
#> n4         NA        NA         NA       NA        NA 0.7620967
#> n5         NA        NA 0.23694050       NA        NA        NA
#> n6         NA        NA         NA       NA        NA        NA
#> n7         NA        NA         NA       NA        NA        NA
#> n8         NA        NA         NA       NA        NA 3.5518988
#> n9         NA        NA 0.05524854       NA        NA 0.1566672
#> n10        NA        NA         NA       NA        NA        NA
#> n11        NA        NA 2.47099933       NA        NA        NA
#> n12 0.2846417        NA         NA       NA        NA        NA
#> n13        NA        NA         NA       NA        NA        NA
#> n14 0.5762159        NA         NA       NA        NA        NA
#> n15        NA        NA         NA       NA        NA        NA
#> n16        NA        NA         NA       NA        NA        NA
#> n17        NA        NA         NA       NA        NA        NA
#> n18        NA 0.2800803         NA 0.232529        NA        NA
#> n19        NA        NA         NA       NA        NA        NA
#> n20        NA        NA         NA       NA        NA        NA
#> n21 0.1232539        NA         NA 1.366956 1.0833494        NA
#> n22        NA        NA         NA       NA        NA        NA
#> n23        NA        NA         NA       NA        NA        NA
#> n24        NA        NA         NA       NA        NA        NA
#> n25        NA 1.0024825         NA       NA        NA        NA
#> n26        NA        NA         NA 0.271860        NA        NA
#> n27        NA        NA         NA       NA        NA        NA
#> n28        NA        NA         NA       NA 0.1146551        NA
#> n29        NA        NA         NA       NA        NA        NA
#> n30        NA        NA         NA       NA        NA        NA
#> 
#> $dce_pvalue
#>     n1 n2 n3        n4        n5      n6 n7        n8 n9       n10       n11
#> n1  NA NA NA 0.7892294 0.7700478      NA NA        NA NA        NA        NA
#> n2  NA NA NA        NA        NA 0.44095 NA        NA NA        NA        NA
#> n3  NA NA NA 0.8809492        NA      NA NA        NA NA 0.6029007        NA
#> n4  NA NA NA        NA 0.6160641      NA NA        NA NA        NA        NA
#> n5  NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n6  NA NA NA        NA        NA      NA NA        NA NA        NA 0.4756794
#> n7  NA NA NA        NA        NA      NA NA 0.2657559 NA        NA        NA
#> n8  NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n9  NA NA NA        NA        NA      NA NA        NA NA 0.4324929        NA
#> n10 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n11 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n12 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n13 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n14 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n15 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n16 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n17 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n18 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n19 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n20 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n21 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n22 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n23 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n24 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n25 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n26 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n27 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n28 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n29 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#> n30 NA NA NA        NA        NA      NA NA        NA NA        NA        NA
#>           n12 n13       n14       n15       n16       n17        n18        n19
#> n1         NA  NA        NA 0.8742518        NA 0.3969938 0.84102232 0.76261260
#> n2         NA  NA        NA        NA        NA        NA         NA         NA
#> n3         NA  NA        NA 0.9307585        NA        NA         NA         NA
#> n4         NA  NA        NA        NA        NA        NA         NA         NA
#> n5         NA  NA        NA        NA        NA        NA         NA         NA
#> n6         NA  NA 0.4010564        NA 0.4922957        NA         NA         NA
#> n7  0.1406660  NA        NA        NA 0.7792070        NA         NA         NA
#> n8         NA  NA        NA        NA 0.7348544 0.7423112         NA         NA
#> n9  0.1098256  NA        NA 0.7472656        NA        NA 0.91743715 0.85623601
#> n10        NA  NA 0.8314770        NA        NA 0.8241731         NA         NA
#> n11 0.0692544  NA 0.5620960        NA        NA        NA         NA         NA
#> n12        NA  NA        NA        NA        NA        NA         NA         NA
#> n13        NA  NA        NA        NA        NA        NA         NA         NA
#> n14        NA  NA        NA        NA        NA        NA 0.07194962         NA
#> n15        NA  NA        NA        NA        NA        NA         NA         NA
#> n16        NA  NA        NA        NA        NA        NA         NA         NA
#> n17        NA  NA        NA        NA        NA        NA         NA         NA
#> n18        NA  NA        NA        NA        NA        NA         NA 0.06581954
#> n19        NA  NA        NA        NA        NA        NA         NA         NA
#> n20        NA  NA        NA        NA        NA        NA         NA         NA
#> n21        NA  NA        NA        NA        NA        NA         NA         NA
#> n22        NA  NA        NA        NA        NA        NA         NA         NA
#> n23        NA  NA        NA        NA        NA        NA         NA         NA
#> n24        NA  NA        NA        NA        NA        NA         NA         NA
#> n25        NA  NA        NA        NA        NA        NA         NA         NA
#> n26        NA  NA        NA        NA        NA        NA         NA         NA
#> n27        NA  NA        NA        NA        NA        NA         NA         NA
#> n28        NA  NA        NA        NA        NA        NA         NA         NA
#> n29        NA  NA        NA        NA        NA        NA         NA         NA
#> n30        NA  NA        NA        NA        NA        NA         NA         NA
#>           n20       n21       n22        n23       n24       n25      n26
#> n1         NA        NA 0.9817455 0.08073505        NA        NA       NA
#> n2         NA        NA        NA         NA 0.5255290        NA       NA
#> n3         NA        NA 0.8443639         NA        NA        NA       NA
#> n4         NA 0.8895846        NA         NA        NA        NA       NA
#> n5         NA        NA        NA         NA        NA        NA       NA
#> n6  0.8010405        NA        NA         NA        NA        NA       NA
#> n7         NA        NA        NA         NA 0.5516469        NA       NA
#> n8         NA        NA        NA         NA        NA        NA       NA
#> n9         NA        NA        NA         NA        NA        NA       NA
#> n10        NA        NA        NA         NA        NA        NA       NA
#> n11        NA        NA        NA         NA        NA        NA       NA
#> n12        NA        NA 0.1294360         NA        NA 0.2317429       NA
#> n13        NA        NA        NA 0.26684778        NA        NA       NA
#> n14        NA        NA        NA         NA        NA 0.5004724       NA
#> n15 0.4580463        NA        NA         NA        NA        NA       NA
#> n16        NA 0.3341640        NA         NA        NA        NA       NA
#> n17        NA 0.2780591        NA         NA        NA        NA       NA
#> n18 0.9045139        NA        NA         NA        NA        NA 0.594988
#> n19        NA        NA        NA         NA        NA        NA       NA
#> n20        NA        NA        NA 0.18885207        NA        NA       NA
#> n21        NA        NA        NA         NA        NA 0.8897836       NA
#> n22        NA        NA        NA         NA 0.8632372        NA       NA
#> n23        NA        NA        NA         NA        NA        NA       NA
#> n24        NA        NA        NA         NA        NA        NA       NA
#> n25        NA        NA        NA         NA        NA        NA 0.775119
#> n26        NA        NA        NA         NA        NA        NA       NA
#> n27        NA        NA        NA         NA        NA        NA       NA
#> n28        NA        NA        NA         NA        NA        NA       NA
#> n29        NA        NA        NA         NA        NA        NA       NA
#> n30        NA        NA        NA         NA        NA        NA       NA
#>           n27       n28        n29       n30
#> n1         NA        NA         NA        NA
#> n2         NA        NA         NA        NA
#> n3         NA        NA         NA        NA
#> n4         NA        NA         NA 0.4494772
#> n5  0.8070881        NA         NA        NA
#> n6         NA        NA         NA        NA
#> n7         NA        NA         NA        NA
#> n8         NA        NA         NA 0.8852554
#> n9  0.5154714        NA         NA 0.6916291
#> n10        NA        NA         NA        NA
#> n11 0.3448878        NA         NA        NA
#> n12        NA        NA         NA        NA
#> n13        NA        NA         NA        NA
#> n14        NA        NA         NA        NA
#> n15        NA        NA         NA        NA
#> n16        NA        NA         NA        NA
#> n17        NA        NA         NA        NA
#> n18        NA 0.3877645         NA        NA
#> n19        NA        NA         NA        NA
#> n20        NA        NA         NA        NA
#> n21        NA 0.8324218 0.55612127        NA
#> n22        NA        NA         NA        NA
#> n23        NA        NA         NA        NA
#> n24        NA        NA         NA        NA
#> n25        NA        NA         NA        NA
#> n26        NA 0.2522181         NA        NA
#> n27        NA        NA         NA        NA
#> n28        NA        NA 0.07178826        NA
#> n29        NA        NA         NA        NA
#> n30        NA        NA         NA        NA
#> 
#> $pathway_pvalue
#> [1] 0.5624535
#> 
#> attr(,"class")
#> [1] "dce"