Differential Causal Effects for negative binomial data

Shortcut for the main function to analyse negative binomial data

dce_nb(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver_args = list(method = "glm.dce.nb.fit", link = "identity"),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "hmp",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph

valid object defining a directed acyclic graph

df_expr_wt

data frame with wild type expression values

df_expr_mt

data from with mutation type expression values

solver_args

additional arguments for the solver function

adjustment_type

character string for the method to define the adjustment set Z for the regression

effect_type

method of computing causal effects

p_method

character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test

either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size

either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding

indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative

logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level

Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce_nb(dag,X.wt,X.mt)
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: iteration limit reached
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: iteration limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: iteration limit reached
#> Warning: alternation limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: alternation limit reached
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: alternation limit reached
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: inner loop 3; cannot correct step size
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: step size truncated due to increasing deviance
#> Warning: alternation limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: iteration limit reached
#> Warning: alternation limit reached
#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  0  0  1  0  1  1  0  0   0   0   0   0   0   0   0   1   0   0   0   1
#> n2   0  0  0  0  0  0  0  1  0   0   0   1   0   1   0   0   1   0   0   0   0
#> n3   0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   0   0   0   0   0   0
#> n4   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0   0   0   1
#> n5   0  0  0  0  0  0  0  0  0   1   0   0   0   0   0   1   0   0   1   0   0
#> n6   0  0  0  0  0  0  0  1  0   0   0   0   0   0   0   1   0   0   0   0   0
#> n7   0  0  0  0  0  0  0  0  0   0   1   0   0   0   0   0   0   0   0   0   0
#> n8   0  0  0  0  0  0  0  0  0   0   0   1   0   0   0   0   0   0   1   0   0
#> n9   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n10  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   1   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   0   0   0   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   1   1   0   0   0   0   0   0
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0   0   0   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   1   0
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    0   0   0   0   0   0   0   0   0
#> n2    1   0   0   0   0   0   0   0   0
#> n3    0   0   0   0   0   0   1   0   0
#> n4    0   0   0   0   0   0   0   0   0
#> n5    1   1   0   0   0   0   0   0   0
#> n6    0   1   0   0   0   0   0   1   0
#> n7    0   0   0   1   0   0   0   0   0
#> n8    0   0   0   0   0   0   0   0   0
#> n9    0   0   0   0   0   0   0   0   0
#> n10   0   0   0   0   0   1   0   0   0
#> n11   1   0   1   0   0   0   0   0   1
#> n12   0   0   0   0   1   0   0   0   0
#> n13   0   1   0   0   0   1   0   0   0
#> n14   0   0   0   0   0   0   0   0   0
#> n15   0   0   1   0   0   0   0   1   0
#> n16   0   0   0   0   0   0   1   0   0
#> n17   0   0   0   0   0   0   0   0   0
#> n18   0   0   0   1   0   0   0   0   0
#> n19   0   0   0   0   0   0   0   0   0
#> n20   0   0   1   0   0   0   0   1   0
#> n21   0   0   0   0   1   1   0   0   1
#> n22   0   0   0   0   1   0   0   0   0
#> n23   0   0   0   0   0   0   1   0   0
#> n24   0   0   0   0   0   0   0   0   0
#> n25   0   0   0   0   0   0   0   0   0
#> n26   0   0   0   0   0   0   0   0   1
#> n27   0   0   0   0   0   0   0   0   0
#> n28   0   0   0   0   0   0   0   0   0
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1 n2 n3          n4 n5          n6         n7           n8 n9          n10
#> n1  NA NA NA 0.009598431 NA 0.003833928 0.00281002           NA NA           NA
#> n2  NA NA NA          NA NA          NA         NA -0.006351443 NA           NA
#> n3  NA NA NA          NA NA          NA         NA           NA NA           NA
#> n4  NA NA NA          NA NA          NA         NA           NA NA           NA
#> n5  NA NA NA          NA NA          NA         NA           NA NA -0.001925493
#> n6  NA NA NA          NA NA          NA         NA -0.069200051 NA           NA
#> n7  NA NA NA          NA NA          NA         NA           NA NA           NA
#> n8  NA NA NA          NA NA          NA         NA           NA NA           NA
#> n9  NA NA NA          NA NA          NA         NA           NA NA           NA
#> n10 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n11 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n12 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n13 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n14 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n15 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n16 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n17 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n18 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n19 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n20 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n21 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n22 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n23 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n24 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n25 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n26 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n27 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n28 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n29 NA NA NA          NA NA          NA         NA           NA NA           NA
#> n30 NA NA NA          NA NA          NA         NA           NA NA           NA
#>            n11          n12 n13        n14          n15        n16         n17
#> n1          NA           NA  NA         NA           NA         NA  0.15264860
#> n2          NA -0.001201424  NA 0.37218517           NA         NA  0.02852566
#> n3  -0.1042891           NA  NA         NA           NA         NA          NA
#> n4          NA           NA  NA         NA           NA         NA -0.11754459
#> n5          NA           NA  NA         NA           NA -0.5356231          NA
#> n6          NA           NA  NA         NA           NA -0.3470092          NA
#> n7  -0.8510483           NA  NA         NA           NA         NA          NA
#> n8          NA  0.133898536  NA         NA           NA         NA          NA
#> n9          NA           NA  NA         NA           NA         NA          NA
#> n10         NA           NA  NA         NA           NA         NA          NA
#> n11         NA           NA  NA         NA           NA -0.1959073          NA
#> n12         NA           NA  NA         NA           NA         NA          NA
#> n13         NA           NA  NA 0.01825183 -0.004013498         NA          NA
#> n14         NA           NA  NA         NA  0.010352180         NA          NA
#> n15         NA           NA  NA         NA           NA         NA          NA
#> n16         NA           NA  NA         NA           NA         NA          NA
#> n17         NA           NA  NA         NA           NA         NA          NA
#> n18         NA           NA  NA         NA           NA         NA          NA
#> n19         NA           NA  NA         NA           NA         NA          NA
#> n20         NA           NA  NA         NA           NA         NA          NA
#> n21         NA           NA  NA         NA           NA         NA          NA
#> n22         NA           NA  NA         NA           NA         NA          NA
#> n23         NA           NA  NA         NA           NA         NA          NA
#> n24         NA           NA  NA         NA           NA         NA          NA
#> n25         NA           NA  NA         NA           NA         NA          NA
#> n26         NA           NA  NA         NA           NA         NA          NA
#> n27         NA           NA  NA         NA           NA         NA          NA
#> n28         NA           NA  NA         NA           NA         NA          NA
#> n29         NA           NA  NA         NA           NA         NA          NA
#> n30         NA           NA  NA         NA           NA         NA          NA
#>     n18       n19         n20       n21          n22        n23        n24
#> n1   NA        NA          NA 0.1964665           NA         NA         NA
#> n2   NA        NA          NA        NA  0.128929749         NA         NA
#> n3   NA        NA          NA        NA           NA         NA         NA
#> n4   NA        NA          NA 0.1151990           NA         NA         NA
#> n5   NA 0.2043502          NA        NA  0.003705346  0.1325395         NA
#> n6   NA        NA          NA        NA           NA -0.4332542         NA
#> n7   NA        NA          NA        NA           NA         NA         NA
#> n8   NA 1.8243552          NA        NA           NA         NA         NA
#> n9   NA        NA -0.05508431        NA           NA         NA         NA
#> n10  NA 0.3941495          NA        NA           NA         NA         NA
#> n11  NA        NA          NA        NA -0.563550381         NA -0.8216571
#> n12  NA        NA          NA        NA           NA         NA         NA
#> n13  NA        NA          NA        NA           NA -0.1758228         NA
#> n14  NA        NA          NA        NA           NA         NA         NA
#> n15  NA        NA          NA        NA           NA         NA -1.1836663
#> n16  NA        NA  0.37730613        NA           NA         NA         NA
#> n17  NA        NA -1.15635476        NA           NA         NA         NA
#> n18  NA        NA          NA        NA           NA         NA         NA
#> n19  NA        NA          NA        NA           NA         NA         NA
#> n20  NA        NA          NA 0.9508950           NA         NA -0.4145703
#> n21  NA        NA          NA        NA           NA         NA         NA
#> n22  NA        NA          NA        NA           NA         NA         NA
#> n23  NA        NA          NA        NA           NA         NA         NA
#> n24  NA        NA          NA        NA           NA         NA         NA
#> n25  NA        NA          NA        NA           NA         NA         NA
#> n26  NA        NA          NA        NA           NA         NA         NA
#> n27  NA        NA          NA        NA           NA         NA         NA
#> n28  NA        NA          NA        NA           NA         NA         NA
#> n29  NA        NA          NA        NA           NA         NA         NA
#> n30  NA        NA          NA        NA           NA         NA         NA
#>            n25        n26       n27         n28         n29         n30
#> n1          NA         NA        NA          NA          NA          NA
#> n2          NA         NA        NA          NA          NA          NA
#> n3          NA         NA        NA -0.01000148          NA          NA
#> n4          NA         NA        NA          NA          NA          NA
#> n5          NA         NA        NA          NA          NA          NA
#> n6          NA         NA        NA          NA  0.12422530          NA
#> n7   0.1333317         NA        NA          NA          NA          NA
#> n8          NA         NA        NA          NA          NA          NA
#> n9          NA         NA        NA          NA          NA          NA
#> n10         NA         NA 0.1265809          NA          NA          NA
#> n11         NA         NA        NA          NA          NA -0.17385888
#> n12         NA -0.2658790        NA          NA          NA          NA
#> n13         NA         NA 0.4434459          NA          NA          NA
#> n14         NA         NA        NA          NA          NA          NA
#> n15         NA         NA        NA          NA -2.36419723          NA
#> n16         NA         NA        NA  0.29425915          NA          NA
#> n17         NA         NA        NA          NA          NA          NA
#> n18 -0.1851477         NA        NA          NA          NA          NA
#> n19         NA         NA        NA          NA          NA          NA
#> n20         NA         NA        NA          NA  0.03254787          NA
#> n21         NA -0.3219329 1.4830725          NA          NA  0.28747208
#> n22         NA  0.3983095        NA          NA          NA          NA
#> n23         NA         NA        NA -1.00736809          NA          NA
#> n24         NA         NA        NA          NA          NA          NA
#> n25         NA         NA        NA          NA          NA          NA
#> n26         NA         NA        NA          NA          NA -0.08511788
#> n27         NA         NA        NA          NA          NA          NA
#> n28         NA         NA        NA          NA          NA          NA
#> n29         NA         NA        NA          NA          NA          NA
#> n30         NA         NA        NA          NA          NA          NA
#> 
#> $dce_stderr
#>     n1 n2 n3          n4 n5          n6          n7         n8 n9         n10
#> n1  NA NA NA 0.008863754 NA 0.009760324 0.008747657         NA NA          NA
#> n2  NA NA NA          NA NA          NA          NA 0.01415069 NA          NA
#> n3  NA NA NA          NA NA          NA          NA         NA NA          NA
#> n4  NA NA NA          NA NA          NA          NA         NA NA          NA
#> n5  NA NA NA          NA NA          NA          NA         NA NA 0.008069067
#> n6  NA NA NA          NA NA          NA          NA 0.25378572 NA          NA
#> n7  NA NA NA          NA NA          NA          NA         NA NA          NA
#> n8  NA NA NA          NA NA          NA          NA         NA NA          NA
#> n9  NA NA NA          NA NA          NA          NA         NA NA          NA
#> n10 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n11 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n12 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n13 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n14 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n15 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n16 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n17 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n18 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n19 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n20 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n21 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n22 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n23 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n24 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n25 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n26 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n27 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n28 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n29 NA NA NA          NA NA          NA          NA         NA NA          NA
#> n30 NA NA NA          NA NA          NA          NA         NA NA          NA
#>            n11         n12 n13        n14         n15       n16        n17 n18
#> n1          NA          NA  NA         NA          NA        NA 0.12834634  NA
#> n2          NA 0.009389227  NA 0.05440526          NA        NA 0.01182577  NA
#> n3  0.02382006          NA  NA         NA          NA        NA         NA  NA
#> n4          NA          NA  NA         NA          NA        NA 1.46803703  NA
#> n5          NA          NA  NA         NA          NA 0.1591343         NA  NA
#> n6          NA          NA  NA         NA          NA 1.3252605         NA  NA
#> n7  0.72634330          NA  NA         NA          NA        NA         NA  NA
#> n8          NA 0.228317702  NA         NA          NA        NA         NA  NA
#> n9          NA          NA  NA         NA          NA        NA         NA  NA
#> n10         NA          NA  NA         NA          NA        NA         NA  NA
#> n11         NA          NA  NA         NA          NA 1.3759610         NA  NA
#> n12         NA          NA  NA         NA          NA        NA         NA  NA
#> n13         NA          NA  NA 0.14649229 0.005847307        NA         NA  NA
#> n14         NA          NA  NA         NA 0.044609583        NA         NA  NA
#> n15         NA          NA  NA         NA          NA        NA         NA  NA
#> n16         NA          NA  NA         NA          NA        NA         NA  NA
#> n17         NA          NA  NA         NA          NA        NA         NA  NA
#> n18         NA          NA  NA         NA          NA        NA         NA  NA
#> n19         NA          NA  NA         NA          NA        NA         NA  NA
#> n20         NA          NA  NA         NA          NA        NA         NA  NA
#> n21         NA          NA  NA         NA          NA        NA         NA  NA
#> n22         NA          NA  NA         NA          NA        NA         NA  NA
#> n23         NA          NA  NA         NA          NA        NA         NA  NA
#> n24         NA          NA  NA         NA          NA        NA         NA  NA
#> n25         NA          NA  NA         NA          NA        NA         NA  NA
#> n26         NA          NA  NA         NA          NA        NA         NA  NA
#> n27         NA          NA  NA         NA          NA        NA         NA  NA
#> n28         NA          NA  NA         NA          NA        NA         NA  NA
#> n29         NA          NA  NA         NA          NA        NA         NA  NA
#> n30         NA          NA  NA         NA          NA        NA         NA  NA
#>            n19       n20        n21        n22        n23       n24        n25
#> n1          NA        NA 0.03582602         NA         NA        NA         NA
#> n2          NA        NA         NA 0.07056366         NA        NA         NA
#> n3          NA        NA         NA         NA         NA        NA         NA
#> n4          NA        NA 0.57085925         NA         NA        NA         NA
#> n5  0.01458461        NA         NA 0.07203955 0.05642308        NA         NA
#> n6          NA        NA         NA         NA 1.34168624        NA         NA
#> n7          NA        NA         NA         NA         NA        NA 0.97935535
#> n8  2.84129847        NA         NA         NA         NA        NA         NA
#> n9          NA 0.2197302         NA         NA         NA        NA         NA
#> n10 0.22487309        NA         NA         NA         NA        NA         NA
#> n11         NA        NA         NA 0.50861416         NA 0.6763962         NA
#> n12         NA        NA         NA         NA         NA        NA         NA
#> n13         NA        NA         NA         NA 0.20483030        NA         NA
#> n14         NA        NA         NA         NA         NA        NA         NA
#> n15         NA        NA         NA         NA         NA 1.6288884         NA
#> n16         NA 1.2092014         NA         NA         NA        NA         NA
#> n17         NA 1.2904043         NA         NA         NA        NA         NA
#> n18         NA        NA         NA         NA         NA        NA 0.02419575
#> n19         NA        NA         NA         NA         NA        NA         NA
#> n20         NA        NA 0.68834842         NA         NA 0.2392031         NA
#> n21         NA        NA         NA         NA         NA        NA         NA
#> n22         NA        NA         NA         NA         NA        NA         NA
#> n23         NA        NA         NA         NA         NA        NA         NA
#> n24         NA        NA         NA         NA         NA        NA         NA
#> n25         NA        NA         NA         NA         NA        NA         NA
#> n26         NA        NA         NA         NA         NA        NA         NA
#> n27         NA        NA         NA         NA         NA        NA         NA
#> n28         NA        NA         NA         NA         NA        NA         NA
#> n29         NA        NA         NA         NA         NA        NA         NA
#> n30         NA        NA         NA         NA         NA        NA         NA
#>           n26       n27        n28        n29        n30
#> n1         NA        NA         NA         NA         NA
#> n2         NA        NA         NA         NA         NA
#> n3         NA        NA 0.07873164         NA         NA
#> n4         NA        NA         NA         NA         NA
#> n5         NA        NA         NA         NA         NA
#> n6         NA        NA         NA 0.57403034         NA
#> n7         NA        NA         NA         NA         NA
#> n8         NA        NA         NA         NA         NA
#> n9         NA        NA         NA         NA         NA
#> n10        NA 1.7361622         NA         NA         NA
#> n11        NA        NA         NA         NA 0.22495834
#> n12 0.4799656        NA         NA         NA         NA
#> n13        NA 0.1692339         NA         NA         NA
#> n14        NA        NA         NA         NA         NA
#> n15        NA        NA         NA 1.71847662         NA
#> n16        NA        NA 0.25576850         NA         NA
#> n17        NA        NA         NA         NA         NA
#> n18        NA        NA         NA         NA         NA
#> n19        NA        NA         NA         NA         NA
#> n20        NA        NA         NA 0.09404295         NA
#> n21 0.5863109 1.2939338         NA         NA 0.15464222
#> n22 0.1898496        NA         NA         NA         NA
#> n23        NA        NA 0.52474749         NA         NA
#> n24        NA        NA         NA         NA         NA
#> n25        NA        NA         NA         NA         NA
#> n26        NA        NA         NA         NA 0.08704132
#> n27        NA        NA         NA         NA         NA
#> n28        NA        NA         NA         NA         NA
#> n29        NA        NA         NA         NA         NA
#> n30        NA        NA         NA         NA         NA
#> 
#> $dce_pvalue
#>     n1 n2 n3        n4 n5        n6        n7        n8 n9       n10
#> n1  NA NA NA 0.2788592 NA 0.6944617 0.7480352        NA NA        NA
#> n2  NA NA NA        NA NA        NA        NA 0.6535448 NA        NA
#> n3  NA NA NA        NA NA        NA        NA        NA NA        NA
#> n4  NA NA NA        NA NA        NA        NA        NA NA        NA
#> n5  NA NA NA        NA NA        NA        NA        NA NA 0.8113953
#> n6  NA NA NA        NA NA        NA        NA 0.7851060 NA        NA
#> n7  NA NA NA        NA NA        NA        NA        NA NA        NA
#> n8  NA NA NA        NA NA        NA        NA        NA NA        NA
#> n9  NA NA NA        NA NA        NA        NA        NA NA        NA
#> n10 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n11 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n12 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n13 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n14 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n15 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n16 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n17 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n18 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n19 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n20 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n21 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n22 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n23 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n24 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n25 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n26 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n27 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n28 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n29 NA NA NA        NA NA        NA        NA        NA NA        NA
#> n30 NA NA NA        NA NA        NA        NA        NA NA        NA
#>              n11       n12 n13          n14       n15          n16        n17
#> n1            NA        NA  NA           NA        NA           NA 0.23430235
#> n2            NA 0.8981824  NA 7.865415e-12        NA           NA 0.01585826
#> n3  1.196618e-05        NA  NA           NA        NA           NA         NA
#> n4            NA        NA  NA           NA        NA           NA 0.93618220
#> n5            NA        NA  NA           NA        NA 0.0007630678         NA
#> n6            NA        NA  NA           NA        NA 0.7934430581         NA
#> n7  2.413221e-01        NA  NA           NA        NA           NA         NA
#> n8            NA 0.5575684  NA           NA        NA           NA         NA
#> n9            NA        NA  NA           NA        NA           NA         NA
#> n10           NA        NA  NA           NA        NA           NA         NA
#> n11           NA        NA  NA           NA        NA 0.8867810082         NA
#> n12           NA        NA  NA           NA        NA           NA         NA
#> n13           NA        NA  NA 9.008462e-01 0.4924709           NA         NA
#> n14           NA        NA  NA           NA 0.8164900           NA         NA
#> n15           NA        NA  NA           NA        NA           NA         NA
#> n16           NA        NA  NA           NA        NA           NA         NA
#> n17           NA        NA  NA           NA        NA           NA         NA
#> n18           NA        NA  NA           NA        NA           NA         NA
#> n19           NA        NA  NA           NA        NA           NA         NA
#> n20           NA        NA  NA           NA        NA           NA         NA
#> n21           NA        NA  NA           NA        NA           NA         NA
#> n22           NA        NA  NA           NA        NA           NA         NA
#> n23           NA        NA  NA           NA        NA           NA         NA
#> n24           NA        NA  NA           NA        NA           NA         NA
#> n25           NA        NA  NA           NA        NA           NA         NA
#> n26           NA        NA  NA           NA        NA           NA         NA
#> n27           NA        NA  NA           NA        NA           NA         NA
#> n28           NA        NA  NA           NA        NA           NA         NA
#> n29           NA        NA  NA           NA        NA           NA         NA
#> n30           NA        NA  NA           NA        NA           NA         NA
#>     n18          n19       n20          n21        n22        n23        n24
#> n1   NA           NA        NA 4.160395e-08         NA         NA         NA
#> n2   NA           NA        NA           NA 0.06767859         NA         NA
#> n3   NA           NA        NA           NA         NA         NA         NA
#> n4   NA           NA        NA 8.400736e-01         NA         NA         NA
#> n5   NA 1.328450e-44        NA           NA 0.95897899 0.01882235         NA
#> n6   NA           NA        NA           NA         NA 0.74675761         NA
#> n7   NA           NA        NA           NA         NA         NA         NA
#> n8   NA 5.208180e-01        NA           NA         NA         NA         NA
#> n9   NA           NA 0.8020533           NA         NA         NA         NA
#> n10  NA 7.964248e-02        NA           NA         NA         NA         NA
#> n11  NA           NA        NA           NA 0.26785681         NA 0.22445876
#> n12  NA           NA        NA           NA         NA         NA         NA
#> n13  NA           NA        NA           NA         NA 0.39068101         NA
#> n14  NA           NA        NA           NA         NA         NA         NA
#> n15  NA           NA        NA           NA         NA         NA 0.46742739
#> n16  NA           NA 0.7550183           NA         NA         NA         NA
#> n17  NA           NA 0.3701897           NA         NA         NA         NA
#> n18  NA           NA        NA           NA         NA         NA         NA
#> n19  NA           NA        NA           NA         NA         NA         NA
#> n20  NA           NA        NA 1.671513e-01         NA         NA 0.08307242
#> n21  NA           NA        NA           NA         NA         NA         NA
#> n22  NA           NA        NA           NA         NA         NA         NA
#> n23  NA           NA        NA           NA         NA         NA         NA
#> n24  NA           NA        NA           NA         NA         NA         NA
#> n25  NA           NA        NA           NA         NA         NA         NA
#> n26  NA           NA        NA           NA         NA         NA         NA
#> n27  NA           NA        NA           NA         NA         NA         NA
#> n28  NA           NA        NA           NA         NA         NA         NA
#> n29  NA           NA        NA           NA         NA         NA         NA
#> n30  NA           NA        NA           NA         NA         NA         NA
#>              n25        n26         n27        n28       n29        n30
#> n1            NA         NA          NA         NA        NA         NA
#> n2            NA         NA          NA         NA        NA         NA
#> n3            NA         NA          NA 0.89891468        NA         NA
#> n4            NA         NA          NA         NA        NA         NA
#> n5            NA         NA          NA         NA        NA         NA
#> n6            NA         NA          NA         NA 0.8286690         NA
#> n7  8.917088e-01         NA          NA         NA        NA         NA
#> n8            NA         NA          NA         NA        NA         NA
#> n9            NA         NA          NA         NA        NA         NA
#> n10           NA         NA 0.941878969         NA        NA         NA
#> n11           NA         NA          NA         NA        NA 0.43961161
#> n12           NA 0.57961011          NA         NA        NA         NA
#> n13           NA         NA 0.008784912         NA        NA         NA
#> n14           NA         NA          NA         NA        NA         NA
#> n15           NA         NA          NA         NA 0.1688985         NA
#> n16           NA         NA          NA 0.24994202        NA         NA
#> n17           NA         NA          NA         NA        NA         NA
#> n18 1.977594e-14         NA          NA         NA        NA         NA
#> n19           NA         NA          NA         NA        NA         NA
#> n20           NA         NA          NA         NA 0.7292707         NA
#> n21           NA 0.58294907 0.251723432         NA        NA 0.06303429
#> n22           NA 0.03590276          NA         NA        NA         NA
#> n23           NA         NA          NA 0.05489332        NA         NA
#> n24           NA         NA          NA         NA        NA         NA
#> n25           NA         NA          NA         NA        NA         NA
#> n26           NA         NA          NA         NA        NA 0.32812282
#> n27           NA         NA          NA         NA        NA         NA
#> n28           NA         NA          NA         NA        NA         NA
#> n29           NA         NA          NA         NA        NA         NA
#> n30           NA         NA          NA         NA        NA         NA
#> 
#> $pathway_pvalue
#> [1] 7.306473e-43
#> 
#> $deconfounding
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "dce"