Differential Causal Effects for negative binomial data

Shortcut for the main function to analyse negative binomial data

dce_nb(
  graph,
  df_expr_wt,
  df_expr_mt,
  solver_args = list(method = "glm.dce.nb.fit", link = "identity"),
  adjustment_type = "parents",
  effect_type = "total",
  p_method = "mean",
  test = "wald",
  lib_size = FALSE,
  deconfounding = FALSE,
  conservative = FALSE,
  log_level = logger::INFO
)

Arguments

graph
valid object defining a directed acyclic graph

df_expr_wt
data frame with wild type expression values

df_expr_mt
data from with mutation type expression values

solver_args
additional arguments for the solver function

adjustment_type
character string for the method to define the adjustment set Z for the regression

effect_type
method of computing causal effects

p_method
character string. "mean", "sum" for standard summary functions, "hmp" for harmonic mean, "test" for the selfcontained test of package 'CombinePValue' or any method from package 'metap', e.g., "meanp" or "sump".

test
either "wald" for testing significance with the wald test or "lr" for using a likelihood ratio test

lib_size
either a numeric vector of the same length as the sum of wild type and mutant samples or a logical. If TRUE, it is recommended that both data sets include not only the genes included in the graph but all genes available in the original data set.

deconfounding
indicates whether adjustment against latent confounding is used. If FALSE, no adjustment is used, if TRUE it adjusts for confounding by automatically estimating the number of latent confounders. The estimated number of latent confounders can be chosen manually by setting this variable to some number.

conservative
logical; if TRUE, does not use the indicator variable for the variables in the adjustment set

log_level
Control verbosity (logger::INFO, logger::DEBUG, ...)

Value

list of matrices with dces and corresponding p-value

Examples

dag <- create_random_DAG(30, 0.2)
X.wt <- simulate_data(dag)
dag.mt <- resample_edge_weights(dag)
X.mt <- simulate_data(dag)
dce_nb(dag,X.wt,X.mt)

#> Warning: Cycle(s) detected in network

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: alternation limit reached

#> Warning: alternation limit reached

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations

#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations

#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: iteration limit reached

#> Warning: iteration limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations

#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: glm.fit2: algorithm did not converge. Try increasing the maximum iterations

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: alternation limit reached

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> Warning: step size truncated due to increasing deviance

#> Warning: step size truncated due to increasing deviance

#> Warning: inner loop 3; cannot correct step size

#> $graph
#>     n1 n2 n3 n4 n5 n6 n7 n8 n9 n10 n11 n12 n13 n14 n15 n16 n17 n18 n19 n20 n21
#> n1   0  0  0  1  0  0  0  0  0   0   0   1   0   0   0   0   0   0   0   0   0
#> n2   0  0  0  0  1  0  0  0  0   0   0   1   0   0   0   0   0   0   0   1   0
#> n3   0  0  0  1  0  1  0  1  0   0   0   0   1   1   1   0   0   0   0   0   1
#> n4   0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0   0   0   0
#> n5   0  0  0  0  0  0  0  0  0   1   0   0   0   1   0   0   0   0   0   0   0
#> n6   0  0  0  0  0  0  0  0  1   0   0   0   1   0   0   0   0   0   0   0   0
#> n7   0  0  0  0  0  0  0  0  0   1   0   0   0   0   1   0   1   0   1   0   0
#> n8   0  0  0  0  0  0  0  0  1   0   0   0   0   0   0   0   0   0   1   1   0
#> n9   0  0  0  0  0  0  0  0  0   1   0   0   0   0   0   0   0   0   0   0   0
#> n10  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   1   0   0   0   0   0
#> n11  0  0  0  0  0  0  0  0  0   0   0   0   1   1   0   0   0   1   0   1   0
#> n12  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n13  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n14  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0   0   0   0
#> n15  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   1   0   0
#> n16  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n17  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   1   0   0   1
#> n18  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   1
#> n19  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n20  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n21  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n22  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n23  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n24  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n25  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n26  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n27  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n28  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n29  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#> n30  0  0  0  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0   0
#>     n22 n23 n24 n25 n26 n27 n28 n29 n30
#> n1    0   0   0   0   0   0   0   0   0
#> n2    0   0   1   1   0   0   0   0   0
#> n3    0   0   0   0   0   0   0   0   0
#> n4    0   0   0   0   0   0   0   0   0
#> n5    0   1   1   0   0   0   0   0   0
#> n6    1   0   0   0   0   0   0   0   0
#> n7    0   1   0   0   0   1   0   0   0
#> n8    0   1   1   0   0   0   0   0   0
#> n9    1   0   0   0   0   0   1   0   0
#> n10   0   0   0   0   0   0   0   0   0
#> n11   0   0   0   0   0   1   0   0   0
#> n12   0   0   0   1   1   0   0   0   0
#> n13   0   0   0   0   0   0   0   0   0
#> n14   0   0   0   0   0   0   0   1   0
#> n15   0   0   0   0   1   0   0   0   0
#> n16   0   0   0   0   0   0   1   0   0
#> n17   0   0   0   0   0   0   0   0   0
#> n18   0   0   0   0   0   0   0   0   0
#> n19   0   0   0   0   0   0   0   1   0
#> n20   0   0   0   0   0   0   0   0   1
#> n21   0   0   0   0   0   0   1   0   1
#> n22   0   0   0   0   0   0   0   0   0
#> n23   0   0   0   1   0   1   0   1   0
#> n24   0   0   0   0   1   0   0   0   0
#> n25   0   0   0   0   0   0   0   0   0
#> n26   0   0   0   0   0   0   0   0   0
#> n27   0   0   0   0   0   0   0   0   1
#> n28   0   0   0   0   0   0   0   0   0
#> n29   0   0   0   0   0   0   0   0   0
#> n30   0   0   0   0   0   0   0   0   0
#> 
#> $dce
#>     n1 n2 n3           n4          n5          n6 n7          n8        n9
#> n1  NA NA NA -0.001915045          NA          NA NA          NA        NA
#> n2  NA NA NA           NA -0.00247593          NA NA          NA        NA
#> n3  NA NA NA  0.012004600          NA 0.006235398 NA -0.01157185        NA
#> n4  NA NA NA           NA          NA          NA NA          NA        NA
#> n5  NA NA NA           NA          NA          NA NA          NA        NA
#> n6  NA NA NA           NA          NA          NA NA          NA 0.9169369
#> n7  NA NA NA           NA          NA          NA NA          NA        NA
#> n8  NA NA NA           NA          NA          NA NA          NA 0.1356265
#> n9  NA NA NA           NA          NA          NA NA          NA        NA
#> n10 NA NA NA           NA          NA          NA NA          NA        NA
#> n11 NA NA NA           NA          NA          NA NA          NA        NA
#> n12 NA NA NA           NA          NA          NA NA          NA        NA
#> n13 NA NA NA           NA          NA          NA NA          NA        NA
#> n14 NA NA NA           NA          NA          NA NA          NA        NA
#> n15 NA NA NA           NA          NA          NA NA          NA        NA
#> n16 NA NA NA           NA          NA          NA NA          NA        NA
#> n17 NA NA NA           NA          NA          NA NA          NA        NA
#> n18 NA NA NA           NA          NA          NA NA          NA        NA
#> n19 NA NA NA           NA          NA          NA NA          NA        NA
#> n20 NA NA NA           NA          NA          NA NA          NA        NA
#> n21 NA NA NA           NA          NA          NA NA          NA        NA
#> n22 NA NA NA           NA          NA          NA NA          NA        NA
#> n23 NA NA NA           NA          NA          NA NA          NA        NA
#> n24 NA NA NA           NA          NA          NA NA          NA        NA
#> n25 NA NA NA           NA          NA          NA NA          NA        NA
#> n26 NA NA NA           NA          NA          NA NA          NA        NA
#> n27 NA NA NA           NA          NA          NA NA          NA        NA
#> n28 NA NA NA           NA          NA          NA NA          NA        NA
#> n29 NA NA NA           NA          NA          NA NA          NA        NA
#> n30 NA NA NA           NA          NA          NA NA          NA        NA
#>             n10 n11          n12         n13         n14         n15
#> n1           NA  NA -0.008882651          NA          NA          NA
#> n2           NA  NA  0.056247088          NA          NA          NA
#> n3           NA  NA           NA  0.05891037  0.08038121  0.14522904
#> n4           NA  NA           NA          NA          NA          NA
#> n5  0.580125439  NA           NA          NA -0.27459213          NA
#> n6           NA  NA           NA -1.10945859          NA          NA
#> n7  0.002657656  NA           NA          NA          NA -0.08590916
#> n8           NA  NA           NA          NA          NA          NA
#> n9  0.513313780  NA           NA          NA          NA          NA
#> n10          NA  NA           NA          NA          NA          NA
#> n11          NA  NA           NA -0.18372418  0.03834805          NA
#> n12          NA  NA           NA          NA          NA          NA
#> n13          NA  NA           NA          NA          NA          NA
#> n14          NA  NA           NA          NA          NA          NA
#> n15          NA  NA           NA          NA          NA          NA
#> n16          NA  NA           NA          NA          NA          NA
#> n17          NA  NA           NA          NA          NA          NA
#> n18          NA  NA           NA          NA          NA          NA
#> n19          NA  NA           NA          NA          NA          NA
#> n20          NA  NA           NA          NA          NA          NA
#> n21          NA  NA           NA          NA          NA          NA
#> n22          NA  NA           NA          NA          NA          NA
#> n23          NA  NA           NA          NA          NA          NA
#> n24          NA  NA           NA          NA          NA          NA
#> n25          NA  NA           NA          NA          NA          NA
#> n26          NA  NA           NA          NA          NA          NA
#> n27          NA  NA           NA          NA          NA          NA
#> n28          NA  NA           NA          NA          NA          NA
#> n29          NA  NA           NA          NA          NA          NA
#> n30          NA  NA           NA          NA          NA          NA
#>             n16         n17         n18         n19        n20         n21
#> n1           NA          NA          NA          NA         NA          NA
#> n2           NA          NA          NA          NA  0.1189344          NA
#> n3           NA          NA          NA          NA         NA -0.01923174
#> n4           NA -1.27314137          NA          NA         NA          NA
#> n5           NA          NA          NA          NA         NA          NA
#> n6           NA          NA          NA          NA         NA          NA
#> n7           NA -0.05862983          NA -0.03687252         NA          NA
#> n8           NA          NA          NA  0.22367603 -0.4491681          NA
#> n9           NA          NA          NA          NA         NA          NA
#> n10 -0.02870708          NA          NA          NA         NA          NA
#> n11          NA          NA -0.03423198          NA  0.1106612          NA
#> n12          NA          NA          NA          NA         NA          NA
#> n13          NA          NA          NA          NA         NA          NA
#> n14          NA -0.31616874          NA          NA         NA          NA
#> n15          NA          NA          NA  0.23604196         NA          NA
#> n16          NA          NA          NA          NA         NA          NA
#> n17          NA          NA  0.05285949          NA         NA -0.62365202
#> n18          NA          NA          NA          NA         NA  0.04692750
#> n19          NA          NA          NA          NA         NA          NA
#> n20          NA          NA          NA          NA         NA          NA
#> n21          NA          NA          NA          NA         NA          NA
#> n22          NA          NA          NA          NA         NA          NA
#> n23          NA          NA          NA          NA         NA          NA
#> n24          NA          NA          NA          NA         NA          NA
#> n25          NA          NA          NA          NA         NA          NA
#> n26          NA          NA          NA          NA         NA          NA
#> n27          NA          NA          NA          NA         NA          NA
#> n28          NA          NA          NA          NA         NA          NA
#> n29          NA          NA          NA          NA         NA          NA
#> n30          NA          NA          NA          NA         NA          NA
#>            n22         n23         n24           n25       n26        n27
#> n1          NA          NA          NA            NA        NA         NA
#> n2          NA          NA  0.01147046 -0.0547375469        NA         NA
#> n3          NA          NA          NA            NA        NA         NA
#> n4          NA          NA          NA            NA        NA         NA
#> n5          NA  0.73087768 -0.31975838            NA        NA         NA
#> n6  0.07548785          NA          NA            NA        NA         NA
#> n7          NA -0.01831014          NA            NA        NA -0.2351807
#> n8          NA -0.23316022 -0.11824637            NA        NA         NA
#> n9  0.08510950          NA          NA            NA        NA         NA
#> n10         NA          NA          NA            NA        NA         NA
#> n11         NA          NA          NA            NA        NA -0.0611826
#> n12         NA          NA          NA -0.0009812514 0.2234449         NA
#> n13         NA          NA          NA            NA        NA         NA
#> n14         NA          NA          NA            NA        NA         NA
#> n15         NA          NA          NA            NA 0.6869669         NA
#> n16         NA          NA          NA            NA        NA         NA
#> n17         NA          NA          NA            NA        NA         NA
#> n18         NA          NA          NA            NA        NA         NA
#> n19         NA          NA          NA            NA        NA         NA
#> n20         NA          NA          NA            NA        NA         NA
#> n21         NA          NA          NA            NA        NA         NA
#> n22         NA          NA          NA            NA        NA         NA
#> n23         NA          NA          NA  0.1657499078        NA  0.7043841
#> n24         NA          NA          NA            NA 0.3078053         NA
#> n25         NA          NA          NA            NA        NA         NA
#> n26         NA          NA          NA            NA        NA         NA
#> n27         NA          NA          NA            NA        NA         NA
#> n28         NA          NA          NA            NA        NA         NA
#> n29         NA          NA          NA            NA        NA         NA
#> n30         NA          NA          NA            NA        NA         NA
#>             n28         n29        n30
#> n1           NA          NA         NA
#> n2           NA          NA         NA
#> n3           NA          NA         NA
#> n4           NA          NA         NA
#> n5           NA          NA         NA
#> n6           NA          NA         NA
#> n7           NA          NA         NA
#> n8           NA          NA         NA
#> n9   0.08018817          NA         NA
#> n10          NA          NA         NA
#> n11          NA          NA         NA
#> n12          NA          NA         NA
#> n13          NA          NA         NA
#> n14          NA -0.05403548         NA
#> n15          NA          NA         NA
#> n16 -0.29248119          NA         NA
#> n17          NA          NA         NA
#> n18          NA          NA         NA
#> n19          NA  0.17140371         NA
#> n20          NA          NA -0.2749582
#> n21 -0.06465306          NA -2.8798966
#> n22          NA          NA         NA
#> n23          NA  0.02493643         NA
#> n24          NA          NA         NA
#> n25          NA          NA         NA
#> n26          NA          NA         NA
#> n27          NA          NA  0.4890587
#> n28          NA          NA         NA
#> n29          NA          NA         NA
#> n30          NA          NA         NA
#> 
#> $dce_stderr
#>     n1 n2 n3          n4          n5          n6 n7          n8         n9
#> n1  NA NA NA 0.012503711          NA          NA NA          NA         NA
#> n2  NA NA NA          NA 0.009709911          NA NA          NA         NA
#> n3  NA NA NA 0.007702071          NA 0.009422768 NA 0.009646737         NA
#> n4  NA NA NA          NA          NA          NA NA          NA         NA
#> n5  NA NA NA          NA          NA          NA NA          NA         NA
#> n6  NA NA NA          NA          NA          NA NA          NA 0.13588487
#> n7  NA NA NA          NA          NA          NA NA          NA         NA
#> n8  NA NA NA          NA          NA          NA NA          NA 0.07825841
#> n9  NA NA NA          NA          NA          NA NA          NA         NA
#> n10 NA NA NA          NA          NA          NA NA          NA         NA
#> n11 NA NA NA          NA          NA          NA NA          NA         NA
#> n12 NA NA NA          NA          NA          NA NA          NA         NA
#> n13 NA NA NA          NA          NA          NA NA          NA         NA
#> n14 NA NA NA          NA          NA          NA NA          NA         NA
#> n15 NA NA NA          NA          NA          NA NA          NA         NA
#> n16 NA NA NA          NA          NA          NA NA          NA         NA
#> n17 NA NA NA          NA          NA          NA NA          NA         NA
#> n18 NA NA NA          NA          NA          NA NA          NA         NA
#> n19 NA NA NA          NA          NA          NA NA          NA         NA
#> n20 NA NA NA          NA          NA          NA NA          NA         NA
#> n21 NA NA NA          NA          NA          NA NA          NA         NA
#> n22 NA NA NA          NA          NA          NA NA          NA         NA
#> n23 NA NA NA          NA          NA          NA NA          NA         NA
#> n24 NA NA NA          NA          NA          NA NA          NA         NA
#> n25 NA NA NA          NA          NA          NA NA          NA         NA
#> n26 NA NA NA          NA          NA          NA NA          NA         NA
#> n27 NA NA NA          NA          NA          NA NA          NA         NA
#> n28 NA NA NA          NA          NA          NA NA          NA         NA
#> n29 NA NA NA          NA          NA          NA NA          NA         NA
#> n30 NA NA NA          NA          NA          NA NA          NA         NA
#>           n10 n11        n12        n13        n14       n15        n16
#> n1         NA  NA 0.09939040         NA         NA        NA         NA
#> n2         NA  NA 0.02309315         NA         NA        NA         NA
#> n3         NA  NA         NA 0.12183379 0.05409754 0.0420840         NA
#> n4         NA  NA         NA         NA         NA        NA         NA
#> n5  1.4753155  NA         NA         NA 1.28232387        NA         NA
#> n6         NA  NA         NA 0.66650704         NA        NA         NA
#> n7  0.1779553  NA         NA         NA         NA 0.1506978         NA
#> n8         NA  NA         NA         NA         NA        NA         NA
#> n9  0.4983273  NA         NA         NA         NA        NA         NA
#> n10        NA  NA         NA         NA         NA        NA 0.01887388
#> n11        NA  NA         NA 0.01164669 0.15868305        NA         NA
#> n12        NA  NA         NA         NA         NA        NA         NA
#> n13        NA  NA         NA         NA         NA        NA         NA
#> n14        NA  NA         NA         NA         NA        NA         NA
#> n15        NA  NA         NA         NA         NA        NA         NA
#> n16        NA  NA         NA         NA         NA        NA         NA
#> n17        NA  NA         NA         NA         NA        NA         NA
#> n18        NA  NA         NA         NA         NA        NA         NA
#> n19        NA  NA         NA         NA         NA        NA         NA
#> n20        NA  NA         NA         NA         NA        NA         NA
#> n21        NA  NA         NA         NA         NA        NA         NA
#> n22        NA  NA         NA         NA         NA        NA         NA
#> n23        NA  NA         NA         NA         NA        NA         NA
#> n24        NA  NA         NA         NA         NA        NA         NA
#> n25        NA  NA         NA         NA         NA        NA         NA
#> n26        NA  NA         NA         NA         NA        NA         NA
#> n27        NA  NA         NA         NA         NA        NA         NA
#> n28        NA  NA         NA         NA         NA        NA         NA
#> n29        NA  NA         NA         NA         NA        NA         NA
#> n30        NA  NA         NA         NA         NA        NA         NA
#>            n17        n18        n19        n20         n21        n22
#> n1          NA         NA         NA         NA          NA         NA
#> n2          NA         NA         NA 0.09120377          NA         NA
#> n3          NA         NA         NA         NA 0.009619823         NA
#> n4  3.82968791         NA         NA         NA          NA         NA
#> n5          NA         NA         NA         NA          NA         NA
#> n6          NA         NA         NA         NA          NA 0.12696352
#> n7  0.06407428         NA 0.01309919         NA          NA         NA
#> n8          NA         NA 0.33486717 1.01741273          NA         NA
#> n9          NA         NA         NA         NA          NA 0.08779198
#> n10         NA         NA         NA         NA          NA         NA
#> n11         NA 0.08878281         NA 0.18947009          NA         NA
#> n12         NA         NA         NA         NA          NA         NA
#> n13         NA         NA         NA         NA          NA         NA
#> n14 1.18398864         NA         NA         NA          NA         NA
#> n15         NA         NA 0.06869856         NA          NA         NA
#> n16         NA         NA         NA         NA          NA         NA
#> n17         NA 0.26187188         NA         NA 0.118171968         NA
#> n18         NA         NA         NA         NA 0.464955398         NA
#> n19         NA         NA         NA         NA          NA         NA
#> n20         NA         NA         NA         NA          NA         NA
#> n21         NA         NA         NA         NA          NA         NA
#> n22         NA         NA         NA         NA          NA         NA
#> n23         NA         NA         NA         NA          NA         NA
#> n24         NA         NA         NA         NA          NA         NA
#> n25         NA         NA         NA         NA          NA         NA
#> n26         NA         NA         NA         NA          NA         NA
#> n27         NA         NA         NA         NA          NA         NA
#> n28         NA         NA         NA         NA          NA         NA
#> n29         NA         NA         NA         NA          NA         NA
#> n30         NA         NA         NA         NA          NA         NA
#>            n23       n24        n25       n26       n27       n28       n29
#> n1          NA        NA         NA        NA        NA        NA        NA
#> n2          NA 0.1333830 0.01528856        NA        NA        NA        NA
#> n3          NA        NA         NA        NA        NA        NA        NA
#> n4          NA        NA         NA        NA        NA        NA        NA
#> n5  0.54472853 1.1769770         NA        NA        NA        NA        NA
#> n6          NA        NA         NA        NA        NA        NA        NA
#> n7  0.07269033        NA         NA        NA 0.1339378        NA        NA
#> n8  0.43414567 0.4004135         NA        NA        NA        NA        NA
#> n9          NA        NA         NA        NA        NA 0.2090260        NA
#> n10         NA        NA         NA        NA        NA        NA        NA
#> n11         NA        NA         NA        NA 0.1023979        NA        NA
#> n12         NA        NA 0.11960225 0.2130493        NA        NA        NA
#> n13         NA        NA         NA        NA        NA        NA        NA
#> n14         NA        NA         NA        NA        NA        NA 0.2386859
#> n15         NA        NA         NA 0.5787444        NA        NA        NA
#> n16         NA        NA         NA        NA        NA 4.2348848        NA
#> n17         NA        NA         NA        NA        NA        NA        NA
#> n18         NA        NA         NA        NA        NA        NA        NA
#> n19         NA        NA         NA        NA        NA        NA 1.5326990
#> n20         NA        NA         NA        NA        NA        NA        NA
#> n21         NA        NA         NA        NA        NA 0.4525824        NA
#> n22         NA        NA         NA        NA        NA        NA        NA
#> n23         NA        NA 0.22293239        NA 0.9527962        NA 0.3004643
#> n24         NA        NA         NA 0.2063593        NA        NA        NA
#> n25         NA        NA         NA        NA        NA        NA        NA
#> n26         NA        NA         NA        NA        NA        NA        NA
#> n27         NA        NA         NA        NA        NA        NA        NA
#> n28         NA        NA         NA        NA        NA        NA        NA
#> n29         NA        NA         NA        NA        NA        NA        NA
#> n30         NA        NA         NA        NA        NA        NA        NA
#>           n30
#> n1         NA
#> n2         NA
#> n3         NA
#> n4         NA
#> n5         NA
#> n6         NA
#> n7         NA
#> n8         NA
#> n9         NA
#> n10        NA
#> n11        NA
#> n12        NA
#> n13        NA
#> n14        NA
#> n15        NA
#> n16        NA
#> n17        NA
#> n18        NA
#> n19        NA
#> n20 0.6450921
#> n21 1.1830747
#> n22        NA
#> n23        NA
#> n24        NA
#> n25        NA
#> n26        NA
#> n27 0.7433656
#> n28        NA
#> n29        NA
#> n30        NA
#> 
#> $dce_pvalue
#>     n1 n2 n3        n4        n5        n6 n7      n8           n9       n10
#> n1  NA NA NA 0.8782735        NA        NA NA      NA           NA        NA
#> n2  NA NA NA        NA 0.7987309        NA NA      NA           NA        NA
#> n3  NA NA NA 0.1190864        NA 0.5081396 NA 0.23031           NA        NA
#> n4  NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n5  NA NA NA        NA        NA        NA NA      NA           NA 0.6941561
#> n6  NA NA NA        NA        NA        NA NA      NA 1.500050e-11        NA
#> n7  NA NA NA        NA        NA        NA NA      NA           NA 0.9880845
#> n8  NA NA NA        NA        NA        NA NA      NA 8.308513e-02        NA
#> n9  NA NA NA        NA        NA        NA NA      NA           NA 0.3029755
#> n10 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n11 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n12 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n13 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n14 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n15 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n16 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n17 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n18 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n19 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n20 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n21 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n22 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n23 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n24 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n25 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n26 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n27 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n28 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n29 NA NA NA        NA        NA        NA NA      NA           NA        NA
#> n30 NA NA NA        NA        NA        NA NA      NA           NA        NA
#>     n11        n12          n13       n14          n15       n16       n17
#> n1   NA 0.92878682           NA        NA           NA        NA        NA
#> n2   NA 0.01486459           NA        NA           NA        NA        NA
#> n3   NA         NA 6.287190e-01 0.1373169 0.0005586543        NA        NA
#> n4   NA         NA           NA        NA           NA        NA 0.7395571
#> n5   NA         NA           NA 0.8304407           NA        NA        NA
#> n6   NA         NA 9.599529e-02        NA           NA        NA        NA
#> n7   NA         NA           NA        NA 0.5686262477        NA 0.3601764
#> n8   NA         NA           NA        NA           NA        NA        NA
#> n9   NA         NA           NA        NA           NA        NA        NA
#> n10  NA         NA           NA        NA           NA 0.1282611        NA
#> n11  NA         NA 4.638445e-56 0.8090402           NA        NA        NA
#> n12  NA         NA           NA        NA           NA        NA        NA
#> n13  NA         NA           NA        NA           NA        NA        NA
#> n14  NA         NA           NA        NA           NA        NA 0.7894407
#> n15  NA         NA           NA        NA           NA        NA        NA
#> n16  NA         NA           NA        NA           NA        NA        NA
#> n17  NA         NA           NA        NA           NA        NA        NA
#> n18  NA         NA           NA        NA           NA        NA        NA
#> n19  NA         NA           NA        NA           NA        NA        NA
#> n20  NA         NA           NA        NA           NA        NA        NA
#> n21  NA         NA           NA        NA           NA        NA        NA
#> n22  NA         NA           NA        NA           NA        NA        NA
#> n23  NA         NA           NA        NA           NA        NA        NA
#> n24  NA         NA           NA        NA           NA        NA        NA
#> n25  NA         NA           NA        NA           NA        NA        NA
#> n26  NA         NA           NA        NA           NA        NA        NA
#> n27  NA         NA           NA        NA           NA        NA        NA
#> n28  NA         NA           NA        NA           NA        NA        NA
#> n29  NA         NA           NA        NA           NA        NA        NA
#> n30  NA         NA           NA        NA           NA        NA        NA
#>           n18          n19       n20          n21       n22       n23       n24
#> n1         NA           NA        NA           NA        NA        NA        NA
#> n2         NA           NA 0.1922162           NA        NA        NA 0.9314693
#> n3         NA           NA        NA 4.558909e-02        NA        NA        NA
#> n4         NA           NA        NA           NA        NA        NA        NA
#> n5         NA           NA        NA           NA        NA 0.1796841 0.7858699
#> n6         NA           NA        NA           NA 0.5521354        NA        NA
#> n7         NA 0.0048797026        NA           NA        NA 0.8011242        NA
#> n8         NA 0.5041626500 0.6588650           NA        NA 0.5912294 0.7677566
#> n9         NA           NA        NA           NA 0.3323232        NA        NA
#> n10        NA           NA        NA           NA        NA        NA        NA
#> n11 0.6998152           NA 0.5591826           NA        NA        NA        NA
#> n12        NA           NA        NA           NA        NA        NA        NA
#> n13        NA           NA        NA           NA        NA        NA        NA
#> n14        NA           NA        NA           NA        NA        NA        NA
#> n15        NA 0.0005905704        NA           NA        NA        NA        NA
#> n16        NA           NA        NA           NA        NA        NA        NA
#> n17 0.8400321           NA        NA 1.309615e-07        NA        NA        NA
#> n18        NA           NA        NA 9.196068e-01        NA        NA        NA
#> n19        NA           NA        NA           NA        NA        NA        NA
#> n20        NA           NA        NA           NA        NA        NA        NA
#> n21        NA           NA        NA           NA        NA        NA        NA
#> n22        NA           NA        NA           NA        NA        NA        NA
#> n23        NA           NA        NA           NA        NA        NA        NA
#> n24        NA           NA        NA           NA        NA        NA        NA
#> n25        NA           NA        NA           NA        NA        NA        NA
#> n26        NA           NA        NA           NA        NA        NA        NA
#> n27        NA           NA        NA           NA        NA        NA        NA
#> n28        NA           NA        NA           NA        NA        NA        NA
#> n29        NA           NA        NA           NA        NA        NA        NA
#> n30        NA           NA        NA           NA        NA        NA        NA
#>              n25       n26        n27       n28       n29        n30
#> n1            NA        NA         NA        NA        NA         NA
#> n2  0.0003432082        NA         NA        NA        NA         NA
#> n3            NA        NA         NA        NA        NA         NA
#> n4            NA        NA         NA        NA        NA         NA
#> n5            NA        NA         NA        NA        NA         NA
#> n6            NA        NA         NA        NA        NA         NA
#> n7            NA        NA 0.07910643        NA        NA         NA
#> n8            NA        NA         NA        NA        NA         NA
#> n9            NA        NA         NA 0.7012544        NA         NA
#> n10           NA        NA         NA        NA        NA         NA
#> n11           NA        NA 0.55017446        NA        NA         NA
#> n12 0.9934539984 0.2942726         NA        NA        NA         NA
#> n13           NA        NA         NA        NA        NA         NA
#> n14           NA        NA         NA        NA 0.8209001         NA
#> n15           NA 0.2352295         NA        NA        NA         NA
#> n16           NA        NA         NA 0.9449381        NA         NA
#> n17           NA        NA         NA        NA        NA         NA
#> n18           NA        NA         NA        NA        NA         NA
#> n19           NA        NA         NA        NA 0.9109572         NA
#> n20           NA        NA         NA        NA        NA 0.66993952
#> n21           NA        NA         NA 0.8864057        NA 0.01492279
#> n22           NA        NA         NA        NA        NA         NA
#> n23 0.4571799110        NA 0.45973643        NA 0.9338571         NA
#> n24           NA 0.1358044         NA        NA        NA         NA
#> n25           NA        NA         NA        NA        NA         NA
#> n26           NA        NA         NA        NA        NA         NA
#> n27           NA        NA         NA        NA        NA 0.51060376
#> n28           NA        NA         NA        NA        NA         NA
#> n29           NA        NA         NA        NA        NA         NA
#> n30           NA        NA         NA        NA        NA         NA
#> 
#> $pathway_pvalue
#> [1] 0.4852498
#> 
#> attr(,"class")
#> [1] "dce"