Partial correlation

Robust partial correlation of column variables of a numeric matrix

pcor(x, ...)

Arguments

x
matrix

...
additional arguments for functions 'cor'

Value

matrix of partial correlations

Examples

x <- matrix(rnorm(100),10,10)
pcor(x)

#> Warning: The inverse of variance-covariance matrix is calculated using Moore-Penrose generalized matrix invers due to its determinant of zero.

#> Warning: NaNs produced

#>               [,1]        [,2]         [,3]        [,4]        [,5]        [,6]
#>  [1,]  1.000000000 -0.47643824 -0.519424131 -0.19341784 -0.70019139 -0.33209625
#>  [2,] -0.476438237  1.00000000  0.121129771 -0.50580165 -0.11083204 -0.08294416
#>  [3,] -0.519424131  0.12112977  1.000000000  0.08122390 -0.76838250 -0.31497524
#>  [4,] -0.193417835 -0.50580165  0.081223896  1.00000000  0.29204058 -0.35246619
#>  [5,] -0.700191393 -0.11083204 -0.768382497  0.29204058  1.00000000 -0.27056717
#>  [6,] -0.332096251 -0.08294416 -0.314975242 -0.35246619 -0.27056717  1.00000000
#>  [7,] -0.910861466 -0.59048069 -0.451913584 -0.11163437 -0.71101344 -0.36114567
#>  [8,]  0.013976222 -0.44330905  0.368934376 -0.59132566  0.28564131  0.03988216
#>  [9,]  0.006683528 -0.23730899 -0.002465881  0.01827767  0.02876053  0.29465360
#> [10,]  0.691428551  0.79723690  0.061170922  0.53666528  0.12830444  0.10475626
#>               [,7]         [,8]         [,9]      [,10]
#>  [1,] -0.910861466  0.013976222  0.006683528 0.69142855
#>  [2,] -0.590480690 -0.443309052 -0.237308994 0.79723690
#>  [3,] -0.451913584  0.368934376 -0.002465881 0.06117092
#>  [4,] -0.111634375 -0.591325659  0.018277669 0.53666528
#>  [5,] -0.711013438  0.285641309  0.028760531 0.12830444
#>  [6,] -0.361145671  0.039882160  0.294653605 0.10475626
#>  [7,]  1.000000000  0.009783456 -0.051177889 0.62603665
#>  [8,]  0.009783456  1.000000000  0.095218857 0.48610999
#>  [9,] -0.051177889  0.095218857  1.000000000 0.20617417
#> [10,]  0.626036653  0.486109993  0.206174174 1.00000000