Partial correlation

Robust partial correlation of column variables of a numeric matrix

pcor(x, ...)

Arguments

x
matrix

...
additional arguments for functions 'cor'

Value

matrix of partial correlations

Examples

x <- matrix(rnorm(100),10,10)
pcor(x)

#> Warning: The inverse of variance-covariance matrix is calculated using Moore-Penrose generalized matrix invers due to its determinant of zero.

#> Warning: NaNs produced

#>              [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,]  1.00000000 -0.4648420 -0.3055647 -0.1638395 -0.5263009 -0.4250535
#>  [2,] -0.46484199  1.0000000  0.4315344 -0.7418433 -0.2609140 -0.5277272
#>  [3,] -0.30556468  0.4315344  1.0000000  0.6158839  0.2096111  0.4417038
#>  [4,] -0.16383946 -0.7418433  0.6158839  1.0000000 -0.4401912 -0.6226528
#>  [5,] -0.52630088 -0.2609140  0.2096111 -0.4401912  1.0000000 -0.8589295
#>  [6,] -0.42505350 -0.5277272  0.4417038 -0.6226528 -0.8589295  1.0000000
#>  [7,] -0.52414515 -0.3376446  0.1813455 -0.4766589 -0.9174823 -0.8517800
#>  [8,] -0.87369886 -0.3190566 -0.3882858 -0.2062549 -0.3648386 -0.2199199
#>  [9,] -0.03800619  0.2372324 -0.6332630  0.4432337  0.6940810  0.7849656
#> [10,]  0.45813085  0.5115461 -0.4563019  0.7140373  0.8318327  0.8792743
#>             [,7]       [,8]        [,9]      [,10]
#>  [1,] -0.5241452 -0.8736989 -0.03800619  0.4581308
#>  [2,] -0.3376446 -0.3190566  0.23723238  0.5115461
#>  [3,]  0.1813455 -0.3882858 -0.63326299 -0.4563019
#>  [4,] -0.4766589 -0.2062549  0.44323368  0.7140373
#>  [5,] -0.9174823 -0.3648386  0.69408103  0.8318327
#>  [6,] -0.8517800 -0.2199199  0.78496558  0.8792743
#>  [7,]  1.0000000 -0.3768195  0.66760213  0.7910594
#>  [8,] -0.3768195  1.0000000 -0.15992978  0.3060713
#>  [9,]  0.6676021 -0.1599298  1.00000000 -0.6580579
#> [10,]  0.7910594  0.3060713 -0.65805793  1.0000000