Partial correlation

Robust partial correlation of column variables of a numeric matrix

pcor(x, ...)

Arguments

x
matrix

...
additional arguments for functions 'cor'

Value

matrix of partial correlations

Examples

x <- matrix(rnorm(100),10,10)
pcor(x)

#> Warning: The inverse of variance-covariance matrix is calculated using Moore-Penrose generalized matrix invers due to its determinant of zero.

#> Warning: NaNs produced

#>             [,1]        [,2]        [,3]        [,4]        [,5]       [,6]
#>  [1,]  1.0000000 -0.44865076 -0.58946583 -0.56163911 -0.44941155 -0.4040139
#>  [2,] -0.4486508  1.00000000 -0.02899803  0.20630409 -0.01327311 -0.6063535
#>  [3,] -0.5894658 -0.02899803  1.00000000 -0.85728219  0.06412155 -0.2892818
#>  [4,] -0.5616391  0.20630409 -0.85728219  1.00000000 -0.08021953 -0.2827109
#>  [5,] -0.4494115 -0.01327311  0.06412155 -0.08021953  1.00000000  0.2190529
#>  [6,] -0.4040139 -0.60635352 -0.28928177 -0.28271091  0.21905290  1.0000000
#>  [7,] -0.6410698 -0.84185943 -0.39447112 -0.18363110  0.06785296 -0.7934120
#>  [8,] -0.6327043 -0.67815890 -0.53100329 -0.35159110  0.09122319 -0.8090953
#>  [9,] -0.5328247 -0.54806224 -0.61538273 -0.39938995  0.33978703 -0.5942670
#> [10,]  0.3557150 -0.38676086  0.54824881  0.59836806  0.25810393 -0.1181388
#>              [,7]        [,8]       [,9]       [,10]
#>  [1,] -0.64106982 -0.63270435 -0.5328247  0.35571503
#>  [2,] -0.84185943 -0.67815890 -0.5480622 -0.38676086
#>  [3,] -0.39447112 -0.53100329 -0.6153827  0.54824881
#>  [4,] -0.18363110 -0.35159110 -0.3993899  0.59836806
#>  [5,]  0.06785296  0.09122319  0.3397870  0.25810393
#>  [6,] -0.79341205 -0.80909534 -0.5942670 -0.11813884
#>  [7,]  1.00000000 -0.92298051 -0.8328776 -0.10093630
#>  [8,] -0.92298051  1.00000000 -0.8598639  0.02877084
#>  [9,] -0.83287763 -0.85986385  1.0000000  0.15570666
#> [10,] -0.10093630  0.02877084  0.1557067  1.00000000