Partial correlation

Robust partial correlation of column variables of a numeric matrix

pcor(x, method = "spearman")

Arguments

x
matrix

Value

matrix of partial correlations

Examples

x <- matrix(rnorm(100),10,10)
pcor(x)

#> Warning: The inverse of variance-covariance matrix is calculated using Moore-Penrose generalized matrix invers due to its determinant of zero.

#> Warning: NaNs produced

#>              [,1]        [,2]        [,3]        [,4]       [,5]       [,6]
#>  [1,]  1.00000000  0.09344332 -0.84638444  0.82585471  0.2588387 -0.7164361
#>  [2,]  0.09344332  1.00000000  0.08714005  0.30384321  0.5736012 -0.1637110
#>  [3,] -0.84638444  0.08714005  1.00000000  0.70822457  0.1049553 -0.6385582
#>  [4,]  0.82585471  0.30384321  0.70822457  1.00000000 -0.4899031  0.7475578
#>  [5,]  0.25883870  0.57360121  0.10495527 -0.48990310  1.0000000  0.5211515
#>  [6,] -0.71643613 -0.16371096 -0.63855821  0.74755777  0.5211515  1.0000000
#>  [7,]  0.50838129  0.45293116  0.43131347 -0.55226962 -0.8215965  0.5802971
#>  [8,] -0.17469482 -0.50861984 -0.12970772  0.46981229  0.6992893 -0.3669544
#>  [9,]  0.31742620 -0.35481067  0.32870275 -0.04775157  0.3106241  0.1491976
#> [10,] -0.85201139 -0.33437321 -0.76286291  0.94925751  0.5675123 -0.8440719
#>              [,7]        [,8]        [,9]       [,10]
#>  [1,]  0.50838129 -0.17469482  0.31742620 -0.85201139
#>  [2,]  0.45293116 -0.50861984 -0.35481067 -0.33437321
#>  [3,]  0.43131347 -0.12970772  0.32870275 -0.76286291
#>  [4,] -0.55226962  0.46981229 -0.04775157  0.94925751
#>  [5,] -0.82159648  0.69928935  0.31062407  0.56751226
#>  [6,]  0.58029714 -0.36695441  0.14919761 -0.84407189
#>  [7,]  1.00000000  0.72940260  0.05848719  0.65210672
#>  [8,]  0.72940260  1.00000000 -0.09089233 -0.49515395
#>  [9,]  0.05848719 -0.09089233  1.00000000  0.09829961
#> [10,]  0.65210672 -0.49515395  0.09829961  1.00000000