Partial correlation

Robust partial correlation of column variables of a numeric matrix

pcor(x, g = NULL, adjustment_type = "parents", ...)

Arguments

x
matrix

g
related graph as adjacency matrix (optional)

adjustment_type
character string for the method to define the adjustment set Z for the regression

...
additional arguments for function 'cor'

Value

matrix of partial correlations

Examples

x <- matrix(rnorm(100),10,10)
pcor(x)

#> Warning: The inverse of variance-covariance matrix is calculated using Moore-Penrose generalized matrix invers due to its determinant of zero.

#> Warning: NaNs produced

#>             [,1]        [,2]       [,3]        [,4]         [,5]       [,6]
#>  [1,]  1.0000000 -0.70055605  0.6979819  0.62859084  0.775515686 -0.8665417
#>  [2,] -0.7005560  1.00000000  0.9860288  0.72706574  0.968294586 -0.9376398
#>  [3,]  0.6979819  0.98602878  1.0000000 -0.71563606 -0.957360956  0.9351657
#>  [4,]  0.6285908  0.72706574 -0.7156361  1.00000000 -0.658494940  0.7279581
#>  [5,]  0.7755157  0.96829459 -0.9573610 -0.65849494  1.000000000  0.9502226
#>  [6,] -0.8665417 -0.93763984  0.9351657  0.72795814  0.950222565  1.0000000
#>  [7,]  0.7997415  0.98125003 -0.9762323 -0.74382805 -0.965824742  0.9750838
#>  [8,]  0.8381308  0.86001764 -0.8858015 -0.65146627 -0.859238465  0.9403969
#>  [9,]  0.2342361 -0.08592856  0.1168653  0.01078242  0.004489181  0.1695102
#> [10,] -0.6333095 -0.94430452  0.9450270  0.69940280  0.889756452 -0.9083258
#>              [,7]       [,8]         [,9]      [,10]
#>  [1,]  0.79974149  0.8381308  0.234236077 -0.6333095
#>  [2,]  0.98125003  0.8600176 -0.085928564 -0.9443045
#>  [3,] -0.97623227 -0.8858015  0.116865275  0.9450270
#>  [4,] -0.74382805 -0.6514663  0.010782425  0.6994028
#>  [5,] -0.96582474 -0.8592385  0.004489181  0.8897565
#>  [6,]  0.97508382  0.9403969  0.169510178 -0.9083258
#>  [7,]  1.00000000 -0.9223709 -0.017162694  0.9457788
#>  [8,] -0.92237087  1.0000000 -0.188467530  0.8819946
#>  [9,] -0.01716269 -0.1884675  1.000000000 -0.0122611
#> [10,]  0.94577884  0.8819946 -0.012261102  1.0000000