Partial correlation

Robust partial correlation of column variables of a numeric matrix

pcor(x, g = NULL, adjustment_type = "parents", ...)

Arguments

x

matrix

g

related graph as adjacency matrix (optional)

adjustment_type

character string for the method to define the adjustment set Z for the regression

...

additional arguments for function 'cor'

Value

matrix of partial correlations

Examples

x <- matrix(rnorm(100),10,10)
pcor(x)
#> Warning: The inverse of variance-covariance matrix is calculated using Moore-Penrose generalized matrix invers due to its determinant of zero.
#> Warning: NaNs produced
#>             [,1]        [,2]       [,3]        [,4]       [,5]        [,6]
#>  [1,]  1.0000000  0.47241852 -0.1996859  0.76716432  0.9440677 -0.60533292
#>  [2,]  0.4724185  1.00000000  0.8381907 -0.89376274 -0.3646377 -0.23490868
#>  [3,] -0.1996859  0.83819073  1.0000000  0.67735484  0.1190281  0.39326611
#>  [4,]  0.7671643 -0.89376274  0.6773548  1.00000000 -0.6339549  0.02804701
#>  [5,]  0.9440677 -0.36463768  0.1190281 -0.63395486  1.0000000  0.72058248
#>  [6,] -0.6053329 -0.23490868  0.3932661  0.02804701  0.7205825  1.00000000
#>  [7,]  0.1165516  0.43306669 -0.5044895  0.38375327 -0.1912623  0.57053644
#>  [8,] -0.7318509  0.03630074  0.2596694  0.36978326  0.7500713 -0.73816535
#>  [9,] -0.8072361  0.77642010 -0.5502345  0.91061095  0.7292420 -0.26257695
#> [10,]  0.1422828 -0.09806457  0.2379480 -0.08374888 -0.2128333  0.14006417
#>             [,7]        [,8]        [,9]       [,10]
#>  [1,]  0.1165516 -0.73185093 -0.80723605  0.14228280
#>  [2,]  0.4330667  0.03630074  0.77642010 -0.09806457
#>  [3,] -0.5044895  0.25966939 -0.55023451  0.23794796
#>  [4,]  0.3837533  0.36978326  0.91061095 -0.08374888
#>  [5,] -0.1912623  0.75007126  0.72924197 -0.21283326
#>  [6,]  0.5705364 -0.73816535 -0.26257695  0.14006417
#>  [7,]  1.0000000  0.23470833 -0.28853603 -0.28211988
#>  [8,]  0.2347083  1.00000000 -0.56603775 -0.23943201
#>  [9,] -0.2885360 -0.56603775  1.00000000 -0.08125521
#> [10,] -0.2821199 -0.23943201 -0.08125521  1.00000000